@@ -30,46 +41,53 @@
   },
   "dependencies": {
     "jsonschema": "^1.0.2",
-    "setimmediate": "^1.0.4"
+    "lodash.topath": "^4.5.2",
+    "prop-types": "^15.5.8",
+    "setimmediate": "^1.0.5"
   },
   "devDependencies": {
     "atob": "^2.0.3",
-    "babel-cli": "^6.6.5",
-    "babel-core": "^6.7.4",
-    "babel-eslint": "^6.0.2",
-    "babel-loader": "^6.2.4",
+    "babel-cli": "^6.18.0",
+    "babel-core": "^6.18.2",
+    "babel-eslint": "^6.1.2",
+    "babel-loader": "^6.2.7",
     "babel-plugin-react-transform": "^2.0.2",
-    "babel-plugin-transform-class-properties": "^6.6.0",
-    "babel-plugin-transform-object-rest-spread": "^6.6.5",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
+    "babel-plugin-transform-class-properties": "^6.18.0",
+    "babel-plugin-transform-object-rest-spread": "^6.16.0",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.16.0",
+    "babel-register": "^6.18.0",
     "chai": "^3.3.0",
-    "codemirror": "^5.13.2",
+    "codemirror": "^5.20.2",
+    "cross-env": "^2.0.1",
     "css-loader": "^0.23.1",
-    "eslint": "^2.9.0",
+    "eslint": "^3.15.0",
     "eslint-plugin-react": "^4.2.3",
     "estraverse": "^4.2.0",
     "estraverse-fb": "^1.3.1",
-    "express": "^4.13.3",
+    "express": "^4.14.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "gh-pages": "^0.11.0",
     "html": "0.0.10",
+    "husky": "^0.13.2",
     "jsdom": "^8.3.0",
+    "lint-staged": "^3.3.1",
     "mocha": "^2.5.3",
-    "react": "^15.0.0",
-    "react-addons-test-utils": "^15.0.1",
+    "prettier": "^1.5.2",
+    "react": "^15.5.0",
+    "react-addons-test-utils": "^15.3.2",
     "react-codemirror": "^0.2.3",
-    "react-dom": "^15.0.0",
+    "react-dom": "^15.3.2",
     "react-transform-catch-errors": "^1.0.0",
     "react-transform-hmr": "^1.0.1",
-    "redbox-react": "^1.2.3",
-    "rimraf": "^2.4.4",
-    "sinon": "^1.17.2",
+    "redbox-react": "^1.3.3",
+    "rimraf": "^2.5.4",
+    "sinon": "^1.17.6",
     "style-loader": "^0.13.1",
     "toctoc": "^0.2.3",
-    "webpack": "^1.10.5",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-hot-middleware": "^2.6.0"
+    "webpack": "^1.13.3",
+    "webpack-dev-middleware": "^1.8.4",
+    "webpack-hot-middleware": "^2.13.2"
   },
   "directories": {
     "test": "test"
