@@ -1,9 +1,8 @@
-import React, { PropTypes } from "react";
-
-import AltDateWidget from "./AltDateWidget";
-
+import React from "react";
+import PropTypes from "prop-types";
 
 function AltDateTimeWidget(props) {
+  const { AltDateWidget } = props.registry.widgets;
   return <AltDateWidget time {...props} />;
 }
 
@@ -11,7 +10,7 @@ if (process.env.NODE_ENV !== "production") {
   AltDateTimeWidget.propTypes = {
     schema: PropTypes.object.isRequired,
     id: PropTypes.string.isRequired,
-    value: React.PropTypes.string,
+    value: PropTypes.string,
     required: PropTypes.bool,
     onChange: PropTypes.func,
   };
