@@ -1,9 +1,13 @@
-import React, { PropTypes } from "react";
+import React from "react";
+import PropTypes from "prop-types";
 
-
-function HiddenWidget({id, value}) {
+function HiddenWidget({ id, value }) {
   return (
-    <input type="hidden" id={id} value={typeof value === "undefined" ? "" : value} />
+    <input
+      type="hidden"
+      id={id}
+      value={typeof value === "undefined" ? "" : value}
+    />
   );
 }
 
@@ -11,9 +15,9 @@ if (process.env.NODE_ENV !== "production") {
   HiddenWidget.propTypes = {
     id: PropTypes.string.isRequired,
     value: PropTypes.oneOfType([
-      React.PropTypes.string,
-      React.PropTypes.number,
-      React.PropTypes.bool,
+      PropTypes.string,
+      PropTypes.number,
+      PropTypes.bool,
     ]),
   };
 }
