@@ -20,10 +20,12 @@ import "codemirror/theme/eclipse.css";
 
 // Patching CodeMirror#componentWillReceiveProps so it's executed synchronously
 // Ref https://github.com/mozilla-services/react-jsonschema-form/issues/174
-Codemirror.prototype.componentWillReceiveProps = function (nextProps) {
-  if (this.codeMirror &&
-      nextProps.value !== undefined &&
-      this.codeMirror.getValue() != nextProps.value) {
+Codemirror.prototype.componentWillReceiveProps = function(nextProps) {
+  if (
+    this.codeMirror &&
+    nextProps.value !== undefined &&
+    this.codeMirror.getValue() != nextProps.value
+  ) {
     this.codeMirror.setValue(nextProps.value);
   }
   if (typeof nextProps.options === "object") {
@@ -35,10 +37,10 @@ Codemirror.prototype.componentWillReceiveProps = function (nextProps) {
   }
 };
 
-const log = (type) => console.log.bind(console, type);
-const fromJson = (json) => JSON.parse(json);
-const toJson = (val) => JSON.stringify(val, null, 2);
-const liveValidateSchema = {type: "boolean", title: "Live validation"};
+const log = type => console.log.bind(console, type);
+const fromJson = json => JSON.parse(json);
+const toJson = val => JSON.stringify(val, null, 2);
+const liveValidateSchema = { type: "boolean", title: "Live validation" };
 const cmOptions = {
   theme: "default",
   height: "auto",
@@ -55,70 +57,89 @@ const cmOptions = {
 };
 const themes = {
   default: {
-    stylesheet: "//maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css"
+    stylesheet:
+      "//maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css",
   },
   cerulean: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/cerulean/bootstrap.min.css"
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/cerulean/bootstrap.min.css",
   },
   cosmo: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/cosmo/bootstrap.min.css"
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/cosmo/bootstrap.min.css",
   },
   cyborg: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/cyborg/bootstrap.min.css",
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/cyborg/bootstrap.min.css",
     editor: "blackboard",
   },
   darkly: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/darkly/bootstrap.min.css",
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/darkly/bootstrap.min.css",
     editor: "mbo",
   },
   flatly: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/flatly/bootstrap.min.css",
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/flatly/bootstrap.min.css",
     editor: "ttcn",
   },
   journal: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/journal/bootstrap.min.css"
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/journal/bootstrap.min.css",
   },
   lumen: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/lumen/bootstrap.min.css"
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/lumen/bootstrap.min.css",
   },
   paper: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/paper/bootstrap.min.css"
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/paper/bootstrap.min.css",
   },
   readable: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/readable/bootstrap.min.css"
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/readable/bootstrap.min.css",
   },
   sandstone: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/sandstone/bootstrap.min.css",
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/sandstone/bootstrap.min.css",
     editor: "solarized",
   },
   simplex: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/simplex/bootstrap.min.css",
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/simplex/bootstrap.min.css",
     editor: "ttcn",
   },
   slate: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/slate/bootstrap.min.css",
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/slate/bootstrap.min.css",
     editor: "monokai",
   },
   spacelab: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/spacelab/bootstrap.min.css"
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/spacelab/bootstrap.min.css",
   },
   "solarized-dark": {
-    stylesheet: "//cdn.rawgit.com/aalpern/bootstrap-solarized/master/bootstrap-solarized-dark.css",
+    stylesheet:
+      "//cdn.rawgit.com/aalpern/bootstrap-solarized/master/bootstrap-solarized-dark.css",
     editor: "dracula",
   },
   "solarized-light": {
-    stylesheet: "//cdn.rawgit.com/aalpern/bootstrap-solarized/master/bootstrap-solarized-light.css",
+    stylesheet:
+      "//cdn.rawgit.com/aalpern/bootstrap-solarized/master/bootstrap-solarized-light.css",
     editor: "solarized",
   },
   superhero: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/superhero/bootstrap.min.css",
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/superhero/bootstrap.min.css",
     editor: "dracula",
   },
   united: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/united/bootstrap.min.css"
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/united/bootstrap.min.css",
   },
   yeti: {
-    stylesheet: "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/yeti/bootstrap.min.css",
+    stylesheet:
+      "//cdnjs.cloudflare.com/ajax/libs/bootswatch/3.3.6/yeti/bootstrap.min.css",
     editor: "eclipse",
   },
 };
@@ -126,34 +147,46 @@ const themes = {
 class GeoPosition extends Component {
   constructor(props) {
     super(props);
-    this.state = {...props.formData};
+    this.state = { ...props.formData };
   }
 
   onChange(name) {
-    return (event) => {
-      this.setState({[name]: parseFloat(event.target.value)});
+    return event => {
+      this.setState({ [name]: parseFloat(event.target.value) });
       setImmediate(() => this.props.onChange(this.state));
     };
   }
 
   render() {
-    const {lat, lon} = this.state;
+    const { lat, lon } = this.state;
     return (
       <div className="geo">
         <h3>Hey, I'm a custom component</h3>
-        <p>I'm registered as <code>geo</code> and referenced in
-        <code>uiSchema</code> as the <code>ui:field</code> to use for this
-        schema.</p>
+        <p>
+          I'm registered as <code>geo</code> and referenced in
+          <code>uiSchema</code> as the <code>ui:field</code> to use for this
+          schema.
+        </p>
         <div className="row">
           <div className="col-sm-6">
             <label>Latitude</label>
-            <input className="form-control" type="number" value={lat} step="0.00001"
-              onChange={this.onChange("lat")} />
+            <input
+              className="form-control"
+              type="number"
+              value={lat}
+              step="0.00001"
+              onChange={this.onChange("lat")}
+            />
           </div>
           <div className="col-sm-6">
             <label>Longitude</label>
-            <input className="form-control" type="number" value={lon} step="0.00001"
-              onChange={this.onChange("lon")} />
+            <input
+              className="form-control"
+              type="number"
+              value={lon}
+              step="0.00001"
+              onChange={this.onChange("lon")}
+            />
           </div>
         </div>
       </div>
@@ -164,31 +197,30 @@ class GeoPosition extends Component {
 class Editor extends Component {
   constructor(props) {
     super(props);
-    this.state = {valid: true, code: props.code};
+    this.state = { valid: true, code: props.code };
   }
 
   componentWillReceiveProps(props) {
-    this.setState({valid: true, code: props.code});
+    this.setState({ valid: true, code: props.code });
   }
 
   shouldComponentUpdate(nextProps, nextState) {
     return shouldRender(this, nextProps, nextState);
   }
 
-  onCodeChange = (code) => {
-    this.setState({valid: true, code});
+  onCodeChange = code => {
+    this.setState({ valid: true, code });
     setImmediate(() => {
       try {
         this.props.onChange(fromJson(this.state.code));
-      } catch(err) {
-        console.error(err);
-        this.setState({valid: false, code});
+      } catch (err) {
+        this.setState({ valid: false, code });
       }
     });
   };
 
   render() {
-    const {title, theme} = this.props;
+    const { title, theme } = this.props;
     const icon = this.state.valid ? "ok" : "remove";
     const cls = this.state.valid ? "valid" : "invalid";
     return (
@@ -200,7 +232,8 @@ class Editor extends Component {
         <Codemirror
           value={this.state.code}
           onChange={this.onCodeChange}
-          options={Object.assign({}, cmOptions, {theme})} />
+          options={Object.assign({}, cmOptions, { theme })}
+        />
       </div>
     );
   }
@@ -209,59 +242,97 @@ class Editor extends Component {
 class Selector extends Component {
   constructor(props) {
     super(props);
-    this.state = {current: "Simple"};
+    this.state = { current: "Simple" };
   }
 
   shouldComponentUpdate(nextProps, nextState) {
     return shouldRender(this, nextProps, nextState);
   }
 
-  onLabelClick = (label) => {
-    return (event) => {
+  onLabelClick = label => {
+    return event => {
       event.preventDefault();
-      this.setState({current: label});
+      this.setState({ current: label });
       setImmediate(() => this.props.onSelected(samples[label]));
     };
   };
 
   render() {
     return (
-      <ul className="nav nav-pills">{
-        Object.keys(samples).map((label, i) => {
+      <ul className="nav nav-pills">
+        {Object.keys(samples).map((label, i) => {
           return (
-            <li key={i} role="presentation"
+            <li
+              key={i}
+              role="presentation"
               className={this.state.current === label ? "active" : ""}>
-              <a href="#"
-                onClick={this.onLabelClick(label)}>
+              <a href="#" onClick={this.onLabelClick(label)}>
                 {label}
               </a>
             </li>
           );
-        })
-      }</ul>
+        })}
+      </ul>
     );
   }
 }
 
-function ThemeSelector({theme, select}) {
+function ThemeSelector({ theme, select }) {
   const themeSchema = {
     type: "string",
-    enum: Object.keys(themes)
+    enum: Object.keys(themes),
   };
   return (
-    <Form schema={themeSchema}
-          formData={theme}
-          onChange={({formData}) => select(formData, themes[formData])}>
-      <div/>
+    <Form
+      schema={themeSchema}
+      formData={theme}
+      onChange={({ formData }) => select(formData, themes[formData])}>
+      <div />
     </Form>
   );
+}
+
+class CopyLink extends Component {
+  onCopyClick = event => {
+    this.input.select();
+    document.execCommand("copy");
+  };
+
+  render() {
+    const { shareURL, onShare } = this.props;
+    if (!shareURL) {
+      return (
+        <button className="btn btn-default" type="button" onClick={onShare}>
+          Share
+        </button>
+      );
+    }
+    return (
+      <div className="input-group">
+        <input
+          type="text"
+          ref={input => (this.input = input)}
+          className="form-control"
+          defaultValue={shareURL}
+        />
+        <span className="input-group-btn">
+          <button
+            className="btn btn-default"
+            type="button"
+            onClick={this.onCopyClick}>
+            <i className="glyphicon glyphicon-copy" />
+          </button>
+        </span>
+      </div>
+    );
+  }
 }
 
 class App extends Component {
   constructor(props) {
     super(props);
     // initialize state with Simple data sample
-    const {schema, uiSchema, formData, validate} = samples.Simple;
+    const { schema, uiSchema, formData, validate } = samples.Simple;
     this.state = {
       form: false,
       schema,
@@ -271,40 +342,65 @@ class App extends Component {
       editor: "default",
       theme: "default",
       liveValidate: true,
+      shareURL: null,
     };
   }
 
   componentDidMount() {
-    this.load(samples.Simple);
+    const hash = document.location.hash.match(/#(.*)/);
+    if (hash && typeof hash[1] === "string" && hash[1].length > 0) {
+      try {
+        this.load(JSON.parse(atob(hash[1])));
+      } catch (err) {
+        alert("Unable to load form setup data.");
+      }
+    } else {
+      this.load(samples.Simple);
+    }
   }
 
   shouldComponentUpdate(nextProps, nextState) {
     return shouldRender(this, nextProps, nextState);
   }
 
-  load = (data) => {
+  load = data => {
+    // Reset the ArrayFieldTemplate whenever you load new data
+    const { ArrayFieldTemplate } = data;
     // force resetting form component instance
-    this.setState({form: false},
-      _ => this.setState({...data, form: true}));
+    this.setState({ form: false }, _ =>
+      this.setState({ ...data, form: true, ArrayFieldTemplate })
+    );
   };
 
-  onSchemaEdited   = (schema) => this.setState({schema});
+  onSchemaEdited = schema => this.setState({ schema, shareURL: null });
 
-  onUISchemaEdited = (uiSchema) => this.setState({uiSchema});
+  onUISchemaEdited = uiSchema => this.setState({ uiSchema, shareURL: null });
 
-  onFormDataEdited = (formData) => this.setState({formData});
+  onFormDataEdited = formData => this.setState({ formData, shareURL: null });
 
-  onThemeSelected  = (theme, {stylesheet, editor}) => {
-    this.setState({theme, editor: editor ? editor : "default"});
+  onThemeSelected = (theme, { stylesheet, editor }) => {
+    this.setState({ theme, editor: editor ? editor : "default" });
     setImmediate(() => {
       // Side effect!
       document.getElementById("theme").setAttribute("href", stylesheet);
     });
   };
 
-  setLiveValidate = ({formData}) => this.setState({liveValidate: formData});
+  setLiveValidate = ({ formData }) => this.setState({ liveValidate: formData });
 
-  onFormDataChange = ({formData}) => this.setState({formData});
+  onFormDataChange = ({ formData }) =>
+    this.setState({ formData, shareURL: null });
+
+  onShare = () => {
+    const { formData, schema, uiSchema } = this.state;
+    const { location: { origin, pathname } } = document;
+    try {
+      const hash = btoa(JSON.stringify({ formData, schema, uiSchema }));
+      this.setState({ shareURL: `${origin}${pathname}#${hash}` });
+    } catch (err) {
+      this.setState({ shareURL: null });
+    }
+  };
 
   render() {
     const {
@@ -314,7 +410,9 @@ class App extends Component {
       liveValidate,
       validate,
       theme,
-      editor
+      editor,
+      ArrayFieldTemplate,
+      transformErrors,
     } = this.state;
 
     return (
@@ -326,9 +424,12 @@ class App extends Component {
               <Selector onSelected={this.load} />
             </div>
             <div className="col-sm-2">
-              <Form schema={liveValidateSchema}
-                    formData={liveValidate}
-                    onChange={this.setLiveValidate}><div/></Form>
+              <Form
+                schema={liveValidateSchema}
+                formData={liveValidate}
+                onChange={this.setLiveValidate}>
+                <div />
+              </Form>
             </div>
             <div className="col-sm-2">
               <ThemeSelector theme={theme} select={this.onThemeSelected} />
@@ -336,30 +437,62 @@ class App extends Component {
           </div>
         </div>
         <div className="col-sm-7">
-          <Editor title="JSONSchema" theme={editor} code={toJson(schema)}
-            onChange={this.onSchemaEdited} />
+          <Editor
+            title="JSONSchema"
+            theme={editor}
+            code={toJson(schema)}
+            onChange={this.onSchemaEdited}
+          />
           <div className="row">
             <div className="col-sm-6">
-              <Editor title="UISchema" theme={editor} code={toJson(uiSchema)}
-                onChange={this.onUISchemaEdited} />
+              <Editor
+                title="UISchema"
+                theme={editor}
+                code={toJson(uiSchema)}
+                onChange={this.onUISchemaEdited}
+              />
             </div>
             <div className="col-sm-6">
-              <Editor title="formData" theme={editor} code={toJson(formData)}
-                onChange={this.onFormDataEdited} />
+              <Editor
+                title="formData"
+                theme={editor}
+                code={toJson(formData)}
+                onChange={this.onFormDataEdited}
+              />
             </div>
           </div>
         </div>
         <div className="col-sm-5">
-          {!this.state.form ? null :
+          {this.state.form &&
             <Form
+              ArrayFieldTemplate={ArrayFieldTemplate}
               liveValidate={liveValidate}
               schema={schema}
               uiSchema={uiSchema}
               formData={formData}
               onChange={this.onFormDataChange}
-              fields={{geo: GeoPosition}}
+              onSubmit={({ formData }) =>
+                console.log("submitted formData", formData)}
+              fields={{ geo: GeoPosition }}
               validate={validate}
-              onError={log("errors")} />}
+              onBlur={(id, value) =>
+                console.log(`Touched ${id} with value ${value}`)}
+              transformErrors={transformErrors}
+              onError={log("errors")}>
+              <div className="row">
+                <div className="col-sm-3">
+                  <button className="btn btn-primary" type="submit">
+                    Submit
+                  </button>
+                </div>
+                <div className="col-sm-9 text-right">
+                  <CopyLink
+                    shareURL={this.state.shareURL}
+                    onShare={this.onShare}
+                  />
+                </div>
+              </div>
+            </Form>}
         </div>
       </div>
     );
