@@ -1,5 +1,16 @@
 module.exports = {
   schema: {
+    definitions: {
+      Thing: {
+        type: "object",
+        properties: {
+          name: {
+            type: "string",
+            default: "Default name",
+          },
+        },
+      },
+    },
     type: "object",
     properties: {
       listOfStrings: {
@@ -7,8 +18,8 @@ module.exports = {
         title: "A list of strings",
         items: {
           type: "string",
-          default: "bazinga"
-        }
+          default: "bazinga",
+        },
       },
       multipleChoicesList: {
         type: "array",
@@ -17,7 +28,7 @@ module.exports = {
           type: "string",
           enum: ["foo", "bar", "fuzz", "qux"],
         },
-        uniqueItems: true
+        uniqueItems: true,
       },
       fixedItemsList: {
         type: "array",
@@ -26,17 +37,25 @@ module.exports = {
           {
             title: "A string value",
             type: "string",
-            default: "lorem ipsum"
+            default: "lorem ipsum",
           },
           {
             title: "a boolean value",
-            type: "boolean"
-          }
+            type: "boolean",
+          },
         ],
         additionalItems: {
           title: "Additional item",
-          type: "number"
-        }
+          type: "number",
+        },
+      },
+      minItemsList: {
+        type: "array",
+        title: "A list with a minimal number of items",
+        minItems: 3,
+        items: {
+          $ref: "#/definitions/Thing",
+        },
       },
       nestedList: {
         type: "array",
@@ -46,30 +65,103 @@ module.exports = {
           title: "Inner list",
           items: {
             type: "string",
-            default: "lorem ipsum"
-          }
-        }
-      }
-    }
+            default: "lorem ipsum",
+          },
+        },
+      },
+      unorderable: {
+        title: "Unorderable items",
+        type: "array",
+        items: {
+          type: "string",
+          default: "lorem ipsum",
+        },
+      },
+      unremovable: {
+        title: "Unremovable items",
+        type: "array",
+        items: {
+          type: "string",
+          default: "lorem ipsum",
+        },
+      },
+      noToolbar: {
+        title: "No add, remove and order buttons",
+        type: "array",
+        items: {
+          type: "string",
+          default: "lorem ipsum",
+        },
+      },
+      fixedNoToolbar: {
+        title: "Fixed array without buttons",
+        type: "array",
+        items: [
+          {
+            title: "A number",
+            type: "number",
+            default: 42,
+          },
+          {
+            title: "A boolean",
+            type: "boolean",
+            default: false,
+          },
+        ],
+        additionalItems: {
+          title: "A string",
+          type: "string",
+          default: "lorem ipsum",
+        },
+      },
+    },
   },
   uiSchema: {
+    listOfStrings: {
+      items: { "ui:emptyValue": "" },
+    },
     multipleChoicesList: {
-      "ui:widget": "checkboxes"
+      "ui:widget": "checkboxes",
     },
     fixedItemsList: {
-      items: [
-        {"ui:widget": "textarea"},
-        {"ui:widget": "select"}
-      ],
+      items: [{ "ui:widget": "textarea" }, { "ui:widget": "select" }],
       additionalItems: {
-        "ui:widget": "updown"
-      }
-    }
+        "ui:widget": "updown",
+      },
+    },
+    unorderable: {
+      "ui:options": {
+        orderable: false,
+      },
+    },
+    unremovable: {
+      "ui:options": {
+        removable: false,
+      },
+    },
+    noToolbar: {
+      "ui:options": {
+        addable: false,
+        orderable: false,
+        removable: false,
+      },
+    },
+    fixedNoToolbar: {
+      "ui:options": {
+        addable: false,
+        orderable: false,
+        removable: false,
+      },
+    },
   },
   formData: {
     listOfStrings: ["foo", "bar"],
     multipleChoicesList: ["foo", "bar"],
     fixedItemsList: ["Some text", true, 123],
-    nestedList: [["lorem", "ipsum"], ["dolor"]]
-  }
+    nestedList: [["lorem", "ipsum"], ["dolor"]],
+    unorderable: ["one", "two"],
+    unremovable: ["one", "two"],
+    noToolbar: ["one", "two"],
+    fixedNoToolbar: [42, true, "additional item one", "additional item two"],
+  },
 };
