@@ -5,18 +5,19 @@ module.exports = {
     properties: {
       native: {
         title: "Native",
-        description: "May not work on some browsers, notably Firefox Desktop and IE.",
+        description:
+          "May not work on some browsers, notably Firefox Desktop and IE.",
         type: "object",
         properties: {
-          "datetime": {
+          datetime: {
             type: "string",
-            format: "date-time"
+            format: "date-time",
           },
-          "date": {
+          date: {
             type: "string",
-            format: "date"
-          }
-        }
+            format: "date",
+          },
+        },
       },
       alternative: {
         title: "Alternative",
@@ -25,25 +26,25 @@ module.exports = {
         properties: {
           "alt-datetime": {
             type: "string",
-            format: "date-time"
+            format: "date-time",
           },
           "alt-date": {
             type: "string",
-            format: "date"
-          }
-        }
-      }
-    }
+            format: "date",
+          },
+        },
+      },
+    },
   },
   uiSchema: {
     alternative: {
       "alt-datetime": {
-        "ui:widget": "alt-datetime"
+        "ui:widget": "alt-datetime",
       },
       "alt-date": {
-        "ui:widget": "alt-date"
-      }
-    }
+        "ui:widget": "alt-date",
+      },
+    },
   },
-  formData: {}
+  formData: {},
 };
