@@ -1,6 +1,6 @@
 function largeEnum(n) {
   const list = [];
-  for (let i=0; i<n;i++) {
+  for (let i = 0; i < n; i++) {
     list.push("option #" + i);
   }
   return list;
@@ -9,7 +9,7 @@ function largeEnum(n) {
 module.exports = {
   schema: {
     definitions: {
-      largeEnum: {type: "string", enum: largeEnum(100)}
+      largeEnum: { type: "string", enum: largeEnum(100) },
     },
     title: "A rather large form",
     type: "object",
@@ -18,18 +18,22 @@ module.exports = {
         type: "string",
         title: "Some string",
       },
-      choice1: {$ref: "#/definitions/largeEnum"},
-      choice2: {$ref: "#/definitions/largeEnum"},
-      choice3: {$ref: "#/definitions/largeEnum"},
-      choice4: {$ref: "#/definitions/largeEnum"},
-      choice5: {$ref: "#/definitions/largeEnum"},
-      choice6: {$ref: "#/definitions/largeEnum"},
-      choice7: {$ref: "#/definitions/largeEnum"},
-      choice8: {$ref: "#/definitions/largeEnum"},
-      choice9: {$ref: "#/definitions/largeEnum"},
-      choice10: {$ref: "#/definitions/largeEnum"},
-    }
+      choice1: { $ref: "#/definitions/largeEnum" },
+      choice2: { $ref: "#/definitions/largeEnum" },
+      choice3: { $ref: "#/definitions/largeEnum" },
+      choice4: { $ref: "#/definitions/largeEnum" },
+      choice5: { $ref: "#/definitions/largeEnum" },
+      choice6: { $ref: "#/definitions/largeEnum" },
+      choice7: { $ref: "#/definitions/largeEnum" },
+      choice8: { $ref: "#/definitions/largeEnum" },
+      choice9: { $ref: "#/definitions/largeEnum" },
+      choice10: { $ref: "#/definitions/largeEnum" },
+    },
   },
-  uiSchema: {},
-  formData: {}
+  uiSchema: {
+    choice1: {
+      "ui:placeholder": "Choose one",
+    },
+  },
+  formData: {},
 };
