@@ -6,7 +6,7 @@ module.exports = {
     properties: {
       title: {
         type: "string",
-        title: "Task list title"
+        title: "Task list title",
       },
       tasks: {
         type: "array",
@@ -18,45 +18,47 @@ module.exports = {
             title: {
               type: "string",
               title: "Title",
-              description: "A sample title"
+              description: "A sample title",
             },
             details: {
               type: "string",
               title: "Task details",
-              description: "Enter the task details"
+              description: "Enter the task details",
             },
             done: {
               type: "boolean",
               title: "Done?",
-              default: false
-            }
-          }
-        }
-      }
-    }
+              default: false,
+            },
+          },
+        },
+      },
+    },
   },
   uiSchema: {
     tasks: {
       items: {
         details: {
-          "ui:widget": "textarea"
-        }
-      }
-    }
+          "ui:widget": "textarea",
+        },
+      },
+    },
   },
   formData: {
     title: "My current tasks",
     tasks: [
       {
         title: "My first task",
-        details: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
-        done: true
+        details:
+          "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+        done: true,
       },
       {
         title: "My second task",
-        details: "Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur",
-        done: false
-      }
-    ]
-  }
+        details:
+          "Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur",
+        done: false,
+      },
+    ],
+  },
 };
