@@ -15,7 +15,7 @@ module.exports = {
       },
       age: {
         type: "integer",
-        title: "Age"
+        title: "Age",
       },
       bio: {
         type: "string",
@@ -24,30 +24,46 @@ module.exports = {
       password: {
         type: "string",
         title: "Password",
-        minLength: 3
-      }
-    }
+        minLength: 3,
+      },
+      telephone: {
+        type: "string",
+        title: "Telephone",
+        minLength: 10,
+      },
+    },
   },
   uiSchema: {
+    firstName: {
+      "ui:autofocus": true,
+      "ui:emptyValue": "",
+    },
     age: {
-      "ui:widget": "updown"
+      "ui:widget": "updown",
+      "ui:title": "Age of person",
+      "ui:description": "(earthian year)",
     },
     bio: {
-      "ui:widget": "textarea"
+      "ui:widget": "textarea",
     },
     password: {
       "ui:widget": "password",
-      "ui:help": "Hint: Make it strong!"
+      "ui:help": "Hint: Make it strong!",
     },
     date: {
-      "ui:widget": "alt-datetime"
-    }
+      "ui:widget": "alt-datetime",
+    },
+    telephone: {
+      "ui:options": {
+        inputType: "tel",
+      },
+    },
   },
   formData: {
     firstName: "Chuck",
     lastName: "Norris",
     age: 75,
     bio: "Roundhouse kicking asses since 1940",
-    password: "noneed"
-  }
+    password: "noneed",
+  },
 };
