@@ -1,6 +1,5 @@
 import React from "react";
 
-
 module.exports = {
   schema: {
     title: "Widgets",
@@ -12,13 +11,13 @@ module.exports = {
         properties: {
           email: {
             type: "string",
-            format: "email"
+            format: "email",
           },
           uri: {
             type: "string",
-            format: "uri"
-          }
-        }
+            format: "uri",
+          },
+        },
       },
       boolean: {
         type: "object",
@@ -26,17 +25,20 @@ module.exports = {
         properties: {
           default: {
             type: "boolean",
-            title: "checkbox (default)"
+            title: "checkbox (default)",
+            description: "This is the checkbox-description",
           },
           radio: {
             type: "boolean",
-            title: "radio buttons"
+            title: "radio buttons",
+            description: "This is the radio-description",
           },
           select: {
             type: "boolean",
-            title: "select box"
-          }
-        }
+            title: "select box",
+            description: "This is the select-description",
+          },
+        },
       },
       string: {
         type: "object",
@@ -44,32 +46,32 @@ module.exports = {
         properties: {
           default: {
             type: "string",
-            title: "text input (default)"
+            title: "text input (default)",
           },
           textarea: {
             type: "string",
-            title: "textarea"
+            title: "textarea",
           },
           color: {
             type: "string",
             title: "color picker",
-            default: "#151ce6"
-          }
-        }
+            default: "#151ce6",
+          },
+        },
       },
       secret: {
         type: "string",
-        default: "I'm a hidden string."
+        default: "I'm a hidden string.",
       },
       disabled: {
         type: "string",
         title: "A disabled field",
-        default: "I am disabled."
+        default: "I am disabled.",
       },
       readonly: {
         type: "string",
         title: "A readonly field",
-        default: "I am read-only."
+        default: "I am read-only.",
       },
       widgetOptions: {
         title: "Custom widget with options",
@@ -81,70 +83,76 @@ module.exports = {
         type: "string",
         enum: ["foo", "bar"],
         enumNames: ["Foo", "Bar"],
-      }
-    }
+      },
+    },
   },
   uiSchema: {
     boolean: {
       radio: {
-        "ui:widget": "radio"
+        "ui:widget": "radio",
       },
       select: {
-        "ui:widget": "select"
-      }
+        "ui:widget": "select",
+      },
     },
     string: {
       textarea: {
-        "ui:widget": "textarea"
+        "ui:widget": "textarea",
+        "ui:options": {
+          rows: 5,
+        },
       },
       color: {
-        "ui:widget": "color"
-      }
+        "ui:widget": "color",
+      },
     },
     secret: {
-      "ui:widget": "hidden"
+      "ui:widget": "hidden",
     },
     disabled: {
-      "ui:disabled": true
+      "ui:disabled": true,
     },
     readonly: {
-      "ui:readonly": true
+      "ui:readonly": true,
     },
     widgetOptions: {
-      "ui:widget": {
-        component: ({value, onChange, options}) => {
-          const {backgroundColor} = options;
-          return (
-            <input className="form-control"
-              onChange={(event) => onChange(event.target.value)}
-              style={{backgroundColor}}
-              value={value} />
-          );
-        },
-        options: {
-          backgroundColor: "yellow",
-        }
-      }
+      "ui:widget": ({ value, onChange, options }) => {
+        const { backgroundColor } = options;
+        return (
+          <input
+            className="form-control"
+            onChange={event => onChange(event.target.value)}
+            style={{ backgroundColor }}
+            value={value}
+          />
+        );
+      },
+      "ui:options": {
+        backgroundColor: "yellow",
+      },
     },
     selectWidgetOptions: {
-      "ui:widget": {
-        component: ({value, onChange, options}) => {
-          const {enumOptions, backgroundColor} = options;
-          return (
-            <select className="form-control"
-              style={{backgroundColor}}
-              value={value}
-              onChange={(event) => onChange(event.target.value)}>{
-              enumOptions.map(({label, value}, i) => {
-                return <option key={i} value={value}>{label}</option>;
-              })
-            }</select>
-          );
-        },
-        options: {
-          backgroundColor: "pink",
-        }
-      }
+      "ui:widget": ({ value, onChange, options }) => {
+        const { enumOptions, backgroundColor } = options;
+        return (
+          <select
+            className="form-control"
+            style={{ backgroundColor }}
+            value={value}
+            onChange={event => onChange(event.target.value)}>
+            {enumOptions.map(({ label, value }, i) => {
+              return (
+                <option key={i} value={value}>
+                  {label}
+                </option>
+              );
+            })}
+          </select>
+        );
+      },
+      "ui:options": {
+        backgroundColor: "pink",
+      },
     },
   },
   formData: {
@@ -155,12 +163,12 @@ module.exports = {
     boolean: {
       default: true,
       radio: true,
-      select: true
+      select: true,
     },
     string: {
       default: "Hello...",
-      textarea: "... World"
+      textarea: "... World",
     },
-    secret: "I'm a hidden string."
-  }
+    secret: "I'm a hidden string.",
+  },
 };
