@@ -1,18 +1,15 @@
-import React, { Component, PropTypes } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 
-import SchemaField from "./fields/SchemaField";
-import TitleField from "./fields/TitleField";
-import DescriptionField from "./fields/DescriptionField";
-
-import ErrorList from "./ErrorList";
+import { default as DefaultErrorList } from "./ErrorList";
 import {
   getDefaultFormState,
   shouldRender,
   toIdSchema,
   setState,
+  getDefaultRegistry,
 } from "../utils";
 import validateFormData from "../validate";
-
 
 export default class Form extends Component {
   static defaultProps = {
@@ -20,7 +17,9 @@ export default class Form extends Component {
     noValidate: false,
     liveValidate: false,
     safeRenderCompletion: false,
-  }
+    noHtml5Validate: false,
+    ErrorList: DefaultErrorList,
+  };
 
   constructor(props) {
     super(props);
@@ -38,14 +37,19 @@ export default class Form extends Component {
     const edit = typeof props.formData !== "undefined";
     const liveValidate = props.liveValidate || this.props.liveValidate;
     const mustValidate = edit && !props.noValidate && liveValidate;
-    const {definitions} = schema;
+    const { definitions } = schema;
     const formData = getDefaultFormState(schema, props.formData, definitions);
-    const {errors, errorSchema} = mustValidate ?
-      this.validate(formData, schema) : {
-        errors: state.errors || [],
-        errorSchema: state.errorSchema || {}
-      };
-    const idSchema = toIdSchema(schema, uiSchema["ui:rootFieldId"], definitions);
+    const { errors, errorSchema } = mustValidate
+      ? this.validate(formData, schema)
+      : {
+          errors: state.errors || [],
+          errorSchema: state.errorSchema || {},
+        };
+    const idSchema = toIdSchema(
+      schema,
+      uiSchema["ui:rootFieldId"],
+      definitions
+    );
     return {
       status: "initial",
       schema,
@@ -54,7 +58,7 @@ export default class Form extends Component {
       formData,
       edit,
       errors,
-      errorSchema
+      errorSchema,
     };
   }
 
@@ -63,13 +67,18 @@ export default class Form extends Component {
   }
 
   validate(formData, schema) {
-    const {validate} = this.props;
-    return validateFormData(formData, schema || this.props.schema, validate);
+    const { validate, transformErrors } = this.props;
+    return validateFormData(
+      formData,
+      schema || this.props.schema,
+      validate,
+      transformErrors
+    );
   }
 
   renderErrors() {
-    const {status, errors} = this.state;
-    const {showErrorList} = this.props;
+    const { status, errors } = this.state;
+    const { ErrorList, showErrorList } = this.props;
 
     if (status !== "editing" && errors.length && showErrorList != false) {
       return <ErrorList errors={errors} />;
@@ -77,12 +86,13 @@ export default class Form extends Component {
     return null;
   }
 
-  onChange = (formData, options={validate: false}) => {
-    const mustValidate = !this.props.noValidate && (this.props.liveValidate || options.validate);
-    let state = {status: "editing", formData};
+  onChange = (formData, options = { validate: false }) => {
+    const mustValidate =
+      !this.props.noValidate && (this.props.liveValidate || options.validate);
+    let state = { status: "editing", formData };
     if (mustValidate) {
-      const {errors, errorSchema} = this.validate(formData);
-      state = {...state, errors, errorSchema};
+      const { errors, errorSchema } = this.validate(formData);
+      state = { ...state, errors, errorSchema };
     }
     setState(this, state, () => {
       if (this.props.onChange) {
@@ -91,14 +101,20 @@ export default class Form extends Component {
     });
   };
 
-  onSubmit = (event) => {
+  onBlur = (...args) => {
+    if (this.props.onBlur) {
+      this.props.onBlur(...args);
+    }
+  };
+
+  onSubmit = event => {
     event.preventDefault();
-    this.setState({status: "submitted"});
+    this.setState({ status: "submitted" });
 
     if (!this.props.noValidate) {
-      const {errors, errorSchema} = this.validate(this.state.formData);
+      const { errors, errorSchema } = this.validate(this.state.formData);
       if (Object.keys(errors).length > 0) {
-        setState(this, {errors, errorSchema}, () => {
+        setState(this, { errors, errorSchema }, () => {
           if (this.props.onError) {
             this.props.onError(errors);
           } else {
@@ -112,25 +128,18 @@ export default class Form extends Component {
     if (this.props.onSubmit) {
       this.props.onSubmit(this.state);
     }
-    this.setState({status: "initial", errors: [], errorSchema: {}});
+    this.setState({ status: "initial", errors: [], errorSchema: {} });
   };
 
   getRegistry() {
     // For BC, accept passed SchemaField and TitleField props and pass them to
     // the "fields" registry one.
-    const _SchemaField = this.props.SchemaField || SchemaField;
-    const _TitleField = this.props.TitleField || TitleField;
-    const _DescriptionField = this.props.DescriptionField || DescriptionField;
-
-    const fields = Object.assign({
-      SchemaField: _SchemaField,
-      TitleField: _TitleField,
-      DescriptionField: _DescriptionField,
-    }, this.props.fields);
+    const { fields, widgets } = getDefaultRegistry();
     return {
-      fields,
+      fields: { ...fields, ...this.props.fields },
+      widgets: { ...widgets, ...this.props.widgets },
+      ArrayFieldTemplate: this.props.ArrayFieldTemplate,
       FieldTemplate: this.props.FieldTemplate,
-      widgets: this.props.widgets || {},
       definitions: this.props.schema.definitions || {},
       formContext: this.props.formContext || {},
     };
@@ -148,15 +157,17 @@ export default class Form extends Component {
       action,
       autocomplete,
       enctype,
-      acceptcharset
+      acceptcharset,
+      noHtml5Validate,
     } = this.props;
 
-    const {schema, uiSchema, formData, errorSchema, idSchema} = this.state;
+    const { schema, uiSchema, formData, errorSchema, idSchema } = this.state;
     const registry = this.getRegistry();
     const _SchemaField = registry.fields.SchemaField;
 
     return (
-      <form className={className ? className : "rjsf"}
+      <form
+        className={className ? className : "rjsf"}
         id={id}
         name={name}
         method={method}
@@ -165,6 +176,7 @@ export default class Form extends Component {
         autoComplete={autocomplete}
         encType={enctype}
         acceptCharset={acceptcharset}
+        noValidate={noHtml5Validate}
         onSubmit={this.onSubmit}>
         {this.renderErrors()}
         <_SchemaField
@@ -174,12 +186,17 @@ export default class Form extends Component {
           idSchema={idSchema}
           formData={formData}
           onChange={this.onChange}
+          onBlur={this.onBlur}
           registry={registry}
-          safeRenderCompletion={safeRenderCompletion} />
-        { children ? children :
-          <p>
-            <button type="submit" className="btn btn-info">Submit</button>
-          </p> }
+          safeRenderCompletion={safeRenderCompletion}
+        />
+        {children
+          ? children
+          : <p>
+              <button type="submit" className="btn btn-info">
+                Submit
+              </button>
+            </p>}
       </form>
     );
   }
@@ -190,12 +207,13 @@ if (process.env.NODE_ENV !== "production") {
     schema: PropTypes.object.isRequired,
     uiSchema: PropTypes.object,
     formData: PropTypes.any,
-    widgets: PropTypes.objectOf(PropTypes.oneOfType([
-      PropTypes.func,
-      PropTypes.object,
-    ])),
+    widgets: PropTypes.objectOf(
+      PropTypes.oneOfType([PropTypes.func, PropTypes.object])
+    ),
     fields: PropTypes.objectOf(PropTypes.func),
+    ArrayFieldTemplate: PropTypes.func,
     FieldTemplate: PropTypes.func,
+    ErrorList: PropTypes.func,
     onChange: PropTypes.func,
     onError: PropTypes.func,
     showErrorList: PropTypes.bool,
@@ -210,7 +228,10 @@ if (process.env.NODE_ENV !== "production") {
     enctype: PropTypes.string,
     acceptcharset: PropTypes.string,
     noValidate: PropTypes.bool,
+    noHtml5Validate: PropTypes.bool,
     liveValidate: PropTypes.bool,
+    validate: PropTypes.func,
+    transformErrors: PropTypes.func,
     safeRenderCompletion: PropTypes.bool,
     formContext: PropTypes.object,
   };
