@@ -1,7 +1,10 @@
-import React, { Component, PropTypes } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 
 import {
+  getWidget,
   getDefaultFormState,
+  getUiOptions,
   isMultiSelect,
   isFilesArray,
   isFixedItems,
@@ -9,144 +12,278 @@ import {
   optionsList,
   retrieveSchema,
   toIdSchema,
-  shouldRender,
   getDefaultRegistry,
-  setState
 } from "../../utils";
-import SelectWidget from "./../widgets/SelectWidget";
-import FileWidget from "./../widgets/FileWidget";
-import CheckboxesWidget from "./../widgets/CheckboxesWidget";
 
-
-function ArrayFieldTitle({TitleField, idSchema, title, required}) {
+function ArrayFieldTitle({ TitleField, idSchema, title, required }) {
   if (!title) {
-    return null;
+    // See #312: Ensure compatibility with old versions of React.
+    return <div />;
   }
   const id = `${idSchema.$id}__title`;
   return <TitleField id={id} title={title} required={required} />;
 }
 
-function ArrayFieldDescription({DescriptionField, idSchema, description}) {
+function ArrayFieldDescription({ DescriptionField, idSchema, description }) {
   if (!description) {
-    return null;
+    // See #312: Ensure compatibility with old versions of React.
+    return <div />;
   }
   const id = `${idSchema.$id}__description`;
   return <DescriptionField id={id} description={description} />;
 }
 
+function IconBtn(props) {
+  const { type = "default", icon, className, ...otherProps } = props;
+  return (
+    <button
+      type="button"
+      className={`btn btn-${type} ${className}`}
+      {...otherProps}>
+      <i className={`glyphicon glyphicon-${icon}`} />
+    </button>
+  );
+}
+
+// Used in the two templates
+function DefaultArrayItem(props) {
+  const btnStyle = {
+    flex: 1,
+    paddingLeft: 6,
+    paddingRight: 6,
+    fontWeight: "bold",
+  };
+  return (
+    <div key={props.index} className={props.className}>
+      <div className={props.hasToolbar ? "col-xs-9" : "col-xs-12"}>
+        {props.children}
+      </div>
+
+      {props.hasToolbar &&
+        <div className="col-xs-3 array-item-toolbox">
+          <div
+            className="btn-group"
+            style={{ display: "flex", justifyContent: "space-around" }}>
+            {(props.hasMoveUp || props.hasMoveDown) &&
+              <IconBtn
+                icon="arrow-up"
+                className="array-item-move-up"
+                tabIndex="-1"
+                style={btnStyle}
+                disabled={props.disabled || props.readonly || !props.hasMoveUp}
+                onClick={props.onReorderClick(props.index, props.index - 1)}
+              />}
+
+            {(props.hasMoveUp || props.hasMoveDown) &&
+              <IconBtn
+                icon="arrow-down"
+                className="array-item-move-down"
+                tabIndex="-1"
+                style={btnStyle}
+                disabled={
+                  props.disabled || props.readonly || !props.hasMoveDown
+                }
+                onClick={props.onReorderClick(props.index, props.index + 1)}
+              />}
+
+            {props.hasRemove &&
+              <IconBtn
+                type="danger"
+                icon="remove"
+                className="array-item-remove"
+                tabIndex="-1"
+                style={btnStyle}
+                disabled={props.disabled || props.readonly}
+                onClick={props.onDropIndexClick(props.index)}
+              />}
+          </div>
+        </div>}
+    </div>
+  );
+}
+
+function DefaultFixedArrayFieldTemplate(props) {
+  return (
+    <fieldset className={props.className}>
+      <ArrayFieldTitle
+        key={`array-field-title-${props.idSchema.$id}`}
+        TitleField={props.TitleField}
+        idSchema={props.idSchema}
+        title={props.title}
+        required={props.required}
+      />
+
+      {props.schema.description &&
+        <div
+          className="field-description"
+          key={`field-description-${props.idSchema.$id}`}>
+          {props.schema.description}
+        </div>}
+
+      <div
+        className="row array-item-list"
+        key={`array-item-list-${props.idSchema.$id}`}>
+        {props.items && props.items.map(DefaultArrayItem)}
+      </div>
+
+      {props.canAdd &&
+        <AddButton
+          onClick={props.onAddClick}
+          disabled={props.disabled || props.readonly}
+        />}
+    </fieldset>
+  );
+}
+
+function DefaultNormalArrayFieldTemplate(props) {
+  return (
+    <fieldset className={props.className}>
+      <ArrayFieldTitle
+        key={`array-field-title-${props.idSchema.$id}`}
+        TitleField={props.TitleField}
+        idSchema={props.idSchema}
+        title={props.title}
+        required={props.required}
+      />
+
+      {props.schema.description &&
+        <ArrayFieldDescription
+          key={`array-field-description-${props.idSchema.$id}`}
+          DescriptionField={props.DescriptionField}
+          idSchema={props.idSchema}
+          description={props.schema.description}
+        />}
+
+      <div
+        className="row array-item-list"
+        key={`array-item-list-${props.idSchema.$id}`}>
+        {props.items && props.items.map(p => DefaultArrayItem(p))}
+      </div>
+
+      {props.canAdd &&
+        <AddButton
+          onClick={props.onAddClick}
+          disabled={props.disabled || props.readonly}
+        />}
+    </fieldset>
+  );
+}
+
 class ArrayField extends Component {
   static defaultProps = {
     uiSchema: {},
+    formData: [],
     idSchema: {},
-    registry: getDefaultRegistry(),
     required: false,
     disabled: false,
     readonly: false,
+    autofocus: false,
   };
 
-  constructor(props) {
-    super(props);
-    this.state = this.getStateFromProps(props);
-  }
-
-  componentWillReceiveProps(nextProps) {
-    this.setState(this.getStateFromProps(nextProps));
-  }
-
-  getStateFromProps(props) {
-    const formData = Array.isArray(props.formData) ? props.formData : null;
-    const {definitions} = this.props.registry;
-    return {
-      items: getDefaultFormState(props.schema, formData, definitions) || []
-    };
-  }
-
-  shouldComponentUpdate(nextProps, nextState) {
-    return shouldRender(this, nextProps, nextState);
-  }
-
   get itemTitle() {
-    const {schema} = this.props;
+    const { schema } = this.props;
     return schema.items.title || schema.items.description || "Item";
   }
 
-  isItemRequired(itemsSchema) {
-    return itemsSchema.type === "string" && itemsSchema.minLength > 0;
+  isItemRequired(itemSchema) {
+    if (Array.isArray(itemSchema.type)) {
+      // While we don't yet support composite/nullable jsonschema types, it's
+      // future-proof to check for requirement against these.
+      return !itemSchema.type.includes("null");
+    }
+    // All non-null array item types are inherently required by design
+    return itemSchema.type !== "null";
   }
 
-  asyncSetState(state, options={validate: false}) {
-    setState(this, state, () => {
-      this.props.onChange(this.state.items, options);
-    });
+  canAddItem(formItems) {
+    const { schema, uiSchema } = this.props;
+    let { addable } = getUiOptions(uiSchema);
+    if (addable !== false) {
+      // if ui:options.addable was not explicitly set to false, we can add
+      // another item if we have not exceeded maxItems yet
+      if (schema.maxItems !== undefined) {
+        addable = formItems.length < schema.maxItems;
+      } else {
+        addable = true;
+      }
+    }
+    return addable;
   }
 
-  onAddClick = (event) => {
+  onAddClick = event => {
     event.preventDefault();
-    const {items} = this.state;
-    const {schema, registry} = this.props;
-    const {definitions} = registry;
+    const { schema, formData, registry = getDefaultRegistry() } = this.props;
+    const { definitions } = registry;
     let itemSchema = schema.items;
     if (isFixedItems(schema) && allowAdditionalItems(schema)) {
       itemSchema = schema.additionalItems;
     }
-    this.asyncSetState({
-      items: items.concat([
-        getDefaultFormState(itemSchema, undefined, definitions)
-      ])
-    });
+    this.props.onChange(
+      [...formData, getDefaultFormState(itemSchema, undefined, definitions)],
+      { validate: false }
+    );
   };
 
-  onDropIndexClick = (index) => {
-    return (event) => {
-      event.preventDefault();
-      this.asyncSetState({
-        items: this.state.items.filter((_, i) => i !== index)
-      }, {validate: true}); // refs #195
+  onDropIndexClick = index => {
+    return event => {
+      if (event) {
+        event.preventDefault();
+      }
+      const { formData, onChange } = this.props;
+      // refs #195: revalidate to ensure properly reindexing errors
+      onChange(formData.filter((_, i) => i !== index), { validate: true });
     };
   };
 
   onReorderClick = (index, newIndex) => {
-    return (event) => {
-      event.preventDefault();
-      event.target.blur();
-      const {items} = this.state;
-      this.asyncSetState({
-        items: items.map((item, i) => {
+    return event => {
+      if (event) {
+        event.preventDefault();
+        event.target.blur();
+      }
+      const { formData, onChange } = this.props;
+      onChange(
+        formData.map((item, i) => {
           if (i === newIndex) {
-            return items[index];
+            return formData[index];
           } else if (i === index) {
-            return items[newIndex];
+            return formData[newIndex];
           } else {
             return item;
           }
-        })
-      }, {validate: true});
+        }),
+        { validate: true }
+      );
     };
   };
 
-  onChangeForIndex = (index) => {
-    return (value) => {
-      this.asyncSetState({
-        items: this.state.items.map((item, i) => {
-          return index === i ? value : item;
-        })
+  onChangeForIndex = index => {
+    return value => {
+      const { formData, onChange } = this.props;
+      const newFormData = formData.map((item, i) => {
+        // We need to treat undefined items as nulls to have validation.
+        // See https://github.com/tdegrunt/jsonschema/issues/206
+        const jsonValue = typeof value === "undefined" ? null : value;
+        return index === i ? jsonValue : item;
       });
+      onChange(newFormData, { validate: false });
     };
   };
 
-  onSelectChange = (value) => {
-    this.asyncSetState({items: value});
+  onSelectChange = value => {
+    this.props.onChange(value, { validate: false });
   };
 
   render() {
-    const {schema, uiSchema} = this.props;
-    if (isFilesArray(schema, uiSchema)) {
-      return this.renderFiles();
-    }
+    const { schema, uiSchema, registry = getDefaultRegistry() } = this.props;
+    const { definitions } = registry;
     if (isFixedItems(schema)) {
       return this.renderFixedArray();
     }
-    if (isMultiSelect(schema)) {
+    if (isFilesArray(schema, uiSchema, definitions)) {
+      return this.renderFiles();
+    }
+    if (isMultiSelect(schema, definitions)) {
       return this.renderMultiSelect();
     }
     return this.renderNormalArray();
@@ -156,91 +293,128 @@ class ArrayField extends Component {
     const {
       schema,
       uiSchema,
+      formData,
       errorSchema,
       idSchema,
       name,
       required,
       disabled,
-      readonly
+      readonly,
+      autofocus,
+      registry = getDefaultRegistry(),
+      formContext,
+      onBlur,
     } = this.props;
-    const title = (schema.title === undefined) ? name : schema.title;
-    const {items} = this.state;
-    const {definitions, fields} = this.props.registry;
-    const {TitleField, DescriptionField} = fields;
+    const title = schema.title === undefined ? name : schema.title;
+    const { ArrayFieldTemplate, definitions, fields } = registry;
+    const { TitleField, DescriptionField } = fields;
     const itemsSchema = retrieveSchema(schema.items, definitions);
+    const arrayProps = {
+      canAdd: this.canAddItem(formData),
+      items: formData.map((item, index) => {
+        const itemErrorSchema = errorSchema ? errorSchema[index] : undefined;
+        const itemIdPrefix = idSchema.$id + "_" + index;
+        const itemIdSchema = toIdSchema(itemsSchema, itemIdPrefix, definitions);
+        return this.renderArrayFieldItem({
+          index,
+          canMoveUp: index > 0,
+          canMoveDown: index < formData.length - 1,
+          itemSchema: itemsSchema,
+          itemIdSchema,
+          itemErrorSchema,
+          itemData: formData[index],
+          itemUiSchema: uiSchema.items,
+          autofocus: autofocus && index === 0,
+          onBlur,
+        });
+      }),
+      className: `field field-array field-array-of-${itemsSchema.type}`,
+      DescriptionField,
+      disabled,
+      idSchema,
+      uiSchema,
+      onAddClick: this.onAddClick,
+      readonly,
+      required,
+      schema,
+      title,
+      TitleField,
+      formContext,
+    };
 
-    return (
-      <fieldset
-        className={`field field-array field-array-of-${itemsSchema.type}`}>
-        <ArrayFieldTitle
-          TitleField={TitleField}
-          idSchema={idSchema}
-          title={title}
-          required={required} />
-        {schema.description ?
-          <ArrayFieldDescription
-            DescriptionField={DescriptionField}
-            idSchema={idSchema}
-            description={schema.description} /> : null}
-        <div className="row array-item-list">{
-          items.map((item, index) => {
-            const itemErrorSchema = errorSchema ? errorSchema[index] : undefined;
-            const itemIdPrefix = idSchema.$id + "_" + index;
-            const itemIdSchema = toIdSchema(itemsSchema, itemIdPrefix, definitions);
-            return this.renderArrayFieldItem({
-              index,
-              canMoveUp: index > 0,
-              canMoveDown: index < items.length - 1,
-              itemSchema: itemsSchema,
-              itemIdSchema,
-              itemErrorSchema,
-              itemData: items[index],
-              itemUiSchema: uiSchema.items
-            });
-          })
-        }</div>
-        <AddButton
-          onClick={this.onAddClick} disabled={disabled || readonly} />
-      </fieldset>
-    );
+    // Check if a custom render function was passed in
+    const Component = ArrayFieldTemplate || DefaultNormalArrayFieldTemplate;
+    return <Component {...arrayProps} />;
   }
 
   renderMultiSelect() {
-    const {schema, idSchema, uiSchema, disabled, readonly} = this.props;
-    const {items} = this.state;
-    const {definitions} = this.props.registry;
+    const {
+      schema,
+      idSchema,
+      uiSchema,
+      disabled,
+      readonly,
+      autofocus,
+      onBlur,
+      registry = getDefaultRegistry(),
+    } = this.props;
+    const items = this.props.formData;
+    const { widgets, definitions, formContext } = registry;
     const itemsSchema = retrieveSchema(schema.items, definitions);
-
-    const multipleCheckboxes = uiSchema["ui:widget"] === "checkboxes";
-    const Widget = (multipleCheckboxes) ? CheckboxesWidget : SelectWidget;
+    const enumOptions = optionsList(itemsSchema);
+    const { widget = "select", ...options } = {
+      ...getUiOptions(uiSchema),
+      enumOptions,
+    };
+    const Widget = getWidget(schema, widget, widgets);
     return (
       <Widget
         id={idSchema && idSchema.$id}
         multiple
         onChange={this.onSelectChange}
-        options={{enumOptions: optionsList(itemsSchema)}}
+        onBlur={onBlur}
+        options={options}
         schema={schema}
         value={items}
         disabled={disabled}
         readonly={readonly}
+        formContext={formContext}
+        autofocus={autofocus}
       />
     );
   }
 
   renderFiles() {
-    const {schema, idSchema, name, disabled, readonly} = this.props;
+    const {
+      schema,
+      uiSchema,
+      idSchema,
+      name,
+      disabled,
+      readonly,
+      autofocus,
+      onBlur,
+      registry = getDefaultRegistry(),
+    } = this.props;
     const title = schema.title || name;
-    const {items} = this.state;
+    const items = this.props.formData;
+    const { widgets, formContext } = registry;
+    const { widget = "files", ...options } = getUiOptions(uiSchema);
+    const Widget = getWidget(schema, widget, widgets);
     return (
-      <FileWidget
+      <Widget
+        options={options}
         id={idSchema && idSchema.$id}
         multiple
         onChange={this.onSelectChange}
+        onBlur={onBlur}
         schema={schema}
         title={title}
         value={items}
         disabled={disabled}
         readonly={readonly}
+        formContext={formContext}
+        autofocus={autofocus}
       />
     );
   }
@@ -254,16 +428,21 @@ class ArrayField extends Component {
       name,
       required,
       disabled,
-      readonly
+      readonly,
+      autofocus,
+      registry = getDefaultRegistry(),
+      onBlur,
     } = this.props;
     const title = schema.title || name;
-    let {items} = this.state;
-    const {definitions, fields} = this.props.registry;
-    const {TitleField} = fields;
+    let items = this.props.formData;
+    const { ArrayFieldTemplate, definitions, fields } = registry;
+    const { TitleField } = fields;
     const itemSchemas = schema.items.map(item =>
-      retrieveSchema(item, definitions));
-    const additionalSchema = allowAdditionalItems(schema) ?
-      retrieveSchema(schema.additionalItems, definitions) : null;
+      retrieveSchema(item, definitions)
+    );
+    const additionalSchema = allowAdditionalItems(schema)
+      ? retrieveSchema(schema.additionalItems, definitions)
+      : null;
 
     if (!items || items.length < itemSchemas.length) {
       // to make sure at least all fixed items are generated
@@ -271,122 +450,128 @@ class ArrayField extends Component {
       items = items.concat(new Array(itemSchemas.length - items.length));
     }
 
-    return (
-      <fieldset className="field field-array field-array-fixed-items">
-        <ArrayFieldTitle
-          TitleField={TitleField}
-          idSchema={idSchema}
-          title={title}
-          required={required} />
-        {schema.description ?
-          <div className="field-description">{schema.description}</div> : null}
-        <div className="row array-item-list">{
-          items.map((item, index) => {
-            const additional = index >= itemSchemas.length;
-            const itemSchema = additional ?
-              additionalSchema : itemSchemas[index];
-            const itemIdPrefix = idSchema.$id + "_" + index;
-            const itemIdSchema = toIdSchema(itemSchema, itemIdPrefix, definitions);
-            const itemUiSchema = additional ?
-              uiSchema.additionalItems || {} :
-              Array.isArray(uiSchema.items) ?
-                uiSchema.items[index] : uiSchema.items || {};
-            const itemErrorSchema = errorSchema ? errorSchema[index] : undefined;
+    // These are the props passed into the render function
+    const arrayProps = {
+      canAdd: this.canAddItem(items) && additionalSchema,
+      className: "field field-array field-array-fixed-items",
+      disabled,
+      idSchema,
+      items: items.map((item, index) => {
+        const additional = index >= itemSchemas.length;
+        const itemSchema = additional ? additionalSchema : itemSchemas[index];
+        const itemIdPrefix = idSchema.$id + "_" + index;
+        const itemIdSchema = toIdSchema(itemSchema, itemIdPrefix, definitions);
+        const itemUiSchema = additional
+          ? uiSchema.additionalItems || {}
+          : Array.isArray(uiSchema.items)
+            ? uiSchema.items[index]
+            : uiSchema.items || {};
+        const itemErrorSchema = errorSchema ? errorSchema[index] : undefined;
 
-            return this.renderArrayFieldItem({
-              index,
-              removable: additional,
-              canMoveUp: index >= itemSchemas.length + 1,
-              canMoveDown: additional && index < items.length - 1,
-              itemSchema,
-              itemData: item,
-              itemUiSchema,
-              itemIdSchema,
-              itemErrorSchema
-            });
-          })
-        }</div>
-        {
-          additionalSchema ? <AddButton
-                               onClick={this.onAddClick}
-                               disabled={disabled || readonly} /> : null
-        }
-      </fieldset>
-    );
+        return this.renderArrayFieldItem({
+          index,
+          canRemove: additional,
+          canMoveUp: index >= itemSchemas.length + 1,
+          canMoveDown: additional && index < items.length - 1,
+          itemSchema,
+          itemData: item,
+          itemUiSchema,
+          itemIdSchema,
+          itemErrorSchema,
+          autofocus: autofocus && index === 0,
+          onBlur,
+        });
+      }),
+      onAddClick: this.onAddClick,
+      readonly,
+      required,
+      schema,
+      uiSchema,
+      title,
+      TitleField,
+    };
+
+    // Check if a custom template template was passed in
+    const Template = ArrayFieldTemplate || DefaultFixedArrayFieldTemplate;
+    return <Template {...arrayProps} />;
   }
 
-  renderArrayFieldItem({
-    index,
-    removable=true,
-    canMoveUp=true,
-    canMoveDown=true,
-    itemSchema,
-    itemData,
-    itemUiSchema,
-    itemIdSchema,
-    itemErrorSchema
-  }) {
-    const {SchemaField} = this.props.registry.fields;
-    const {disabled, readonly} = this.props;
-    const hasToolbar = removable || canMoveUp || canMoveDown;
-    const btnStyle = {flex: 1, paddingLeft: 6, paddingRight: 6, fontWeight: "bold"};
+  renderArrayFieldItem(props) {
+    const {
+      index,
+      canRemove = true,
+      canMoveUp = true,
+      canMoveDown = true,
+      itemSchema,
+      itemData,
+      itemUiSchema,
+      itemIdSchema,
+      itemErrorSchema,
+      autofocus,
+      onBlur,
+    } = props;
+    const {
+      disabled,
+      readonly,
+      uiSchema,
+      registry = getDefaultRegistry(),
+    } = this.props;
+    const { fields: { SchemaField } } = registry;
+    const { orderable, removable } = {
+      orderable: true,
+      removable: true,
+      ...uiSchema["ui:options"],
+    };
+    const has = {
+      moveUp: orderable && canMoveUp,
+      moveDown: orderable && canMoveDown,
+      remove: removable && canRemove,
+    };
+    has.toolbar = Object.keys(has).some(key => has[key]);
 
-    return (
-      <div key={index} className="array-item">
-        <div className={hasToolbar ? "col-xs-10" : "col-xs-12"}>
-          <SchemaField
-            schema={itemSchema}
-            uiSchema={itemUiSchema}
-            formData={itemData}
-            errorSchema={itemErrorSchema}
-            idSchema={itemIdSchema}
-            required={this.isItemRequired(itemSchema)}
-            onChange={this.onChangeForIndex(index)}
-            registry={this.props.registry}
-            disabled={this.props.disabled}
-            readonly={this.props.readonly} />
-        </div>
-        {
-          hasToolbar ?
-            <div className="col-xs-2 array-item-toolbox text-right">
-              <div className="btn-group" style={{display: "flex"}}>
-                {canMoveUp || canMoveDown ?
-                  <button type="button" className="btn btn-default array-item-move-up"
-                          style={btnStyle}
-                          tabIndex="-1"
-                          disabled={disabled || readonly || !canMoveUp}
-                          onClick={this.onReorderClick(index, index - 1)}>⬆</button>
-                  : null}
-                {canMoveUp || canMoveDown ?
-                  <button type="button" className="btn btn-default array-item-move-down"
-                          style={btnStyle}
-                          tabIndex="-1"
-                          disabled={disabled || readonly || !canMoveDown}
-                          onClick={this.onReorderClick(index, index + 1)}>⬇</button>
-                  : null}
-                {removable ?
-                  <button type="button" className="btn btn-danger array-item-remove"
-                          style={btnStyle}
-                          tabIndex="-1"
-                          disabled={disabled || readonly}
-                          onClick={this.onDropIndexClick(index)}>✖</button>
-                  : null}
-              </div>
-            </div>
-          : null
-        }
-      </div>
-    );
+    return {
+      children: (
+        <SchemaField
+          schema={itemSchema}
+          uiSchema={itemUiSchema}
+          formData={itemData}
+          errorSchema={itemErrorSchema}
+          idSchema={itemIdSchema}
+          required={this.isItemRequired(itemSchema)}
+          onChange={this.onChangeForIndex(index)}
+          onBlur={onBlur}
+          registry={this.props.registry}
+          disabled={this.props.disabled}
+          readonly={this.props.readonly}
+          autofocus={autofocus}
+        />
+      ),
+      className: "array-item",
+      disabled,
+      hasToolbar: has.toolbar,
+      hasMoveUp: has.moveUp,
+      hasMoveDown: has.moveDown,
+      hasRemove: has.remove,
+      index,
+      onDropIndexClick: this.onDropIndexClick,
+      onReorderClick: this.onReorderClick,
+      readonly,
+    };
   }
 }
 
-function AddButton({onClick, disabled}) {
+function AddButton({ onClick, disabled }) {
   return (
     <div className="row">
-      <p className="col-xs-2 col-xs-offset-10 array-item-add text-right">
-        <button type="button" className="btn btn-info col-xs-12"
-                tabIndex="-1" onClick={onClick}
-                disabled={disabled} style={{fontWeight: "bold"}}>➕</button>
+      <p className="col-xs-3 col-xs-offset-9 array-item-add text-right">
+        <IconBtn
+          type="info"
+          icon="plus"
+          className="btn-add col-xs-12"
+          tabIndex="0"
+          onClick={onClick}
+          disabled={disabled}
+        />
       </p>
     </div>
   );
@@ -395,22 +580,29 @@ function AddButton({onClick, disabled}) {
 if (process.env.NODE_ENV !== "production") {
   ArrayField.propTypes = {
     schema: PropTypes.object.isRequired,
-    uiSchema: PropTypes.object,
+    uiSchema: PropTypes.shape({
+      "ui:options": PropTypes.shape({
+        addable: PropTypes.bool,
+        orderable: PropTypes.bool,
+        removable: PropTypes.bool,
+      }),
+    }),
     idSchema: PropTypes.object,
     errorSchema: PropTypes.object,
     onChange: PropTypes.func.isRequired,
+    onBlur: PropTypes.func,
     formData: PropTypes.array,
     required: PropTypes.bool,
     disabled: PropTypes.bool,
     readonly: PropTypes.bool,
+    autofocus: PropTypes.bool,
     registry: PropTypes.shape({
-      widgets: PropTypes.objectOf(PropTypes.oneOfType([
-        PropTypes.func,
-        PropTypes.object,
-      ])).isRequired,
+      widgets: PropTypes.objectOf(
+        PropTypes.oneOfType([PropTypes.func, PropTypes.object])
+      ).isRequired,
       fields: PropTypes.objectOf(PropTypes.func).isRequired,
       definitions: PropTypes.object.isRequired,
-      formContext: PropTypes.object.isRequired
+      formContext: PropTypes.object.isRequired,
     }),
   };
 }
