@@ -1,24 +1,31 @@
-import React, {PropTypes} from "react";
+import React from "react";
+import PropTypes from "prop-types";
 
 function DescriptionField(props) {
-  const {id, description} = props;
+  const { id, description } = props;
   if (!description) {
-    return null;
+    // See #312: Ensure compatibility with old versions of React.
+    return <div />;
   }
   if (typeof description === "string") {
-    return <p id={id} className="field-description">{description}</p>;
+    return (
+      <p id={id} className="field-description">
+        {description}
+      </p>
+    );
   } else {
-    return <div id={id} className="field-description">{description}</div>;
+    return (
+      <div id={id} className="field-description">
+        {description}
+      </div>
+    );
   }
 }
 
 if (process.env.NODE_ENV !== "production") {
   DescriptionField.propTypes = {
     id: PropTypes.string,
-    description: PropTypes.oneOfType([
-      PropTypes.string,
-      PropTypes.element,
-    ])
+    description: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
   };
 }
 
