@@ -1,11 +1,16 @@
-import React, {PropTypes} from "react";
+import React from "react";
+import PropTypes from "prop-types";
 
 const REQUIRED_FIELD_SYMBOL = "*";
 
 function TitleField(props) {
-  const {id, title, required} = props;
+  const { id, title, required } = props;
   const legend = required ? title + REQUIRED_FIELD_SYMBOL : title;
-  return <legend id={id}>{legend}</legend>;
+  return (
+    <legend id={id}>
+      {legend}
+    </legend>
+  );
 }
 
 if (process.env.NODE_ENV !== "production") {
