@@ -1,7 +1,6 @@
 import React from "react";
 
-
-export default function UnsupportedField({schema}) {
+export default function UnsupportedField({ schema }) {
   // XXX render json as string so dev can inspect faulty subschema
   return (
     <div className="unsupported-field">
