@@ -1,5 +1,5 @@
-import React, { PropTypes } from "react";
-
+import React from "react";
+import PropTypes from "prop-types";
 
 function BaseInput(props) {
   // Note: since React 15.2.0 we can't forward unknown element attributes, so we
@@ -7,19 +7,31 @@ function BaseInput(props) {
   const {
     value,
     readonly,
-    onChange,
-    options,  // eslint-disable-line
-    schema,   // eslint-disable-line
-    formContext,  // eslint-disable-line
+    disabled,
+    autofocus,
+    onBlur,
+    options,
+    schema,
+    formContext,
+    registry,
     ...inputProps
   } = props;
+
+  inputProps.type = options.inputType || inputProps.type || "text";
+  const _onChange = ({ target: { value } }) => {
+    return props.onChange(value === "" ? options.emptyValue : value);
+  };
   return (
     <input
-      {...inputProps}
       className="form-control"
       readOnly={readonly}
-      value={typeof value === "undefined" ? "" : value}
-      onChange={(event) => onChange(event.target.value)} />
+      disabled={disabled}
+      autoFocus={autofocus}
+      value={value == null ? "" : value}
+      {...inputProps}
+      onChange={_onChange}
+      onBlur={onBlur && (event => onBlur(inputProps.id, event.target.value))}
+    />
   );
 }
 
@@ -28,6 +40,7 @@ BaseInput.defaultProps = {
   required: false,
   disabled: false,
   readonly: false,
+  autofocus: false,
 };
 
 if (process.env.NODE_ENV !== "production") {
@@ -38,7 +51,9 @@ if (process.env.NODE_ENV !== "production") {
     required: PropTypes.bool,
     disabled: PropTypes.bool,
     readonly: PropTypes.bool,
+    autofocus: PropTypes.bool,
     onChange: PropTypes.func,
+    onBlur: PropTypes.func,
   };
 }
 
