@@ -1,36 +1,55 @@
-import React, { PropTypes } from "react";
+import React from "react";
+import PropTypes from "prop-types";
+import DescriptionField from "../fields/DescriptionField.js";
 
-
-function CheckboxWidget({
-  schema,
-  id,
-  value,
-  required,
-  disabled,
-  onChange,
-  label,
-}) {
+function CheckboxWidget(props) {
+  const {
+    schema,
+    id,
+    value,
+    required,
+    disabled,
+    readonly,
+    label,
+    autofocus,
+    onChange,
+  } = props;
   return (
-    <div className={`checkbox ${disabled ? "disabled" : ""}`}>
+    <div className={`checkbox ${disabled || readonly ? "disabled" : ""}`}>
+      {schema.description &&
+        <DescriptionField description={schema.description} />}
       <label>
-        <input type="checkbox"
+        <input
+          type="checkbox"
           id={id}
           checked={typeof value === "undefined" ? false : value}
           required={required}
-          disabled={disabled}
-          onChange={(event) => onChange(event.target.checked)} />
-        <strong>{label}</strong>
+          disabled={disabled || readonly}
+          autoFocus={autofocus}
+          onChange={event => onChange(event.target.checked)}
+        />
+        <span>
+          {label}
+        </span>
       </label>
     </div>
   );
 }
+
+CheckboxWidget.defaultProps = {
+  autofocus: false,
+};
+
 if (process.env.NODE_ENV !== "production") {
   CheckboxWidget.propTypes = {
     schema: PropTypes.object.isRequired,
     id: PropTypes.string.isRequired,
-    onChange: PropTypes.func,
     value: PropTypes.bool,
     required: PropTypes.bool,
+    disabled: PropTypes.bool,
+    readonly: PropTypes.bool,
+    autofocus: PropTypes.bool,
+    onChange: PropTypes.func,
   };
 }
 
