@@ -1,5 +1,5 @@
-import React, { PropTypes } from "react";
-
+import React from "react";
+import PropTypes from "prop-types";
 
 function selectValue(value, selected, all) {
   const at = all.indexOf(value);
@@ -14,35 +14,55 @@ function deselectValue(value, selected) {
 }
 
 function CheckboxesWidget(props) {
-  const {id, disabled, options, value, onChange} = props;
-  const {enumOptions} = options;
+  const { id, disabled, options, value, autofocus, readonly, onChange } = props;
+  const { enumOptions, inline } = options;
   return (
-    <div className="checkboxes" id={id}>{
-      enumOptions.map((option, index) => {
+    <div className="checkboxes" id={id}>
+      {enumOptions.map((option, index) => {
         const checked = value.indexOf(option.value) !== -1;
-        return (
-          <div key={index} className="checkbox">
-            <label>
-              <input type="checkbox"
-                id={`${id}_${index}`}
-                checked={checked}
-                disabled={disabled}
-                onChange={(event) => {
-                  const all = enumOptions.map(({value}) => value);
-                  if (event.target.checked) {
-                    onChange(selectValue(option.value, value, all));
-                  } else {
-                    onChange(deselectValue(option.value, value));
-                  }
-                }} />
-              <strong>{option.label}</strong>
-            </label>
-          </div>
+        const disabledCls = disabled || readonly ? "disabled" : "";
+        const checkbox = (
+          <span>
+            <input
+              type="checkbox"
+              id={`${id}_${index}`}
+              checked={checked}
+              disabled={disabled || readonly}
+              autoFocus={autofocus && index === 0}
+              onChange={event => {
+                const all = enumOptions.map(({ value }) => value);
+                if (event.target.checked) {
+                  onChange(selectValue(option.value, value, all));
+                } else {
+                  onChange(deselectValue(option.value, value));
+                }
+              }}
+            />
+            <span>
+              {option.label}
+            </span>
+          </span>
         );
-      })
-    }</div>
+        return inline
+          ? <label key={index} className={`checkbox-inline ${disabledCls}`}>
+              {checkbox}
+            </label>
+          : <div key={index} className={`checkbox ${disabledCls}`}>
+              <label>
+                {checkbox}
+              </label>
+            </div>;
+      })}
+    </div>
   );
 }
+
+CheckboxesWidget.defaultProps = {
+  autofocus: false,
+  options: {
+    inline: false,
+  },
+};
 
 if (process.env.NODE_ENV !== "production") {
   CheckboxesWidget.propTypes = {
@@ -50,10 +70,14 @@ if (process.env.NODE_ENV !== "production") {
     id: PropTypes.string.isRequired,
     options: PropTypes.shape({
       enumOptions: PropTypes.array,
+      inline: PropTypes.bool,
     }).isRequired,
     value: PropTypes.any,
     required: PropTypes.bool,
+    readonly: PropTypes.bool,
+    disabled: PropTypes.bool,
     multiple: PropTypes.bool,
+    autofocus: PropTypes.bool,
     onChange: PropTypes.func,
   };
 }
