@@ -1,14 +1,20 @@
-import React, { PropTypes } from "react";
+import React from "react";
+import PropTypes from "prop-types";
 
 import { asNumber } from "../../utils";
-
 
 /**
  * This is a silly limitation in the DOM where option change event values are
  * always retrieved as strings.
  */
-function processValue({type, items}, value) {
-  if (type === "array" && items && ["number", "integer"].includes(items.type)) {
+function processValue({ type, items }, value) {
+  if (value === "") {
+    return undefined;
+  } else if (
+    type === "array" &&
+    items &&
+    ["number", "integer"].includes(items.type)
+  ) {
     return value.map(asNumber);
   } else if (type === "boolean") {
     return value === "true";
@@ -18,43 +24,73 @@ function processValue({type, items}, value) {
   return value;
 }
 
-function SelectWidget({
-  schema,
-  id,
-  options,
-  value,
-  required,
-  disabled,
-  readonly,
-  multiple,
-  onChange
-}) {
-  const {enumOptions} = options;
+function getValue(event, multiple) {
+  if (multiple) {
+    return [].slice
+      .call(event.target.options)
+      .filter(o => o.selected)
+      .map(o => o.value);
+  } else {
+    return event.target.value;
+  }
+}
+
+function SelectWidget(props) {
+  const {
+    schema,
+    id,
+    options,
+    value,
+    required,
+    disabled,
+    readonly,
+    multiple,
+    autofocus,
+    onChange,
+    onBlur,
+    placeholder,
+  } = props;
+  const { enumOptions } = options;
+  const emptyValue = multiple ? [] : "";
   return (
     <select
       id={id}
       multiple={multiple}
       className="form-control"
-      value={value}
+      value={typeof value === "undefined" ? emptyValue : value}
       required={required}
-      disabled={disabled}
-      readOnly={readonly}
-      onChange={(event) => {
-        let newValue;
-        if (multiple) {
-          newValue = [].filter.call(
-            event.target.options, o => o.selected).map(o => o.value);
-        } else {
-          newValue = event.target.value;
-        }
+      disabled={disabled || readonly}
+      autoFocus={autofocus}
+      onBlur={
+        onBlur &&
+        (event => {
+          const newValue = getValue(event, multiple);
+          onBlur(id, processValue(schema, newValue));
+        })
+      }
+      onChange={event => {
+        const newValue = getValue(event, multiple);
         onChange(processValue(schema, newValue));
-      }}>{
-      enumOptions.map(({value, label}, i) => {
-        return <option key={i} value={value}>{label}</option>;
-      })
-    }</select>
+      }}>
+      {!multiple &&
+        !schema.default &&
+        <option value="">
+          {placeholder}
+        </option>}
+      {enumOptions.map(({ value, label }, i) => {
+        return (
+          <option key={i} value={value}>
+            {label}
+          </option>
+        );
+      })}
+    </select>
   );
 }
+
+SelectWidget.defaultProps = {
+  autofocus: false,
+};
 
 if (process.env.NODE_ENV !== "production") {
   SelectWidget.propTypes = {
@@ -65,8 +101,12 @@ if (process.env.NODE_ENV !== "production") {
     }).isRequired,
     value: PropTypes.any,
     required: PropTypes.bool,
+    disabled: PropTypes.bool,
+    readonly: PropTypes.bool,
     multiple: PropTypes.bool,
+    autofocus: PropTypes.bool,
     onChange: PropTypes.func,
+    onBlur: PropTypes.func,
   };
 }
 
