@@ -1,116 +1,82 @@
+import React from "react";
 import "setimmediate";
 
-import TitleField from "./components/fields/TitleField";
-import DescriptionField from "./components/fields/DescriptionField";
-import PasswordWidget from "./components/widgets/PasswordWidget";
-import RadioWidget from "./components/widgets/RadioWidget";
-import UpDownWidget from "./components/widgets/UpDownWidget";
-import RangeWidget from "./components/widgets/RangeWidget";
-import SelectWidget from "./components/widgets/SelectWidget";
-import TextWidget from "./components/widgets/TextWidget";
-import DateWidget from "./components/widgets/DateWidget";
-import DateTimeWidget from "./components/widgets/DateTimeWidget";
-import AltDateWidget from "./components/widgets/AltDateWidget";
-import AltDateTimeWidget from "./components/widgets/AltDateTimeWidget";
-import EmailWidget from "./components/widgets/EmailWidget";
-import URLWidget from "./components/widgets/URLWidget";
-import TextareaWidget from "./components/widgets/TextareaWidget";
-import HiddenWidget from "./components/widgets/HiddenWidget";
-import ColorWidget from "./components/widgets/ColorWidget";
-import FileWidget from "./components/widgets/FileWidget";
-import CheckboxesWidget from "./components/widgets/CheckboxesWidget";
-
-
-
-const altWidgetMap = {
+const widgetMap = {
   boolean: {
-    radio: RadioWidget,
-    select: SelectWidget,
-    hidden: HiddenWidget,
+    checkbox: "CheckboxWidget",
+    radio: "RadioWidget",
+    select: "SelectWidget",
+    hidden: "HiddenWidget",
   },
   string: {
-    password: PasswordWidget,
-    radio: RadioWidget,
-    select: SelectWidget,
-    textarea: TextareaWidget,
-    hidden: HiddenWidget,
-    date: DateWidget,
-    datetime: DateTimeWidget,
-    "alt-date": AltDateWidget,
-    "alt-datetime": AltDateTimeWidget,
-    color: ColorWidget,
-    file: FileWidget,
+    text: "TextWidget",
+    password: "PasswordWidget",
+    email: "EmailWidget",
+    hostname: "TextWidget",
+    ipv4: "TextWidget",
+    ipv6: "TextWidget",
+    uri: "URLWidget",
+    "data-url": "FileWidget",
+    radio: "RadioWidget",
+    select: "SelectWidget",
+    textarea: "TextareaWidget",
+    hidden: "HiddenWidget",
+    date: "DateWidget",
+    datetime: "DateTimeWidget",
+    "date-time": "DateTimeWidget",
+    "alt-date": "AltDateWidget",
+    "alt-datetime": "AltDateTimeWidget",
+    color: "ColorWidget",
+    file: "FileWidget",
   },
   number: {
-    updown: UpDownWidget,
-    range: RangeWidget,
-    hidden: HiddenWidget,
+    text: "TextWidget",
+    select: "SelectWidget",
+    updown: "UpDownWidget",
+    range: "RangeWidget",
+    radio: "RadioWidget",
+    hidden: "HiddenWidget",
   },
   integer: {
-    updown: UpDownWidget,
-    range: RangeWidget,
-    hidden: HiddenWidget,
+    text: "TextWidget",
+    select: "SelectWidget",
+    updown: "UpDownWidget",
+    range: "RangeWidget",
+    radio: "RadioWidget",
+    hidden: "HiddenWidget",
   },
   array: {
-    checkboxes: CheckboxesWidget,
-  }
-};
-
-const stringFormatWidgets = {
-  "date-time": DateTimeWidget,
-  "date": DateWidget,
-  "email": EmailWidget,
-  "hostname": TextWidget,
-  "ipv4": TextWidget,
-  "ipv6": TextWidget,
-  "uri": URLWidget,
-  "data-url": FileWidget,
+    select: "SelectWidget",
+    checkboxes: "CheckboxesWidget",
+    files: "FileWidget",
+  },
 };
 
 export function getDefaultRegistry() {
   return {
-    fields: {
-      // Prevent a bug where SchemaField is undefined when imported via Babel.
-      // This seems to have been introduced when upgrading React from 0.14 to to
-      // 15.0, which now seems to prevent cyclic references of exported
-      // components.
-      // Investigation hint: getDefaultRegistry is called from within
-      // SchemaField itself.
-      SchemaField: require("./components/fields/SchemaField").default,
-      TitleField,
-      DescriptionField,
-    },
-    widgets: {},
+    fields: require("./components/fields").default,
+    widgets: require("./components/widgets").default,
     definitions: {},
-    formContext: {}
+    formContext: {},
   };
 }
 
-export function defaultFieldValue(formData, schema) {
-  return typeof formData === "undefined" ? schema.default : formData;
-}
+export function getWidget(schema, widget, registeredWidgets = {}) {
+  const { type } = schema;
 
-export function getAlternativeWidget(
-  schema,
-  widget,
-  registeredWidgets={},
-  widgetOptions={}
-) {
-  const {type, format} = schema;
-
-  function setDefaultOptions(widget) {
-    widget.defaultProps = {...widget.defaultProps, options: widgetOptions};
-    return widget;
+  function mergeOptions(Widget) {
+    // cache return value as property of widget for proper react reconciliation
+    if (!Widget.MergedWidget) {
+      const defaultOptions =
+        (Widget.defaultProps && Widget.defaultProps.options) || {};
+      Widget.MergedWidget = ({ options = {}, ...props }) =>
+        <Widget options={{ ...defaultOptions, ...options }} {...props} />;
+    }
+    return Widget.MergedWidget;
   }
 
   if (typeof widget === "function") {
-    return setDefaultOptions(widget);
-  }
-
-  if (isObject(widget)) {
-    const {component, options} = widget;
-    const mergedOptions = {...options, ...widgetOptions};
-    return getAlternativeWidget(schema, component, registeredWidgets, mergedOptions);
+    return mergeOptions(widget);
   }
 
   if (typeof widget !== "string") {
@@ -119,28 +85,22 @@ export function getAlternativeWidget(
 
   if (registeredWidgets.hasOwnProperty(widget)) {
     const registeredWidget = registeredWidgets[widget];
-    return getAlternativeWidget(schema, registeredWidget, registeredWidgets, widgetOptions);
+    return getWidget(schema, registeredWidget, registeredWidgets);
   }
 
-  if (!altWidgetMap.hasOwnProperty(type)) {
-    throw new Error(`No alternative widget for type ${type}`);
+  if (!widgetMap.hasOwnProperty(type)) {
+    throw new Error(`No widget for type "${type}"`);
   }
 
-  if (altWidgetMap[type].hasOwnProperty(widget)) {
-    const altWidget = altWidgetMap[type][widget];
-    return getAlternativeWidget(schema, altWidget, registeredWidgets, widgetOptions);
+  if (widgetMap[type].hasOwnProperty(widget)) {
+    const registeredWidget = registeredWidgets[widgetMap[type][widget]];
+    return getWidget(schema, registeredWidget, registeredWidgets);
   }
 
-  if (type === "string" && stringFormatWidgets.hasOwnProperty(format)) {
-    const stringFormatWidget = stringFormatWidgets[format];
-    return getAlternativeWidget(schema, stringFormatWidget, registeredWidgets, widgetOptions);
-  }
-
-  const info = type === "string" && format ? `/${format}` : "";
-  throw new Error(`No alternative widget "${widget}" for type ${type}${info}`);
+  throw new Error(`No widget "${widget}" for type "${type}"`);
 }
 
-function computeDefaults(schema, parentDefaults, definitions={}) {
+function computeDefaults(schema, parentDefaults, definitions = {}) {
   // Compute the defaults recursively: give highest priority to deepest nodes.
   let defaults = parentDefaults;
   if (isObject(defaults) && isObject(schema.default)) {
@@ -150,46 +110,83 @@ function computeDefaults(schema, parentDefaults, definitions={}) {
   } else if ("default" in schema) {
     // Use schema defaults for this node.
     defaults = schema.default;
-  } else if ("enum" in schema && Array.isArray(schema.enum)) {
-    // For enum with no defined default, select the first entry.
-    defaults = schema.enum[0];
   } else if ("$ref" in schema) {
     // Use referenced schema defaults for this node.
     const refSchema = findSchemaDefinition(schema.$ref, definitions);
     return computeDefaults(refSchema, defaults, definitions);
   } else if (isFixedItems(schema)) {
-    defaults = schema.items.map(itemSchema => computeDefaults(itemSchema, undefined, definitions));
+    defaults = schema.items.map(itemSchema =>
+      computeDefaults(itemSchema, undefined, definitions)
+    );
   }
   // Not defaults defined for this node, fallback to generic typed ones.
-  if (typeof(defaults) === "undefined") {
+  if (typeof defaults === "undefined") {
     defaults = schema.default;
   }
-  // We need to recur for object schema inner default values.
-  if (schema.type === "object") {
-    return Object.keys(schema.properties).reduce((acc, key) => {
-      // Compute the defaults for this node, with the parent defaults we might
-      // have from a previous run: defaults[key].
-      acc[key] = computeDefaults(
-        schema.properties[key], (defaults || {})[key], definitions);
-      return acc;
-    }, {});
+
+  switch (schema.type) {
+    // We need to recur for object schema inner default values.
+    case "object":
+      return Object.keys(schema.properties || {}).reduce((acc, key) => {
+        // Compute the defaults for this node, with the parent defaults we might
+        // have from a previous run: defaults[key].
+        acc[key] = computeDefaults(
+          schema.properties[key],
+          (defaults || {})[key],
+          definitions
+        );
+        return acc;
+      }, {});
+
+    case "array":
+      if (schema.minItems) {
+        return new Array(schema.minItems).fill(
+          computeDefaults(schema.items, defaults, definitions)
+        );
+      }
   }
   return defaults;
 }
 
-export function getDefaultFormState(_schema, formData, definitions={}) {
+export function getDefaultFormState(_schema, formData, definitions = {}) {
   if (!isObject(_schema)) {
     throw new Error("Invalid schema: " + _schema);
   }
   const schema = retrieveSchema(_schema, definitions);
   const defaults = computeDefaults(schema, _schema.default, definitions);
-  if (typeof(formData) === "undefined") { // No form data? Use schema defaults.
+  if (typeof formData === "undefined") {
+    // No form data? Use schema defaults.
     return defaults;
   }
-  if (isObject(formData)) { // Override schema defaults with form data.
+  if (isObject(formData)) {
+    // Override schema defaults with form data.
     return mergeObjects(defaults, formData);
   }
   return formData || defaults;
+}
+
+export function getUiOptions(uiSchema) {
+  // get all passed options from ui:widget, ui:options, and ui:<optionName>
+  return Object.keys(uiSchema)
+    .filter(key => key.indexOf("ui:") === 0)
+    .reduce((options, key) => {
+      const value = uiSchema[key];
+
+      if (key === "ui:widget" && isObject(value)) {
+        console.warn(
+          "Setting options via ui:widget object is deprecated, use ui:options instead"
+        );
+        return {
+          ...options,
+          ...(value.options || {}),
+          widget: value.component,
+        };
+      }
+      if (key === "ui:options" && isObject(value)) {
+        return { ...options, ...value };
+      }
+      return { ...options, [key.substring(3)]: value };
+    }, {});
 }
 
 export function isObject(thing) {
@@ -199,8 +196,9 @@ export function isObject(thing) {
 export function mergeObjects(obj1, obj2, concatArrays = false) {
   // Recursively merge deeply nested objects.
   var acc = Object.assign({}, obj1); // Prevent mutation of source object.
-  return Object.keys(obj2).reduce((acc, key) =>{
-    const left = obj1[key], right = obj2[key];
+  return Object.keys(obj2).reduce((acc, key) => {
+    const left = obj1[key],
+      right = obj2[key];
     if (obj1.hasOwnProperty(key) && isObject(right)) {
       acc[key] = mergeObjects(left, right, concatArrays);
     } else if (concatArrays && Array.isArray(left) && Array.isArray(right)) {
@@ -213,6 +211,9 @@ export function mergeObjects(obj1, obj2, concatArrays = false) {
 }
 
 export function asNumber(value) {
+  if (value === "") {
+    return undefined;
+  }
   if (/\.$/.test(value)) {
     // "3." can't really be considered a number even if it parses in js. The
     // user is most likely entering a float.
@@ -224,6 +225,14 @@ export function asNumber(value) {
   }
   const n = Number(value);
   const valid = typeof n === "number" && !Number.isNaN(n);
+
+  if (/\.\d*0$/.test(value)) {
+    // It's a number, that's cool - but we need it as a string so it doesn't screw
+    // with the user when entering dollar amounts or other values (such as those with
+    // specific precision or number of significant digits)
+    return value;
+  }
+
   return valid ? n : value;
 }
 
@@ -231,32 +240,98 @@ export function orderProperties(properties, order) {
   if (!Array.isArray(order)) {
     return properties;
   }
-  if (order.length !== properties.length) {
+
+  const arrayToHash = arr =>
+    arr.reduce((prev, curr) => {
+      prev[curr] = true;
+      return prev;
+    }, {});
+  const errorPropList = arr =>
+    arr.length > 1
+      ? `properties '${arr.join("', '")}'`
+      : `property '${arr[0]}'`;
+  const propertyHash = arrayToHash(properties);
+  const orderHash = arrayToHash(order);
+  const extraneous = order.filter(prop => prop !== "*" && !propertyHash[prop]);
+  if (extraneous.length) {
     throw new Error(
-      "uiSchema order list length should match object properties length");
+      `uiSchema order list contains extraneous ${errorPropList(extraneous)}`
+    );
   }
-  const fingerprint = (arr) => [].slice.call(arr).sort().toString();
-  if (fingerprint(order) !== fingerprint(properties)) {
-    throw new Error(
-      "uiSchema order list does not match object properties list");
+  const rest = properties.filter(prop => !orderHash[prop]);
+  const restIndex = order.indexOf("*");
+  if (restIndex === -1) {
+    if (rest.length) {
+      throw new Error(
+        `uiSchema order list does not contain ${errorPropList(rest)}`
+      );
+    }
+    return order;
   }
-  return order;
+  if (restIndex !== order.lastIndexOf("*")) {
+    throw new Error("uiSchema order list contains more than one wildcard item");
+  }
+
+  const complete = [...order];
+  complete.splice(restIndex, 1, ...rest);
+  return complete;
 }
 
-export function isMultiSelect(schema) {
-  return Array.isArray(schema.items.enum) && schema.uniqueItems;
-}
-
-export function isFilesArray(schema, uiSchema) {
+/**
+ * This function checks if the given schema matches a single
+ * constant value.
+ */
+export function isConstant(schema) {
   return (
-    schema.items.type === "string" && schema.items.format === "data-url"
-  ) || uiSchema["ui:widget"] === "files";
+    (Array.isArray(schema.enum) && schema.enum.length === 1) ||
+    schema.hasOwnProperty("const")
+  );
+}
+
+export function toConstant(schema) {
+  if (Array.isArray(schema.enum) && schema.enum.length === 1) {
+    return schema.enum[0];
+  } else if (schema.hasOwnProperty("const")) {
+    return schema.const;
+  } else {
+    throw new Error("schema cannot be inferred as a constant");
+  }
+}
+
+export function isSelect(_schema, definitions = {}) {
+  const schema = retrieveSchema(_schema, definitions);
+  const altSchemas = schema.oneOf || schema.anyOf;
+  if (Array.isArray(schema.enum)) {
+    return true;
+  } else if (Array.isArray(altSchemas)) {
+    return altSchemas.every(altSchemas => isConstant(altSchemas));
+  }
+  return false;
+}
+
+export function isMultiSelect(schema, definitions = {}) {
+  if (!schema.uniqueItems || !schema.items) {
+    return false;
+  }
+  return isSelect(schema.items, definitions);
+}
+
+export function isFilesArray(schema, uiSchema, definitions = {}) {
+  if (uiSchema["ui:widget"] === "files") {
+    return true;
+  } else if (schema.items) {
+    const itemsSchema = retrieveSchema(schema.items, definitions);
+    return itemsSchema.type === "string" && itemsSchema.format === "data-url";
+  }
+  return false;
 }
 
 export function isFixedItems(schema) {
-  return Array.isArray(schema.items) &&
-         schema.items.length > 0 &&
-         schema.items.every(item => isObject(item));
+  return (
+    Array.isArray(schema.items) &&
+    schema.items.length > 0 &&
+    schema.items.every(item => isObject(item))
+  );
 }
 
 export function allowAdditionalItems(schema) {
@@ -267,23 +342,44 @@ export function allowAdditionalItems(schema) {
 }
 
 export function optionsList(schema) {
-  return schema.enum.map((value, i) => {
-    const label = schema.enumNames && schema.enumNames[i] || String(value);
-    return {label, value};
-  });
+  if (schema.enum) {
+    return schema.enum.map((value, i) => {
+      const label = (schema.enumNames && schema.enumNames[i]) || String(value);
+      return { label, value };
+    });
+  } else {
+    const altSchemas = schema.oneOf || schema.anyOf;
+    return altSchemas.map((schema, i) => {
+      const value = toConstant(schema);
+      const label = schema.title || String(value);
+      return { label, value };
+    });
+  }
 }
 
-function findSchemaDefinition($ref, definitions={}) {
+function findSchemaDefinition($ref, definitions = {}) {
   // Extract and use the referenced definition if we have it.
-  const match = /#\/definitions\/(.*)$/.exec($ref);
-  if (match && match[1] && definitions.hasOwnProperty(match[1])) {
-    return definitions[match[1]];
+  const match = /^#\/definitions\/(.*)$/.exec($ref);
+  if (match && match[1]) {
+    const parts = match[1].split("/");
+    let current = definitions;
+    for (let part of parts) {
+      part = part.replace(/~1/g, "/").replace(/~0/g, "~");
+      if (current.hasOwnProperty(part)) {
+        current = current[part];
+      } else {
+        // No matching definition found, that's an error (bogus schema?)
+        throw new Error(`Could not find a definition for ${$ref}.`);
+      }
+    }
+    return current;
   }
+
   // No matching definition found, that's an error (bogus schema?)
   throw new Error(`Could not find a definition for ${$ref}.`);
 }
 
-export function retrieveSchema(schema, definitions={}) {
+export function retrieveSchema(schema, definitions = {}) {
   // No $ref attribute found, returning the original schema.
   if (!schema.hasOwnProperty("$ref")) {
     return schema;
@@ -291,12 +387,12 @@ export function retrieveSchema(schema, definitions={}) {
   // Retrieve the referenced schema definition.
   const $refSchema = findSchemaDefinition(schema.$ref, definitions);
   // Drop the $ref property of the source schema.
-  const {$ref, ...localSchema} = schema; // eslint-disable-line no-unused-vars
+  const { $ref, ...localSchema } = schema;
   // Update referenced schema definition with local schema properties.
-  return {...$refSchema, ...localSchema};
+  return { ...$refSchema, ...localSchema };
 }
 
-function isArguments (object) {
+function isArguments(object) {
   return Object.prototype.toString.call(object) === "[object Arguments]";
 }
 
@@ -317,11 +413,13 @@ export function deepEquals(a, b, ca = [], cb = []) {
   } else if (a instanceof Date && b instanceof Date) {
     return a.getTime() === b.getTime();
   } else if (a instanceof RegExp && b instanceof RegExp) {
-    return a.source === b.source &&
-    a.global === b.global &&
-    a.multiline === b.multiline &&
-    a.lastIndex === b.lastIndex &&
-    a.ignoreCase === b.ignoreCase;
+    return (
+      a.source === b.source &&
+      a.global === b.global &&
+      a.multiline === b.multiline &&
+      a.lastIndex === b.lastIndex &&
+      a.ignoreCase === b.ignoreCase
+    );
   } else if (isArguments(a) || isArguments(b)) {
     if (!(isArguments(a) && isArguments(b))) {
       return false;
@@ -376,19 +474,19 @@ export function deepEquals(a, b, ca = [], cb = []) {
 }
 
 export function shouldRender(comp, nextProps, nextState) {
-  const {props, state} = comp;
+  const { props, state } = comp;
   return !deepEquals(props, nextProps) || !deepEquals(state, nextState);
 }
 
 export function toIdSchema(schema, id, definitions) {
   const idSchema = {
-    $id: id || "root"
+    $id: id || "root",
   };
   if ("$ref" in schema) {
     const _schema = retrieveSchema(schema, definitions);
     return toIdSchema(_schema, id, definitions);
   }
-  if ("items" in schema) {
+  if ("items" in schema && !schema.items.$ref) {
     return toIdSchema(schema.items, id, definitions);
   }
   if (schema.type !== "object") {
@@ -427,14 +525,10 @@ export function parseDateString(dateString, includeTime = true) {
   };
 }
 
-export function toDateString({
-  year,
-  month,
-  day,
-  hour=0,
-  minute=0,
-  second=0
-}, time = true) {
+export function toDateString(
+  { year, month, day, hour = 0, minute = 0, second = 0 },
+  time = true
+) {
   const utcTime = Date.UTC(year, month - 1, day, hour, minute, second);
   const datetime = new Date(utcTime).toJSON();
   return time ? datetime : datetime.slice(0, 10);
@@ -449,7 +543,7 @@ export function pad(num, size) {
 }
 
 export function setState(instance, state, callback) {
-  const {safeRenderCompletion} = instance.props;
+  const { safeRenderCompletion } = instance.props;
   if (safeRenderCompletion) {
     instance.setState(state, callback);
   } else {
@@ -482,11 +576,25 @@ export function dataURItoBlob(dataURI) {
   // Built the Uint8Array Blob parameter from the base64 string.
   const binary = atob(splitted[1]);
   const array = [];
-  for(let i = 0; i < binary.length; i++) {
+  for (let i = 0; i < binary.length; i++) {
     array.push(binary.charCodeAt(i));
   }
   // Create the blob object
-  const blob = new window.Blob([new Uint8Array(array)], {type});
+  const blob = new window.Blob([new Uint8Array(array)], { type });
 
-  return {blob, name};
+  return { blob, name };
+}
+
+export function rangeSpec(schema) {
+  const spec = {};
+  if (schema.multipleOf) {
+    spec.step = schema.multipleOf;
+  }
+  if (schema.minimum || schema.minimum === 0) {
+    spec.min = schema.minimum;
+  }
+  if (schema.maximum || schema.maximum === 0) {
+    spec.max = schema.maximum;
+  }
+  return spec;
 }
