@@ -1,25 +1,7 @@
+import toPath from "lodash.topath";
 import { validate as jsonValidate } from "jsonschema";
 
 import { isObject, mergeObjects } from "./utils";
-
-
-const RE_ERROR_ARRAY_PATH = /\[\d+]/g;
-
-function errorPropertyToPath(property) {
-  // Parse array indices, eg. "instance.level1.level2[2].level3"
-  // => ["instance", "level1", "level2", 2, "level3"]
-  return property.split(".").reduce((path, node) => {
-    const match = node.match(RE_ERROR_ARRAY_PATH);
-    if (match) {
-      const nodeName = node.slice(0, node.indexOf("["));
-      const indices = match.map(str => parseInt(str.slice(1, -1), 10));
-      path = path.concat(nodeName, indices);
-    } else {
-      path.push(node);
-    }
-    return path;
-  }, []);
-}
 
 function toErrorSchema(errors) {
   // Transforms a jsonschema validation errors list:
@@ -41,8 +23,8 @@ function toErrorSchema(errors) {
     return {};
   }
   return errors.reduce((errorSchema, error) => {
-    const {property, message} = error;
-    const path = errorPropertyToPath(property);
+    const { property, message } = error;
+    const path = toPath(property);
     let parent = errorSchema;
     for (const segment of path.slice(1)) {
       if (!(segment in parent)) {
@@ -66,11 +48,13 @@ export function toErrorList(errorSchema, fieldName = "root") {
   // XXX: We should transform fieldName as a full field path string.
   let errorList = [];
   if ("__errors" in errorSchema) {
-    errorList = errorList.concat(errorSchema.__errors.map(stack => {
-      return {
-        stack: `${fieldName}: ${stack}`
-      };
-    }));
+    errorList = errorList.concat(
+      errorSchema.__errors.map(stack => {
+        return {
+          stack: `${fieldName}: ${stack}`,
+        };
+      })
+    );
   }
   return Object.keys(errorSchema).reduce((acc, key) => {
     if (key !== "__errors") {
@@ -92,7 +76,12 @@ function createErrorHandler(formData) {
   };
   if (isObject(formData)) {
     return Object.keys(formData).reduce((acc, key) => {
-      return {...acc, [key]: createErrorHandler(formData[key])};
+      return { ...acc, [key]: createErrorHandler(formData[key]) };
+    }, handler);
+  }
+  if (Array.isArray(formData)) {
+    return formData.reduce((acc, value, key) => {
+      return { ...acc, [key]: createErrorHandler(value) };
     }, handler);
   }
   return handler;
@@ -103,9 +92,9 @@ function unwrapErrorHandler(errorHandler) {
     if (key === "addError") {
       return acc;
     } else if (key === "__errors") {
-      return {...acc, [key]: errorHandler[key]};
+      return { ...acc, [key]: errorHandler[key] };
     }
-    return {...acc, [key]: unwrapErrorHandler(errorHandler[key])};
+    return { ...acc, [key]: unwrapErrorHandler(errorHandler[key]) };
   }, {});
 }
 
@@ -114,12 +103,20 @@ function unwrapErrorHandler(errorHandler) {
  * function, which receives the form data and an `errorHandler` object that
  * will be used to add custom validation errors for each field.
  */
-export default function validateFormData(formData, schema, customValidate) {
-  const {errors} = jsonValidate(formData, schema);
+export default function validateFormData(
+  formData,
+  schema,
+  customValidate,
+  transformErrors
+) {
+  let { errors } = jsonValidate(formData, schema);
+  if (typeof transformErrors === "function") {
+    errors = transformErrors(errors);
+  }
   const errorSchema = toErrorSchema(errors);
 
   if (typeof customValidate !== "function") {
-    return {errors, errorSchema};
+    return { errors, errorSchema };
   }
 
   const errorHandler = customValidate(formData, createErrorHandler(formData));
@@ -130,5 +127,5 @@ export default function validateFormData(formData, schema, customValidate) {
   // properties.
   const newErrors = toErrorList(newErrorSchema);
 
-  return {errors: newErrors, errorSchema: newErrorSchema};
+  return { errors: newErrors, errorSchema: newErrorSchema };
 }
