@@ -7,7 +7,6 @@ import { findDOMNode } from "react-dom";
 import Form from "../src";
 import { createFormComponent, createSandbox } from "./test_utils";
 
-
 describe("Form", () => {
   let sandbox;
 
@@ -21,20 +20,19 @@ describe("Form", () => {
 
   describe("Empty schema", () => {
     it("should render a form tag", () => {
-      const {node} = createFormComponent({schema: {}});
+      const { node } = createFormComponent({ schema: {} });
 
       expect(node.tagName).eql("FORM");
     });
 
     it("should render a submit button", () => {
-      const {node} = createFormComponent({schema: {}});
+      const { node } = createFormComponent({ schema: {} });
 
-      expect(node.querySelectorAll("button[type=submit]"))
-        .to.have.length.of(1);
+      expect(node.querySelectorAll("button[type=submit]")).to.have.length.of(1);
     });
 
     it("should render children buttons", () => {
-      const props = {schema: {}};
+      const props = { schema: {} };
       const comp = renderIntoDocument(
         <Form {...props}>
           <button type="submit">Submit</button>
@@ -42,8 +40,7 @@ describe("Form", () => {
         </Form>
       );
       const node = findDOMNode(comp);
-      expect(node.querySelectorAll("button[type=submit]"))
-        .to.have.length.of(2);
+      expect(node.querySelectorAll("button[type=submit]")).to.have.length.of(2);
     });
   });
 
@@ -57,27 +54,57 @@ describe("Form", () => {
           type: "string",
           description: "this is description",
           minLength: 32,
-        }
-      }
+        },
+      },
     };
 
     const uiSchema = {
       foo: {
-        "ui:help": "this is help"
-      }
+        "ui:help": "this is help",
+      },
     };
 
-    const formData = {foo: "invalid"};
+    const formData = { foo: "invalid" };
 
     function FieldTemplate(props) {
-      const {id, classNames, label, help, required, description, errors, children} = props;
+      const {
+        id,
+        classNames,
+        label,
+        help,
+        rawHelp,
+        required,
+        description,
+        rawDescription,
+        errors,
+        rawErrors,
+        children,
+      } = props;
       return (
         <div className={"my-template " + classNames}>
-          <label htmlFor={id}>{label}{required ? "*" : null}</label>
+          <label htmlFor={id}>
+            {label}
+            {required ? "*" : null}
+          </label>
           {description}
           {children}
           {errors}
           {help}
+          <span className="raw-help">
+            {`${rawHelp} rendered from the raw format`}
+          </span>
+          <span className="raw-description">
+            {`${rawDescription} rendered from the raw format`}
+          </span>
+          {rawErrors
+            ? <ul>
+                {rawErrors.map((error, i) =>
+                  <li key={i} className="raw-error">
+                    {error}
+                  </li>
+                )}
+              </ul>
+            : null}
         </div>
       );
     }
@@ -99,25 +126,42 @@ describe("Form", () => {
     });
 
     it("should use the provided template for labels", () => {
-      expect(node.querySelector(".my-template > label").textContent)
-        .eql("root object");
-      expect(node.querySelector(".my-template .field-string > label").textContent)
-        .eql("foo*");
+      expect(node.querySelector(".my-template > label").textContent).eql(
+        "root object"
+      );
+      expect(
+        node.querySelector(".my-template .field-string > label").textContent
+      ).eql("foo*");
     });
 
-    it("should use the provided template for descriptions", () => {
-      expect(node.querySelector("#root_foo__description").textContent)
-        .eql("this is description");
+    it("should pass description as the provided React element", () => {
+      expect(node.querySelector("#root_foo__description").textContent).eql(
+        "this is description"
+      );
     });
 
-    it("should use the provided template for errors", () => {
-      expect(node.querySelectorAll(".error-detail li"))
-        .to.have.length.of(1);
+    it("should pass rawDescription as a string", () => {
+      expect(node.querySelector(".raw-description").textContent).eql(
+        "this is description rendered from the raw format"
+      );
     });
 
-    it("should use the provided template for help", () => {
-      expect(node.querySelector(".help-block").textContent)
-        .eql("this is help");
+    it("should pass errors as the provided React component", () => {
+      expect(node.querySelectorAll(".error-detail li")).to.have.length.of(1);
+    });
+
+    it("should pass rawErrors as an array of strings", () => {
+      expect(node.querySelectorAll(".raw-error")).to.have.length.of(1);
+    });
+
+    it("should pass help as a the provided React element", () => {
+      expect(node.querySelector(".help-block").textContent).eql("this is help");
+    });
+
+    it("should pass rawHelp as a string", () => {
+      expect(node.querySelector(".raw-help").textContent).eql(
+        "this is help rendered from the raw format"
+      );
     });
   });
 
@@ -125,7 +169,7 @@ describe("Form", () => {
     it("should submit the form when clicked", () => {
       const onSubmit = sandbox.spy();
       const comp = renderIntoDocument(
-        <Form onSubmit={ onSubmit } schema={ {} }>
+        <Form onSubmit={onSubmit} schema={{}}>
           <button type="submit">Submit</button>
           <button type="submit">Another submit</button>
         </Form>
@@ -142,139 +186,186 @@ describe("Form", () => {
     it("should use a single schema definition reference", () => {
       const schema = {
         definitions: {
-          testdef: {type: "string"}
+          testdef: { type: "string" },
         },
-        $ref: "#/definitions/testdef"
+        $ref: "#/definitions/testdef",
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
-      expect(node.querySelectorAll("input[type=text]"))
-        .to.have.length.of(1);
+      expect(node.querySelectorAll("input[type=text]")).to.have.length.of(1);
     });
 
     it("should handle multiple schema definition references", () => {
       const schema = {
         definitions: {
-          testdef: {type: "string"}
+          testdef: { type: "string" },
         },
         type: "object",
         properties: {
-          foo: {$ref: "#/definitions/testdef"},
-          bar: {$ref: "#/definitions/testdef"},
-        }
+          foo: { $ref: "#/definitions/testdef" },
+          bar: { $ref: "#/definitions/testdef" },
+        },
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
-      expect(node.querySelectorAll("input[type=text]"))
-        .to.have.length.of(2);
+      expect(node.querySelectorAll("input[type=text]")).to.have.length.of(2);
     });
 
     it("should handle deeply referenced schema definitions", () => {
       const schema = {
         definitions: {
-          testdef: {type: "string"}
+          testdef: { type: "string" },
         },
         type: "object",
         properties: {
           foo: {
             type: "object",
             properties: {
-              bar: {$ref: "#/definitions/testdef"},
-            }
-          }
-        }
+              bar: { $ref: "#/definitions/testdef" },
+            },
+          },
+        },
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
-      expect(node.querySelectorAll("input[type=text]"))
-        .to.have.length.of(1);
+      expect(node.querySelectorAll("input[type=text]")).to.have.length.of(1);
+    });
+
+    it("should handle references to deep schema definitions", () => {
+      const schema = {
+        definitions: {
+          testdef: {
+            type: "object",
+            properties: {
+              bar: { type: "string" },
+            },
+          },
+        },
+        type: "object",
+        properties: {
+          foo: { $ref: "#/definitions/testdef/properties/bar" },
+        },
+      };
+
+      const { node } = createFormComponent({ schema });
+
+      expect(node.querySelectorAll("input[type=text]")).to.have.length.of(1);
     });
 
     it("should handle referenced definitions for array items", () => {
       const schema = {
         definitions: {
-          testdef: {type: "string"}
+          testdef: { type: "string" },
         },
         type: "object",
         properties: {
           foo: {
             type: "array",
-            items: {$ref: "#/definitions/testdef"}
-          }
-        }
+            items: { $ref: "#/definitions/testdef" },
+          },
+        },
       };
 
-      const {node} = createFormComponent({schema, formData: {
-        foo: ["blah"]
-      }});
+      const { node } = createFormComponent({
+        schema,
+        formData: {
+          foo: ["blah"],
+        },
+      });
 
-      expect(node.querySelectorAll("input[type=text]"))
-        .to.have.length.of(1);
+      expect(node.querySelectorAll("input[type=text]")).to.have.length.of(1);
     });
 
     it("should raise for non-existent definitions referenced", () => {
       const schema = {
         type: "object",
         properties: {
-          foo: {$ref: "#/definitions/nonexistent"},
-        }
+          foo: { $ref: "#/definitions/nonexistent" },
+        },
       };
 
-      expect(() => createFormComponent({schema}))
-        .to.Throw(Error, /#\/definitions\/nonexistent/);
+      expect(() => createFormComponent({ schema })).to.Throw(
+        Error,
+        /#\/definitions\/nonexistent/
+      );
     });
 
     it("should propagate referenced definition defaults", () => {
       const schema = {
         definitions: {
-          testdef: {type: "string", default: "hello"}
+          testdef: { type: "string", default: "hello" },
         },
-        $ref: "#/definitions/testdef"
+        $ref: "#/definitions/testdef",
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
-      expect(node.querySelector("input[type=text]").value)
-        .eql("hello");
+      expect(node.querySelector("input[type=text]").value).eql("hello");
     });
 
     it("should propagate nested referenced definition defaults", () => {
       const schema = {
         definitions: {
-          testdef: {type: "string", default: "hello"}
+          testdef: { type: "string", default: "hello" },
         },
         type: "object",
         properties: {
-          foo: {$ref: "#/definitions/testdef"}
-        }
+          foo: { $ref: "#/definitions/testdef" },
+        },
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
-      expect(node.querySelector("input[type=text]").value)
-        .eql("hello");
+      expect(node.querySelector("input[type=text]").value).eql("hello");
     });
 
     it("should propagate referenced definition defaults for array items", () => {
       const schema = {
         definitions: {
-          testdef: {type: "string", default: "hello"}
+          testdef: { type: "string", default: "hello" },
         },
         type: "array",
         items: {
-          $ref: "#/definitions/testdef"
-        }
+          $ref: "#/definitions/testdef",
+        },
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
       Simulate.click(node.querySelector(".array-item-add button"));
 
-      expect(node.querySelector("input[type=text]").value)
-        .eql("hello");
+      expect(node.querySelector("input[type=text]").value).eql("hello");
+    });
+
+    it("should recursively handle referenced definitions", () => {
+      const schema = {
+        $ref: "#/definitions/node",
+        definitions: {
+          node: {
+            type: "object",
+            properties: {
+              name: { type: "string" },
+              children: {
+                type: "array",
+                items: {
+                  $ref: "#/definitions/node",
+                },
+              },
+            },
+          },
+        },
+      };
+
+      const { node } = createFormComponent({ schema });
+
+      expect(node.querySelector("#root_children_0_name")).to.not.exist;
+
+      Simulate.click(node.querySelector(".array-item-add button"));
+
+      expect(node.querySelector("#root_children_0_name")).to.exist;
     });
 
     it("should priorize definition over schema type property", () => {
@@ -282,26 +373,25 @@ describe("Form", () => {
       const schema = {
         type: "object",
         properties: {
-          name: {type: "string"},
+          name: { type: "string" },
           childObj: {
             type: "object",
-            $ref: "#/definitions/childObj"
-          }
+            $ref: "#/definitions/childObj",
+          },
         },
         definitions: {
           childObj: {
             type: "object",
             properties: {
-              otherName: {type: "string"}
-            }
-          }
-        }
+              otherName: { type: "string" },
+            },
+          },
+        },
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
-      expect(node.querySelectorAll("input[type=text]"))
-        .to.have.length.of(2);
+      expect(node.querySelectorAll("input[type=text]")).to.have.length.of(2);
     });
 
     it("should priorize local properties over definition ones", () => {
@@ -311,41 +401,56 @@ describe("Form", () => {
         properties: {
           foo: {
             title: "custom title",
-            $ref: "#/definitions/objectDef"
-          }
+            $ref: "#/definitions/objectDef",
+          },
         },
         definitions: {
           objectDef: {
             type: "object",
             title: "definition title",
             properties: {
-              field: {type: "string"}
-            }
-          }
-        }
+              field: { type: "string" },
+            },
+          },
+        },
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
-      expect(node.querySelector("legend").textContent)
-        .eql("custom title");
+      expect(node.querySelector("legend").textContent).eql("custom title");
     });
 
     it("should propagate and handle a resolved schema definition", () => {
       const schema = {
         definitions: {
-          enumDef: {type: "string", enum: ["a", "b"]}
+          enumDef: { type: "string", enum: ["a", "b"] },
         },
         type: "object",
         properties: {
-          name: {$ref: "#/definitions/enumDef"}
+          name: { $ref: "#/definitions/enumDef" },
         },
       };
 
-      const {node} = createFormComponent({schema});
+      const { node } = createFormComponent({ schema });
 
-      expect(node.querySelectorAll("option"))
-        .to.have.length.of(2);
+      expect(node.querySelectorAll("option")).to.have.length.of(3);
+    });
+  });
+
+  describe("Default value handling on clear", () => {
+    const schema = {
+      type: "string",
+      default: "foo",
+    };
+
+    it("should not set default when a text field is cleared", () => {
+      const { node } = createFormComponent({ schema, formData: "bar" });
+
+      Simulate.change(node.querySelector("input"), {
+        target: { value: "" },
+      });
+
+      expect(node.querySelector("input").value).eql("");
     });
   });
 
@@ -366,18 +471,18 @@ describe("Form", () => {
                 properties: {
                   bool: {
                     type: "boolean",
-                    default: true
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
+                    default: true,
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
     };
 
-    it("should propagate deeply nested defaults to form state", (done) => {
-      const {comp, node} = createFormComponent({schema});
+    it("should propagate deeply nested defaults to form state", done => {
+      const { comp, node } = createFormComponent({ schema });
 
       Simulate.click(node.querySelector(".array-item-add button"));
       Simulate.submit(node);
@@ -388,10 +493,10 @@ describe("Form", () => {
           object: {
             array: [
               {
-                bool: true
-              }
-            ]
-          }
+                bool: true,
+              },
+            ],
+          },
         });
         done();
       }, 250);
@@ -403,14 +508,18 @@ describe("Form", () => {
       const schema = {
         type: "object",
         properties: {
-          foo: {type: "string"},
-        }
+          foo: { type: "string" },
+        },
       };
       const formData = {
-        foo: "bar"
+        foo: "bar",
       };
       const onSubmit = sandbox.spy();
-      const {comp, node} = createFormComponent({schema, formData, onSubmit});
+      const { comp, node } = createFormComponent({
+        schema,
+        formData,
+        onSubmit,
+      });
 
       Simulate.submit(node);
 
@@ -425,14 +534,19 @@ describe("Form", () => {
             type: "string",
             minLength: 1,
           },
-        }
+        },
       };
       const formData = {
-        foo: ""
+        foo: "",
       };
       const onSubmit = sandbox.spy();
       const onError = sandbox.spy();
-      const {node} = createFormComponent({schema, formData, onSubmit, onError});
+      const { node } = createFormComponent({
+        schema,
+        formData,
+        onSubmit,
+        onError,
+      });
 
       Simulate.submit(node);
 
@@ -448,23 +562,47 @@ describe("Form", () => {
           foo: {
             type: "string",
           },
-        }
+        },
       };
       const formData = {
-        foo: ""
+        foo: "",
       };
       const onChange = sandbox.spy();
-      const {node} = createFormComponent({schema, formData, onChange});
+      const { node } = createFormComponent({ schema, formData, onChange });
 
       Simulate.change(node.querySelector("[type=text]"), {
-        target: {value: "new"}
+        target: { value: "new" },
       });
 
       sinon.assert.calledWithMatch(onChange, {
         formData: {
-          foo: "new"
-        }
+          foo: "new",
+        },
       });
+    });
+  });
+  describe("Blur handler", () => {
+    it("should call provided blur handler on form input blur event", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "string",
+          },
+        },
+      };
+      const formData = {
+        foo: "",
+      };
+      const onBlur = sandbox.spy();
+      const { node } = createFormComponent({ schema, formData, onBlur });
+
+      const input = node.querySelector("[type=text]");
+      Simulate.blur(input, {
+        target: { value: "new" },
+      });
+
+      sinon.assert.calledWithMatch(onBlur, input.id, "new");
     });
   });
 
@@ -477,13 +615,13 @@ describe("Form", () => {
             type: "string",
             minLength: 1,
           },
-        }
+        },
       };
       const formData = {
-        foo: ""
+        foo: "",
       };
       const onError = sandbox.spy();
-      const {node} = createFormComponent({schema, formData, onError});
+      const { node } = createFormComponent({ schema, formData, onError });
 
       Simulate.submit(node);
 
@@ -494,45 +632,49 @@ describe("Form", () => {
   describe("External formData updates", () => {
     describe("root level", () => {
       const formProps = {
-        schema: {type: "string"},
+        schema: { type: "string" },
         liveValidate: true,
       };
 
       it("should update form state from new formData prop value", () => {
-        const {comp} = createFormComponent(formProps);
+        const { comp } = createFormComponent(formProps);
 
-        comp.componentWillReceiveProps({formData: "yo"});
+        comp.componentWillReceiveProps({ formData: "yo" });
 
         expect(comp.state.formData).eql("yo");
       });
 
       it("should validate formData when the schema is updated", () => {
-        const {comp} = createFormComponent(formProps);
+        const { comp } = createFormComponent(formProps);
 
-        comp.componentWillReceiveProps({formData: "yo", schema: {type: "number"}});
+        comp.componentWillReceiveProps({
+          formData: "yo",
+          schema: { type: "number" },
+        });
 
         expect(comp.state.errors).to.have.length.of(1);
-        expect(comp.state.errors[0].stack)
-          .to.eql("instance is not of a type(s) number");
+        expect(comp.state.errors[0].stack).to.eql(
+          "instance is not of a type(s) number"
+        );
       });
     });
 
     describe("object level", () => {
       it("should update form state from new formData prop value", () => {
-        const {comp} = createFormComponent({
+        const { comp } = createFormComponent({
           schema: {
             type: "object",
             properties: {
               foo: {
-                type: "string"
-              }
-            }
-          }
+                type: "string",
+              },
+            },
+          },
         });
 
-        comp.componentWillReceiveProps({formData: {foo: "yo"}});
+        comp.componentWillReceiveProps({ formData: { foo: "yo" } });
 
-        expect(comp.state.formData).eql({foo: "yo"});
+        expect(comp.state.formData).eql({ foo: "yo" });
       });
     });
 
@@ -541,12 +683,12 @@ describe("Form", () => {
         const schema = {
           type: "array",
           items: {
-            type: "string"
-          }
+            type: "string",
+          },
         };
-        const {comp} = createFormComponent({schema});
+        const { comp } = createFormComponent({ schema });
 
-        comp.componentWillReceiveProps({formData: ["yo"]});
+        comp.componentWillReceiveProps({ formData: ["yo"] });
 
         expect(comp.state.formData).eql(["yo"]);
       });
@@ -557,48 +699,50 @@ describe("Form", () => {
     describe("on form state updated", () => {
       const schema = {
         type: "string",
-        minLength: 8
+        minLength: 8,
       };
 
       describe("Lazy validation", () => {
         it("should not update the errorSchema when the formData changes", () => {
-          const {comp, node} = createFormComponent({schema});
+          const { comp, node } = createFormComponent({ schema });
 
           Simulate.change(node.querySelector("input[type=text]"), {
-            target: {value: "short"}
+            target: { value: "short" },
           });
 
           expect(comp.state.errorSchema).eql({});
         });
 
         it("should not denote an error in the field", () => {
-          const {node} = createFormComponent({schema});
+          const { node } = createFormComponent({ schema });
 
           Simulate.change(node.querySelector("input[type=text]"), {
-            target: {value: "short"}
+            target: { value: "short" },
           });
 
-          expect(node.querySelectorAll(".field-error"))
-            .to.have.length.of(0);
+          expect(node.querySelectorAll(".field-error")).to.have.length.of(0);
         });
 
         it("should clean contextualized errors up when they're fixed", () => {
           const altSchema = {
             type: "object",
             properties: {
-              field1: {type: "string", minLength: 8},
-              field2: {type: "string", minLength: 8},
-            }
+              field1: { type: "string", minLength: 8 },
+              field2: { type: "string", minLength: 8 },
+            },
           };
-          const {node} = createFormComponent({schema: altSchema, formData: {
-            field1: "short",
-            field2: "short",
-          }});
+          const { node } = createFormComponent({
+            schema: altSchema,
+            formData: {
+              field1: "short",
+              field2: "short",
+            },
+          });
 
           function submit(node) {
             try {
               Simulate.submit(node);
-            } catch(err) {
+            } catch (err) {
               // Validation is expected to fail and call console.error, which is
               // stubbed to actually throw in createSandbox().
             }
@@ -608,64 +752,65 @@ describe("Form", () => {
 
           // Fix the first field
           Simulate.change(node.querySelectorAll("input[type=text]")[0], {
-            target: {value: "fixed error"}
+            target: { value: "fixed error" },
           });
           submit(node);
 
-          expect(node.querySelectorAll(".field-error"))
-            .to.have.length.of(1);
+          expect(node.querySelectorAll(".field-error")).to.have.length.of(1);
 
           // Fix the second field
           Simulate.change(node.querySelectorAll("input[type=text]")[1], {
-            target: {value: "fixed error too"}
+            target: { value: "fixed error too" },
           });
           submit(node);
 
           // No error remaining, shouldn't throw.
           Simulate.submit(node);
 
-          expect(node.querySelectorAll(".field-error"))
-            .to.have.length.of(0);
+          expect(node.querySelectorAll(".field-error")).to.have.length.of(0);
         });
       });
 
       describe("Live validation", () => {
         it("should update the errorSchema when the formData changes", () => {
-          const {comp, node} = createFormComponent({schema, liveValidate: true});
+          const { comp, node } = createFormComponent({
+            schema,
+            liveValidate: true,
+          });
 
           Simulate.change(node.querySelector("input[type=text]"), {
-            target: {value: "short"}
+            target: { value: "short" },
           });
 
           expect(comp.state.errorSchema).eql({
-            __errors: ["does not meet minimum length of 8"]
+            __errors: ["does not meet minimum length of 8"],
           });
         });
 
         it("should denote the new error in the field", () => {
-          const {node} = createFormComponent({schema, liveValidate: true});
+          const { node } = createFormComponent({ schema, liveValidate: true });
 
           Simulate.change(node.querySelector("input[type=text]"), {
-            target: {value: "short"}
+            target: { value: "short" },
           });
 
-          expect(node.querySelectorAll(".field-error"))
-            .to.have.length.of(1);
-          expect(node.querySelector(".field-string .error-detail").textContent)
-            .eql("does not meet minimum length of 8");
+          expect(node.querySelectorAll(".field-error")).to.have.length.of(1);
+          expect(
+            node.querySelector(".field-string .error-detail").textContent
+          ).eql("does not meet minimum length of 8");
         });
       });
 
       describe("Disable validation onChange event", () => {
         it("should not update errorSchema when the formData changes", () => {
-          const {comp, node} = createFormComponent({
+          const { comp, node } = createFormComponent({
             schema,
             noValidate: true,
-            liveValidate: true
+            liveValidate: true,
           });
 
           Simulate.change(node.querySelector("input[type=text]"), {
-            target: {value: "short"}
+            target: { value: "short" },
           });
 
           expect(comp.state.errorSchema).eql({});
@@ -674,13 +819,13 @@ describe("Form", () => {
 
       describe("Disable validation onSubmit event", () => {
         it("should not update errorSchema when the formData changes", () => {
-          const {comp, node} = createFormComponent({
+          const { comp, node } = createFormComponent({
             schema,
-            noValidate: true
+            noValidate: true,
           });
 
           Simulate.change(node.querySelector("input[type=text]"), {
-            target: {value: "short"}
+            target: { value: "short" },
           });
           Simulate.submit(node);
 
@@ -692,35 +837,43 @@ describe("Form", () => {
     describe("on form submitted", () => {
       const schema = {
         type: "string",
-        minLength: 8
+        minLength: 8,
       };
 
       it("should update the errorSchema on form submission", () => {
-        const {comp, node} = createFormComponent({schema, onError: () => {}});
+        const { comp, node } = createFormComponent({
+          schema,
+          onError: () => {},
+        });
 
         Simulate.change(node.querySelector("input[type=text]"), {
-          target: {value: "short"}
+          target: { value: "short" },
         });
         Simulate.submit(node);
 
         expect(comp.state.errorSchema).eql({
-          __errors: ["does not meet minimum length of 8"]
+          __errors: ["does not meet minimum length of 8"],
         });
       });
 
       it("should call the onError handler", () => {
         const onError = sandbox.spy();
-        const {node} = createFormComponent({schema, onError});
+        const { node } = createFormComponent({ schema, onError });
 
         Simulate.change(node.querySelector("input[type=text]"), {
-          target: {value: "short"}
+          target: { value: "short" },
         });
         Simulate.submit(node);
 
-        sinon.assert.calledWithMatch(onError, sinon.match(value => {
-          return value.length === 1 &&
-                 value[0].message === "does not meet minimum length of 8";
-        }));
+        sinon.assert.calledWithMatch(
+          onError,
+          sinon.match(value => {
+            return (
+              value.length === 1 &&
+              value[0].message === "does not meet minimum length of 8"
+            );
+          })
+        );
       });
     });
 
@@ -729,26 +882,26 @@ describe("Form", () => {
         liveValidate: true,
         schema: {
           type: "string",
-          minLength: 8
+          minLength: 8,
         },
-        formData: "short"
+        formData: "short",
       };
 
       it("should reflect the contextualized error in state", () => {
-        const {comp} = createFormComponent(formProps);
+        const { comp } = createFormComponent(formProps);
 
         expect(comp.state.errorSchema).eql({
-          __errors: ["does not meet minimum length of 8"]
+          __errors: ["does not meet minimum length of 8"],
         });
       });
 
       it("should denote the error in the field", () => {
-        const {node} = createFormComponent(formProps);
+        const { node } = createFormComponent(formProps);
 
-        expect(node.querySelectorAll(".field-error"))
-          .to.have.length.of(1);
-        expect(node.querySelector(".field-string .error-detail").textContent)
-          .eql("does not meet minimum length of 8");
+        expect(node.querySelectorAll(".field-error")).to.have.length.of(1);
+        expect(
+          node.querySelector(".field-string .error-detail").textContent
+        ).eql("does not meet minimum length of 8");
       });
     });
 
@@ -758,30 +911,30 @@ describe("Form", () => {
         schema: {
           type: "string",
           minLength: 8,
-          pattern: "\d+",
+          pattern: "d+",
         },
-        formData: "short"
+        formData: "short",
       };
 
       it("should reflect the contextualized error in state", () => {
-        const {comp} = createFormComponent(formProps);
+        const { comp } = createFormComponent(formProps);
         expect(comp.state.errorSchema).eql({
           __errors: [
             "does not meet minimum length of 8",
-            "does not match pattern \"\d+\""
-          ]
+            'does not match pattern "d+"',
+          ],
         });
       });
 
       it("should denote the error in the field", () => {
-        const {node} = createFormComponent(formProps);
+        const { node } = createFormComponent(formProps);
 
         const liNodes = node.querySelectorAll(".field-string .error-detail li");
         const errors = [].map.call(liNodes, li => li.textContent);
 
         expect(errors).eql([
           "does not meet minimum length of 8",
-          "does not match pattern \"\d+\""
+          'does not match pattern "d+"',
         ]);
       });
     });
@@ -795,11 +948,11 @@ describe("Form", () => {
             properties: {
               level2: {
                 type: "string",
-                minLength: 8
-              }
-            }
-          }
-        }
+                minLength: 8,
+              },
+            },
+          },
+        },
       };
 
       const formProps = {
@@ -807,32 +960,33 @@ describe("Form", () => {
         liveValidate: true,
         formData: {
           level1: {
-            level2: "short"
-          }
-        }
+            level2: "short",
+          },
+        },
       };
 
       it("should reflect the contextualized error in state", () => {
-        const {comp} = createFormComponent(formProps);
+        const { comp } = createFormComponent(formProps);
 
         expect(comp.state.errorSchema).eql({
           level1: {
             level2: {
-              __errors: ["does not meet minimum length of 8"]
-            }
-          }
+              __errors: ["does not meet minimum length of 8"],
+            },
+          },
         });
       });
 
       it("should denote the error in the field", () => {
-        const {node} = createFormComponent(formProps);
+        const { node } = createFormComponent(formProps);
         const errorDetail = node.querySelector(
-          ".field-object .field-string .error-detail");
+          ".field-object .field-string .error-detail"
+        );
 
-        expect(node.querySelectorAll(".field-error"))
-          .to.have.length.of(1);
-        expect(errorDetail.textContent)
-          .eql("does not meet minimum length of 8");
+        expect(node.querySelectorAll(".field-error")).to.have.length.of(1);
+        expect(errorDetail.textContent).eql(
+          "does not meet minimum length of 8"
+        );
       });
     });
 
@@ -841,40 +995,39 @@ describe("Form", () => {
         type: "array",
         items: {
           type: "string",
-          minLength: 4
-        }
+          minLength: 4,
+        },
       };
 
       const formProps = {
         schema,
         liveValidate: true,
-        formData: ["good", "bad", "good"]
+        formData: ["good", "bad", "good"],
       };
 
       it("should contextualize the error for array indices", () => {
-        const {comp} = createFormComponent(formProps);
+        const { comp } = createFormComponent(formProps);
 
-        expect(comp.state.errorSchema)
-          .eql({
-            1: {__errors: ["does not meet minimum length of 4"]}
-          });
+        expect(comp.state.errorSchema).eql({
+          1: { __errors: ["does not meet minimum length of 4"] },
+        });
       });
 
       it("should denote the error in the item field in error", () => {
-        const {node} = createFormComponent(formProps);
+        const { node } = createFormComponent(formProps);
         const fieldNodes = node.querySelectorAll(".field-string");
 
-        const liNodes = fieldNodes[1]
-          .querySelectorAll(".field-string .error-detail li");
+        const liNodes = fieldNodes[1].querySelectorAll(
+          ".field-string .error-detail li"
+        );
         const errors = [].map.call(liNodes, li => li.textContent);
 
         expect(fieldNodes[1].classList.contains("field-error")).eql(true);
-        expect(errors)
-          .eql(["does not meet minimum length of 4"]);
+        expect(errors).eql(["does not meet minimum length of 4"]);
       });
 
       it("should not denote errors on non impacted fields", () => {
-        const {node} = createFormComponent(formProps);
+        const { node } = createFormComponent(formProps);
         const fieldNodes = node.querySelectorAll(".field-string");
 
         expect(fieldNodes[0].classList.contains("field-error")).eql(false);
@@ -890,37 +1043,42 @@ describe("Form", () => {
             type: "array",
             items: {
               type: "string",
-              minLength: 4
-            }
-          }
-        }
+              minLength: 4,
+            },
+          },
+        },
       };
 
-      const formProps = {schema, liveValidate: true};
+      const formProps = { schema, liveValidate: true };
 
       it("should contextualize the error for nested array indices", () => {
-        const {comp} = createFormComponent({...formProps, formData: {
-          level1: ["good", "bad", "good", "bad"]
-        }});
+        const { comp } = createFormComponent({
+          ...formProps,
+          formData: {
+            level1: ["good", "bad", "good", "bad"],
+          },
+        });
 
         expect(comp.state.errorSchema).eql({
           level1: {
-            1: {__errors: ["does not meet minimum length of 4"]},
-            3: {__errors: ["does not meet minimum length of 4"]},
-          }
+            1: { __errors: ["does not meet minimum length of 4"] },
+            3: { __errors: ["does not meet minimum length of 4"] },
+          },
         });
       });
 
       it("should denote the error in the nested item field in error", () => {
-        const {node} = createFormComponent({...formProps, formData: {
-          level1: ["good", "bad", "good"]
-        }});
+        const { node } = createFormComponent({
+          ...formProps,
+          formData: {
+            level1: ["good", "bad", "good"],
+          },
+        });
 
         const liNodes = node.querySelectorAll(".field-string .error-detail li");
         const errors = [].map.call(liNodes, li => li.textContent);
 
-        expect(errors)
-          .eql(["does not meet minimum length of 4"]);
+        expect(errors).eql(["does not meet minimum length of 4"]);
       });
     });
 
@@ -934,52 +1092,48 @@ describe("Form", () => {
               type: "array",
               items: {
                 type: "string",
-                minLength: 4
-              }
-            }
-          }
-        }
+                minLength: 4,
+              },
+            },
+          },
+        },
       };
 
       const formData = {
-        outer: [
-          ["good", "bad"],
-          ["bad", "good"]
-        ]
+        outer: [["good", "bad"], ["bad", "good"]],
       };
 
-      const formProps = {schema, formData, liveValidate: true};
+      const formProps = { schema, formData, liveValidate: true };
 
       it("should contextualize the error for nested array indices", () => {
-        const {comp} = createFormComponent(formProps);
+        const { comp } = createFormComponent(formProps);
 
         expect(comp.state.errorSchema).eql({
           outer: {
             0: {
-              1: {__errors: ["does not meet minimum length of 4"]}
+              1: { __errors: ["does not meet minimum length of 4"] },
             },
             1: {
-              0: {__errors: ["does not meet minimum length of 4"]}
-            }
-          }
+              0: { __errors: ["does not meet minimum length of 4"] },
+            },
+          },
         });
       });
 
       it("should denote the error in the nested item field in error", () => {
-        const {node} = createFormComponent(formProps);
+        const { node } = createFormComponent(formProps);
         const fields = node.querySelectorAll(".field-string");
         const errors = [].map.call(fields, field => {
           const li = field.querySelector(".error-detail li");
           return li && li.textContent;
         });
 
-        expect(errors)
-          .eql([
-            null,
-            "does not meet minimum length of 4",
-            "does not meet minimum length of 4",
-            null
-          ]);
+        expect(errors).eql([
+          null,
+          "does not meet minimum length of 4",
+          "does not meet minimum length of 4",
+          null,
+        ]);
       });
     });
 
@@ -991,43 +1145,41 @@ describe("Form", () => {
           properties: {
             foo: {
               type: "string",
-              minLength: 4
-            }
-          }
-        }
+              minLength: 4,
+            },
+          },
+        },
       };
 
       const formProps = {
         schema,
         liveValidate: true,
-        formData: [
-          {foo: "good"}, {foo: "bad"}, {foo: "good"}
-        ]
+        formData: [{ foo: "good" }, { foo: "bad" }, { foo: "good" }],
       };
 
       it("should contextualize the error for array nested items", () => {
-        const {comp} = createFormComponent(formProps);
+        const { comp } = createFormComponent(formProps);
 
         expect(comp.state.errorSchema).eql({
           1: {
             foo: {
-              __errors: ["does not meet minimum length of 4"]
-            }
-          }
+              __errors: ["does not meet minimum length of 4"],
+            },
+          },
         });
       });
 
       it("should denote the error in the array nested item", () => {
-        const {node} = createFormComponent(formProps);
+        const { node } = createFormComponent(formProps);
         const fieldNodes = node.querySelectorAll(".field-string");
 
-        const liNodes = fieldNodes[1]
-          .querySelectorAll(".field-string .error-detail li");
+        const liNodes = fieldNodes[1].querySelectorAll(
+          ".field-string .error-detail li"
+        );
         const errors = [].map.call(liNodes, li => li.textContent);
 
         expect(fieldNodes[1].classList.contains("field-error")).eql(true);
-        expect(errors)
-          .eql(["does not meet minimum length of 4"]);
+        expect(errors).eql(["does not meet minimum length of 4"]);
       });
     });
   });
@@ -1037,50 +1189,50 @@ describe("Form", () => {
     const schema = {
       type: "object",
       properties: {
-        foo: {type: "string"},
-        bar: {type: "string"},
-      }
+        foo: { type: "string" },
+        bar: { type: "string" },
+      },
     };
 
     it("should replace state when formData have keys removed", () => {
-      const formData = {foo: "foo", bar: "bar"};
-      const {comp, node} = createFormComponent({schema, formData});
+      const formData = { foo: "foo", bar: "bar" };
+      const { comp, node } = createFormComponent({ schema, formData });
       comp.componentWillReceiveProps({
         schema: {
           type: "object",
           properties: {
-            bar: {type: "string"},
-          }
+            bar: { type: "string" },
+          },
         },
-        formData: {bar: "bar"},
+        formData: { bar: "bar" },
       });
 
       Simulate.change(node.querySelector("#root_bar"), {
-        target: {value: "baz"}
+        target: { value: "baz" },
       });
 
-      expect(comp.state.formData).eql({bar: "baz"});
+      expect(comp.state.formData).eql({ bar: "baz" });
     });
 
     it("should replace state when formData keys have changed", () => {
-      const formData = {foo: "foo", bar: "bar"};
-      const {comp, node} = createFormComponent({schema, formData});
+      const formData = { foo: "foo", bar: "bar" };
+      const { comp, node } = createFormComponent({ schema, formData });
       comp.componentWillReceiveProps({
         schema: {
           type: "object",
           properties: {
-            foo: {type: "string"},
-            baz: {type: "string"},
-          }
+            foo: { type: "string" },
+            baz: { type: "string" },
+          },
         },
-        formData: {foo: "foo", baz: "bar"},
+        formData: { foo: "foo", baz: "bar" },
       });
 
       Simulate.change(node.querySelector("#root_baz"), {
-        target: {value: "baz"}
+        target: { value: "baz" },
       });
 
-      expect(comp.state.formData).eql({foo: "foo", baz: "baz"});
+      expect(comp.state.formData).eql({ foo: "foo", baz: "baz" });
     });
   });
 
@@ -1095,7 +1247,8 @@ describe("Form", () => {
       action: "/users/list",
       autocomplete: "off",
       enctype: "multipart/form-data",
-      acceptcharset: "ISO-8859-1"
+      acceptcharset: "ISO-8859-1",
+      noHtml5Validate: true,
     };
 
     let node;
@@ -1138,6 +1291,10 @@ describe("Form", () => {
 
     it("should set attr acceptcharset of form", () => {
       expect(node.getAttribute("accept-charset")).eql(formProps.acceptcharset);
+    });
+
+    it("should set attr novalidate of form", () => {
+      expect(node.getAttribute("novalidate")).not.to.be.null;
     });
   });
 });
