@@ -13,5 +13,5 @@ global.atob = require("atob");
 
 // HTML debugging helper
 global.d = function d(node) {
-  console.log(require("html").prettyPrint(node.outerHTML, {indent_size: 2}));
+  console.log(require("html").prettyPrint(node.outerHTML, { indent_size: 2 }));
 };
