@@ -1,10 +1,8 @@
-
 import { expect } from "chai";
 import React from "react";
 import { Simulate } from "react-addons-test-utils";
 
 import { createFormComponent, createSandbox } from "./test_utils";
-
 
 describe("uiSchema", () => {
   let sandbox;
@@ -21,18 +19,18 @@ describe("uiSchema", () => {
     const schema = {
       type: "object",
       properties: {
-        foo: {type: "string"},
-        bar: {type: "string"},
-      }
+        foo: { type: "string" },
+        bar: { type: "string" },
+      },
     };
 
     const uiSchema = {
-      foo: { classNames: "class-for-foo"},
-      bar: { classNames: "class-for-bar another-for-bar"},
+      foo: { classNames: "class-for-foo" },
+      bar: { classNames: "class-for-bar another-for-bar" },
     };
 
     it("should apply custom class names to target widgets", () => {
-      const {node} = createFormComponent({schema, uiSchema});
+      const { node } = createFormComponent({ schema, uiSchema });
       const [foo, bar] = node.querySelectorAll(".field-string");
 
       expect(foo.classList.contains("class-for-foo")).eql(true);
@@ -44,62 +42,208 @@ describe("uiSchema", () => {
   describe("custom widget", () => {
     describe("root widget", () => {
       const schema = {
-        type: "string"
+        type: "string",
       };
 
       const uiSchema = {
-        "ui:widget": (props) => {
+        "ui:widget": props => {
           return (
-            <input type="text"
+            <input
+              type="text"
               className="custom"
               value={props.value}
               defaultValue={props.defaultValue}
               required={props.required}
-              onChange={(event) => props.onChange(event.target.value)} />
+              onChange={event => props.onChange(event.target.value)}
+            />
           );
-        }
+        },
       };
 
       it("should render a root custom widget", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
         expect(node.querySelectorAll(".custom")).to.have.length.of(1);
       });
     });
 
+    describe("custom options", () => {
+      let widget, widgets, schema, uiSchema;
+
+      beforeEach(() => {
+        sandbox.stub(console, "warn");
+
+        widget = ({ label, options }) => <div id={label} style={options} />;
+        widget.defaultProps = {
+          options: { background: "yellow", color: "green" },
+        };
+
+        widgets = { widget };
+
+        // all fields in one schema to catch errors where options passed to one
+        // instance of a widget are persistent across all instances
+        schema = {
+          type: "object",
+          properties: {
+            funcAll: { type: "string" },
+            funcNone: { type: "string" },
+            stringAll: { type: "string" },
+            stringNone: { type: "string" },
+            stringTel: { type: "string" },
+          },
+        };
+
+        uiSchema = {
+          // pass widget as function
+          funcAll: {
+            "ui:widget": {
+              component: widget,
+              options: {
+                background: "purple",
+              },
+            },
+            "ui:options": {
+              margin: "7px",
+            },
+            "ui:padding": "42px",
+          },
+          funcNone: {
+            "ui:widget": widget,
+          },
+
+          // pass widget as string
+          stringAll: {
+            "ui:widget": {
+              component: "widget",
+              options: {
+                background: "blue",
+              },
+            },
+            "ui:options": {
+              margin: "19px",
+            },
+            "ui:padding": "41px",
+          },
+          stringNone: {
+            "ui:widget": "widget",
+          },
+          stringTel: {
+            "ui:options": {
+              inputType: "tel",
+            },
+          },
+        };
+      });
+
+      it("should log warning when deprecated ui:widget: {component, options} api is used", () => {
+        createFormComponent({
+          schema: { type: "string" },
+          uiSchema: { "ui:widget": { component: "widget" } },
+          widgets,
+        });
+        expect(console.warn.calledWithMatch(/ui:widget object is deprecated/))
+          .to.be.true;
+      });
+
+      it("should cache MergedWidget instance", () => {
+        expect(widget.MergedWidget).not.to.be.ok;
+        createFormComponent({
+          schema: { type: "string" },
+          uiSchema: { "ui:widget": "widget" },
+          widgets,
+        });
+        const cached = widget.MergedWidget;
+        expect(cached).to.be.ok;
+        createFormComponent({
+          schema: { type: "string" },
+          uiSchema: { "ui:widget": "widget" },
+          widgets,
+        });
+        expect(widget.MergedWidget).to.equal(cached);
+      });
+
+      it("should render merged ui:widget options for widget referenced as function", () => {
+        const { node } = createFormComponent({ schema, uiSchema, widgets });
+        const widget = node.querySelector("#funcAll");
+
+        expect(widget.style.background).to.equal("purple");
+        expect(widget.style.color).to.equal("green");
+        expect(widget.style.margin).to.equal("7px");
+        expect(widget.style.padding).to.equal("42px");
+      });
+
+      it("should render ui:widget default options for widget referenced as function", () => {
+        const { node } = createFormComponent({ schema, uiSchema, widgets });
+        const widget = node.querySelector("#funcNone");
+
+        expect(widget.style.background).to.equal("yellow");
+        expect(widget.style.color).to.equal("green");
+        expect(widget.style.margin).to.equal("");
+        expect(widget.style.padding).to.equal("");
+      });
+
+      it("should render merged ui:widget options for widget referenced as string", () => {
+        const { node } = createFormComponent({ schema, uiSchema, widgets });
+        const widget = node.querySelector("#stringAll");
+
+        expect(widget.style.background).to.equal("blue");
+        expect(widget.style.color).to.equal("green");
+        expect(widget.style.margin).to.equal("19px");
+        expect(widget.style.padding).to.equal("41px");
+      });
+
+      it("should render ui:widget default options for widget referenced as string", () => {
+        const { node } = createFormComponent({ schema, uiSchema, widgets });
+        const widget = node.querySelector("#stringNone");
+
+        expect(widget.style.background).to.equal("yellow");
+        expect(widget.style.color).to.equal("green");
+        expect(widget.style.margin).to.equal("");
+        expect(widget.style.padding).to.equal("");
+      });
+
+      it("should ui:option inputType for html5 input types", () => {
+        const { node } = createFormComponent({ schema, uiSchema, widgets });
+        const widget = node.querySelector("input[type='tel']");
+        expect(widget).to.not.be.null;
+      });
+    });
+
     describe("nested widget", () => {
       const schema = {
-        "type": "object",
-        "properties": {
-          "field": {
-            "type": "string"
-          }
-        }
+        type: "object",
+        properties: {
+          field: {
+            type: "string",
+          },
+        },
       };
 
       const uiSchema = {
-        "field": {
-          "ui:widget": "custom"
-        }
+        field: {
+          "ui:widget": "custom",
+        },
       };
 
-      const CustomWidget = (props) => {
+      const CustomWidget = props => {
         return (
-          <input type="text"
+          <input
+            type="text"
             className="custom"
             value={props.value}
             defaultValue={props.defaultValue}
             required={props.required}
-            onChange={(event) => props.onChange(event.target.value)} />
+            onChange={event => props.onChange(event.target.value)}
+          />
         );
       };
 
       const widgets = {
-        custom: CustomWidget
+        custom: CustomWidget,
       };
 
       it("should render a nested custom widget", () => {
-        const {node} = createFormComponent({schema, uiSchema, widgets});
+        const { node } = createFormComponent({ schema, uiSchema, widgets });
 
         expect(node.querySelectorAll(".custom")).to.have.length.of(1);
       });
@@ -107,16 +251,16 @@ describe("uiSchema", () => {
 
     describe("options", () => {
       const schema = {
-        "type": "object",
-        "properties": {
-          "field": {
-            "type": "string"
-          }
-        }
+        type: "object",
+        properties: {
+          field: {
+            type: "string",
+          },
+        },
       };
 
-      const CustomWidget = (props) => {
-        const {value, options} = props;
+      const CustomWidget = props => {
+        const { value, options } = props;
         return (
           <input type="text" className={options.className} value={value} />
         );
@@ -124,18 +268,16 @@ describe("uiSchema", () => {
 
       describe("direct reference", () => {
         const uiSchema = {
-          "field": {
-            "ui:widget": {
-              component: CustomWidget,
-              options: {
-                className: "custom"
-              }
-            }
-          }
+          field: {
+            "ui:widget": CustomWidget,
+            "ui:options": {
+              className: "custom",
+            },
+          },
         };
 
         it("should render a custom widget with options", () => {
-          const {node} = createFormComponent({schema, uiSchema});
+          const { node } = createFormComponent({ schema, uiSchema });
 
           expect(node.querySelectorAll(".custom")).to.have.length.of(1);
         });
@@ -143,45 +285,20 @@ describe("uiSchema", () => {
 
       describe("string reference", () => {
         const uiSchema = {
-          "field": {
-            "ui:widget": {
-              component: "custom",
-              options: {
-                className: "custom"
-              }
-            }
-          }
+          field: {
+            "ui:widget": "custom",
+            "ui:options": {
+              className: "custom",
+            },
+          },
         };
 
         const widgets = {
-          custom: CustomWidget
+          custom: CustomWidget,
         };
 
         it("should render a custom widget with options", () => {
-          const {node} = createFormComponent({schema, uiSchema, widgets});
-
-          expect(node.querySelectorAll(".custom")).to.have.length.of(1);
-        });
-      });
-
-      describe("referenced descriptor", () => {
-        const uiSchema = {
-          "field": {
-            "ui:widget": "custom"
-          }
-        };
-
-        const widgets = {
-          custom: {
-            component: CustomWidget,
-            options: {
-              className: "custom"
-            }
-          }
-        };
-
-        it("should render a custom widget with options", () => {
-          const {node} = createFormComponent({schema, uiSchema, widgets});
+          const { node } = createFormComponent({ schema, uiSchema, widgets });
 
           expect(node.querySelectorAll(".custom")).to.have.length.of(1);
         });
@@ -190,38 +307,40 @@ describe("uiSchema", () => {
 
     describe("enum fields native options", () => {
       const schema = {
-        "type": "object",
-        "properties": {
-          "field": {
-            "type": "string",
-            "enum": ["foo", "bar"]
-          }
-        }
+        type: "object",
+        properties: {
+          field: {
+            type: "string",
+            enum: ["foo", "bar"],
+          },
+        },
       };
 
-      const CustomWidget = (props) => {
-        const {value, options} = props;
-        const {enumOptions, className} = options;
+      const CustomWidget = props => {
+        const { value, options } = props;
+        const { enumOptions, className } = options;
         return (
-          <select className={className}>{
-            enumOptions.map(({label, value}, i) => <option key={i}>{value}</option>)
-          }</select>
+          <select className={className}>
+            {enumOptions.map(({ label, value }, i) =>
+              <option key={i}>
+                {value}
+              </option>
+            )}
+          </select>
         );
       };
 
       const uiSchema = {
-        "field": {
-          "ui:widget": {
-            component: CustomWidget,
-            options: {
-              className: "custom"
-            }
-          }
-        }
+        field: {
+          "ui:widget": CustomWidget,
+          "ui:options": {
+            className: "custom",
+          },
+        },
       };
 
       it("should merge enumOptions with custom options", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
         expect(node.querySelectorAll(".custom option")).to.have.length.of(2);
       });
     });
@@ -229,21 +348,216 @@ describe("uiSchema", () => {
 
   describe("ui:help", () => {
     it("should render the provided help text", () => {
-      const schema = {type: "string"};
-      const uiSchema = {"ui:help": "plop"};
+      const schema = { type: "string" };
+      const uiSchema = { "ui:help": "plop" };
 
-      const {node} = createFormComponent({schema, uiSchema});
+      const { node } = createFormComponent({ schema, uiSchema });
 
       expect(node.querySelector("p.help-block").textContent).eql("plop");
     });
+  });
 
-    it("should accept a react element as help", () => {
-      const schema = {type: "string"};
-      const uiSchema = {"ui:help": (<b>plop</b>)};
+  describe("ui:title", () => {
+    it("should render the provided title text", () => {
+      const schema = { type: "string" };
+      const uiSchema = { "ui:title": "plop" };
 
-      const {node} = createFormComponent({schema, uiSchema});
+      const { node } = createFormComponent({ schema, uiSchema });
 
-      expect(node.querySelector("div.help-block").textContent).eql("plop");
+      expect(node.querySelector("label.control-label").textContent).eql("plop");
+    });
+  });
+
+  describe("ui:description", () => {
+    it("should render the provided description text", () => {
+      const schema = { type: "string" };
+      const uiSchema = { "ui:description": "plop" };
+
+      const { node } = createFormComponent({ schema, uiSchema });
+
+      expect(node.querySelector("p.field-description").textContent).eql("plop");
+    });
+  });
+
+  it("should accept a react element as help", () => {
+    const schema = { type: "string" };
+    const uiSchema = { "ui:help": <b>plop</b> };
+
+    const { node } = createFormComponent({ schema, uiSchema });
+
+    expect(node.querySelector("div.help-block").textContent).eql("plop");
+  });
+
+  describe("ui:focus", () => {
+    const shouldFocus = (schema, uiSchema, selector = "input", formData) => {
+      const props = { schema, uiSchema };
+      if (typeof formData !== "undefined") {
+        props.formData = formData;
+      }
+
+      const { node } = createFormComponent(props);
+      expect(node.querySelector(selector)).eql(document.activeElement);
+    };
+
+    describe("number", () => {
+      it("should focus on integer input", () => {
+        shouldFocus({ type: "integer" }, { "ui:autofocus": true });
+      });
+
+      it("should focus on integer input, updown widget", () => {
+        shouldFocus(
+          { type: "integer" },
+          { "ui:widget": "updown", "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on integer input, range widget", () => {
+        shouldFocus(
+          { type: "integer" },
+          { "ui:widget": "range", "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on integer enum input", () => {
+        shouldFocus(
+          { type: "integer", enum: [1, 2, 3] },
+          { "ui:autofocus": true },
+          "select"
+        );
+      });
+    });
+
+    describe("string", () => {
+      it("should focus on text input", () => {
+        shouldFocus({ type: "string" }, { "ui:autofocus": true });
+      });
+
+      it("should focus on textarea", () => {
+        shouldFocus(
+          { type: "string" },
+          { "ui:widget": "textarea", "ui:autofocus": true },
+          "textarea"
+        );
+      });
+
+      it("should focus on password input", () => {
+        shouldFocus(
+          { type: "string" },
+          { "ui:widget": "password", "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on color input", () => {
+        shouldFocus(
+          { type: "string" },
+          { "ui:widget": "color", "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on email input", () => {
+        shouldFocus(
+          { type: "string", format: "email" },
+          { "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on uri input", () => {
+        shouldFocus(
+          { type: "string", format: "uri" },
+          { "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on data-url input", () => {
+        shouldFocus(
+          { type: "string", format: "data-url" },
+          { "ui:autofocus": true }
+        );
+      });
+    });
+
+    describe("object", () => {
+      it("should focus on date input", () => {
+        shouldFocus(
+          { type: "string", format: "date" },
+          { "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on date-time input", () => {
+        shouldFocus(
+          { type: "string", format: "date-time" },
+          { "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on alt-date input", () => {
+        shouldFocus(
+          { type: "string", format: "date" },
+          { "ui:widget": "alt-date", "ui:autofocus": true },
+          "select"
+        );
+      });
+
+      it("should focus on alt-date-time input", () => {
+        shouldFocus(
+          { type: "string", format: "date-time" },
+          { "ui:widget": "alt-datetime", "ui:autofocus": true },
+          "select"
+        );
+      });
+    });
+
+    describe("array", () => {
+      it("should focus on multiple files input", () => {
+        shouldFocus(
+          { type: "array", items: { type: "string", format: "data-url" } },
+          { "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on first item of a list of strings", () => {
+        shouldFocus(
+          { type: "array", items: { type: "string", default: "foo" } },
+          { "ui:autofocus": true },
+          "input",
+          ["foo", "bar"]
+        );
+      });
+
+      it("should focus on first item of a multiple choices list", () => {
+        shouldFocus(
+          {
+            type: "array",
+            items: { type: "string", enum: ["foo", "bar"] },
+            uniqueItems: true,
+          },
+          { "ui:widget": "checkboxes", "ui:autofocus": true },
+          "input",
+          ["bar"]
+        );
+      });
+    });
+
+    describe("boolean", () => {
+      it("should focus on checkbox input", () => {
+        shouldFocus({ type: "boolean" }, { "ui:autofocus": true });
+      });
+
+      it("should focus on radio input", () => {
+        shouldFocus(
+          { type: "boolean" },
+          { "ui:widget": "radio", "ui:autofocus": true }
+        );
+      });
+
+      it("should focus on select input", () => {
+        shouldFocus(
+          { type: "boolean" },
+          { "ui:widget": "select", "ui:autofocus": true },
+          "select"
+        );
+      });
     });
   });
 
@@ -253,68 +567,70 @@ describe("uiSchema", () => {
       properties: {
         foo: {
           type: "string",
-        }
-      }
+        },
+      },
     };
 
     describe("file", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "file"
-        }
+          "ui:widget": "file",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("input[type=file]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("input[type=file]")).to.have.length.of(1);
       });
     });
-
 
     describe("textarea", () => {
       const uiSchema = {
         foo: {
           "ui:widget": "textarea",
           "ui:placeholder": "sample",
-        }
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("textarea"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("textarea")).to.have.length.of(1);
       });
 
-
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "a"
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "a",
+          },
+        });
 
-        expect(node.querySelector("textarea").value)
-          .eql("a");
+        expect(node.querySelector("textarea").value).eql("a");
       });
 
       it("should update state when text is updated", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "a"
-        }});
-
-        Simulate.change(node.querySelector("textarea"), {
-          target: {value: "b"}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "a",
+          },
         });
 
-        expect(comp.state.formData).eql({foo: "b"});
+        Simulate.change(node.querySelector("textarea"), {
+          target: { value: "b" },
+        });
+
+        expect(comp.state.formData).eql({ foo: "b" });
       });
 
       it("should set a placeholder value", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelector("textarea").placeholder)
-          .eql("sample");
+        expect(node.querySelector("textarea").placeholder).eql("sample");
       });
     });
 
@@ -323,108 +639,125 @@ describe("uiSchema", () => {
         foo: {
           "ui:widget": "password",
           "ui:placeholder": "sample",
-        }
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=password]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=password]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "a"
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "a",
+          },
+        });
 
-        expect(node.querySelector("[type=password]").value)
-          .eql("a");
+        expect(node.querySelector("[type=password]").value).eql("a");
       });
 
       it("should update state when text is updated is checked", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "a"
-        }});
-
-        Simulate.change(node.querySelector("[type=password]"), {
-          target: {value: "b"}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "a",
+          },
         });
 
-        expect(comp.state.formData).eql({foo: "b"});
+        Simulate.change(node.querySelector("[type=password]"), {
+          target: { value: "b" },
+        });
+
+        expect(comp.state.formData).eql({ foo: "b" });
       });
 
       it("should set a placeholder value", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelector("[type=password]").placeholder)
-          .eql("sample");
+        expect(node.querySelector("[type=password]").placeholder).eql("sample");
       });
     });
 
     describe("color", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "color"
-        }
+          "ui:widget": "color",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=color]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=color]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "#151ce6"
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "#151ce6",
+          },
+        });
 
-        expect(node.querySelector("[type=color]").value)
-          .eql("#151ce6");
+        expect(node.querySelector("[type=color]").value).eql("#151ce6");
       });
 
       it("should update state when text is updated", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "#151ce6"
-        }});
-
-        Simulate.change(node.querySelector("[type=color]"), {
-          target: {value: "#001122"}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "#151ce6",
+          },
         });
 
-        expect(comp.state.formData).eql({foo: "#001122"});
+        Simulate.change(node.querySelector("[type=color]"), {
+          target: { value: "#001122" },
+        });
+
+        expect(comp.state.formData).eql({ foo: "#001122" });
       });
     });
 
     describe("hidden", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "hidden"
-        }
+          "ui:widget": "hidden",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=hidden]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=hidden]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "a"
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "a",
+          },
+        });
 
-        expect(node.querySelector("[type=hidden]").value)
-          .eql("a");
+        expect(node.querySelector("[type=hidden]").value).eql("a");
       });
 
       it("should map widget value to a typed state one", () => {
-        const {comp} = createFormComponent({schema, uiSchema, formData: {
-          foo: "a"
-        }});
+        const { comp } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "a",
+          },
+        });
 
         expect(comp.state.formData.foo).eql("a");
       });
@@ -438,43 +771,49 @@ describe("uiSchema", () => {
         foo: {
           type: "string",
           enum: ["a", "b"],
-        }
-      }
+        },
+      },
     };
 
     describe("radio", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "radio"
-        }
+          "ui:widget": "radio",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=radio]"))
-          .to.have.length.of(2);
+        expect(node.querySelectorAll("[type=radio]")).to.have.length.of(2);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "b"
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "b",
+          },
+        });
 
-        expect(node.querySelectorAll("[type=radio]")[1].checked)
-          .eql(true);
+        expect(node.querySelectorAll("[type=radio]")[1].checked).eql(true);
       });
 
       it("should update state when value is updated", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: "a"
-        }});
-
-        Simulate.change(node.querySelectorAll("[type=radio]")[1], {
-          target: {checked: true}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: "a",
+          },
         });
 
-        expect(comp.state.formData).eql({foo: "b"});
+        Simulate.change(node.querySelectorAll("[type=radio]")[1], {
+          target: { checked: true },
+        });
+
+        expect(comp.state.formData).eql({ foo: "b" });
       });
     });
   });
@@ -488,50 +827,56 @@ describe("uiSchema", () => {
           multipleOf: 1,
           minimum: 10,
           maximum: 100,
-        }
-      }
+        },
+      },
     };
 
     describe("updown", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "updown"
-        }
+          "ui:widget": "updown",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=number]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=number]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 3.14
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 3.14,
+          },
+        });
 
-        expect(node.querySelector("[type=number]").value)
-          .eql("3.14");
+        expect(node.querySelector("[type=number]").value).eql("3.14");
       });
 
       it("should update state when value is updated", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 3.14
-        }});
-
-        Simulate.change(node.querySelector("[type=number]"), {
-          target: {value: "6.28"}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 3.14,
+          },
         });
 
-        expect(comp.state.formData).eql({foo: 6.28});
+        Simulate.change(node.querySelector("[type=number]"), {
+          target: { value: "6.28" },
+        });
+
+        expect(comp.state.formData).eql({ foo: 6.28 });
       });
 
       describe("Constraint attributes", () => {
         let input;
 
         beforeEach(() => {
-          const {node} = createFormComponent({schema, uiSchema});
+          const { node } = createFormComponent({ schema, uiSchema });
           input = node.querySelector("[type=number]");
         });
 
@@ -543,6 +888,20 @@ describe("uiSchema", () => {
           expect(input.getAttribute("max")).eql("100");
         });
 
+        it("should support '0' as minimum and maximum constraints", () => {
+          const schema = {
+            type: "number",
+            minimum: 0,
+            maximum: 0,
+          };
+          const uiSchema = { "ui:widget": "updown" };
+          const { node } = createFormComponent({ schema, uiSchema });
+          input = node.querySelector("[type=number]");
+
+          expect(input.getAttribute("min")).eql("0");
+          expect(input.getAttribute("max")).eql("0");
+        });
+
         it("should support the multipleOf constraint", () => {
           expect(input.getAttribute("step")).eql("1");
         });
@@ -552,43 +911,49 @@ describe("uiSchema", () => {
     describe("range", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "range"
-        }
+          "ui:widget": "range",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=range]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=range]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 3.14
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 3.14,
+          },
+        });
 
-        expect(node.querySelector("[type=range]").value)
-          .eql("3.14");
+        expect(node.querySelector("[type=range]").value).eql("3.14");
       });
 
       it("should update state when value is updated", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 3.14
-        }});
-
-        Simulate.change(node.querySelector("[type=range]"), {
-          target: {value: "6.28"}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 3.14,
+          },
         });
 
-        expect(comp.state.formData).eql({foo: 6.28});
+        Simulate.change(node.querySelector("[type=range]"), {
+          target: { value: "6.28" },
+        });
+
+        expect(comp.state.formData).eql({ foo: 6.28 });
       });
 
       describe("Constraint attributes", () => {
         let input;
 
         beforeEach(() => {
-          const {node} = createFormComponent({schema, uiSchema});
+          const { node } = createFormComponent({ schema, uiSchema });
           input = node.querySelector("[type=range]");
         });
 
@@ -600,39 +965,111 @@ describe("uiSchema", () => {
           expect(input.getAttribute("max")).eql("100");
         });
 
+        it("should support '0' as minimum and maximum constraints", () => {
+          const schema = {
+            type: "number",
+            minimum: 0,
+            maximum: 0,
+          };
+          const uiSchema = { "ui:widget": "range" };
+          const { node } = createFormComponent({ schema, uiSchema });
+          input = node.querySelector("[type=range]");
+
+          expect(input.getAttribute("min")).eql("0");
+          expect(input.getAttribute("max")).eql("0");
+        });
+
         it("should support the multipleOf constraint", () => {
           expect(input.getAttribute("step")).eql("1");
         });
       });
     });
 
-    describe("hidden", () => {
+    describe("radio", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "number",
+            enum: [3.14159, 2.718, 1.4142],
+          },
+        },
+      };
+
       const uiSchema = {
         foo: {
-          "ui:widget": "hidden"
-        }
+          "ui:widget": "radio",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=hidden]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=radio]")).to.have.length.of(3);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 42
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 2.718,
+          },
+        });
 
-        expect(node.querySelector("[type=hidden]").value)
-          .eql("42");
+        expect(node.querySelectorAll("[type=radio]")[1].checked).eql(true);
+      });
+
+      it("should update state when value is updated", () => {
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 1.4142,
+          },
+        });
+
+        Simulate.change(node.querySelectorAll("[type=radio]")[2], {
+          target: { checked: true },
+        });
+
+        expect(comp.state.formData).eql({ foo: 1.4142 });
+      });
+    });
+
+    describe("hidden", () => {
+      const uiSchema = {
+        foo: {
+          "ui:widget": "hidden",
+        },
+      };
+
+      it("should accept a uiSchema object", () => {
+        const { node } = createFormComponent({ schema, uiSchema });
+
+        expect(node.querySelectorAll("[type=hidden]")).to.have.length.of(1);
+      });
+
+      it("should support formData", () => {
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 42,
+          },
+        });
+
+        expect(node.querySelector("[type=hidden]").value).eql("42");
       });
 
       it("should map widget value to a typed state one", () => {
-        const {comp} = createFormComponent({schema, uiSchema, formData: {
-          foo: 42
-        }});
+        const { comp } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 42,
+          },
+        });
 
         expect(comp.state.formData.foo).eql(42);
       });
@@ -645,109 +1082,179 @@ describe("uiSchema", () => {
       properties: {
         foo: {
           type: "integer",
-        }
-      }
+        },
+      },
     };
 
     describe("updown", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "updown"
-        }
+          "ui:widget": "updown",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=number]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=number]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 3
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 3,
+          },
+        });
 
-        expect(node.querySelector("[type=number]").value)
-          .eql("3");
+        expect(node.querySelector("[type=number]").value).eql("3");
       });
 
       it("should update state when value is updated", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 3
-        }});
-
-        Simulate.change(node.querySelector("[type=number]"), {
-          target: {value: "6"}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 3,
+          },
         });
 
-        expect(comp.state.formData).eql({foo: 6});
+        Simulate.change(node.querySelector("[type=number]"), {
+          target: { value: "6" },
+        });
+
+        expect(comp.state.formData).eql({ foo: 6 });
       });
     });
 
     describe("range", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "range"
-        }
+          "ui:widget": "range",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=range]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=range]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 3
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 3,
+          },
+        });
 
-        expect(node.querySelector("[type=range]").value)
-          .eql("3");
+        expect(node.querySelector("[type=range]").value).eql("3");
       });
 
       it("should update state when value is updated", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 3
-        }});
-
-        Simulate.change(node.querySelector("[type=range]"), {
-          target: {value: "6"}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 3,
+          },
         });
 
-        expect(comp.state.formData).eql({foo: 6});
+        Simulate.change(node.querySelector("[type=range]"), {
+          target: { value: "6" },
+        });
+
+        expect(comp.state.formData).eql({ foo: 6 });
+      });
+    });
+
+    describe("radio", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "integer",
+            enum: [1, 2],
+          },
+        },
+      };
+
+      const uiSchema = {
+        foo: {
+          "ui:widget": "radio",
+        },
+      };
+
+      it("should accept a uiSchema object", () => {
+        const { node } = createFormComponent({ schema, uiSchema });
+
+        expect(node.querySelectorAll("[type=radio]")).to.have.length.of(2);
+      });
+
+      it("should support formData", () => {
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 2,
+          },
+        });
+
+        expect(node.querySelectorAll("[type=radio]")[1].checked).eql(true);
+      });
+
+      it("should update state when value is updated", () => {
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 1,
+          },
+        });
+
+        Simulate.change(node.querySelectorAll("[type=radio]")[1], {
+          target: { checked: true },
+        });
+
+        expect(comp.state.formData).eql({ foo: 2 });
       });
     });
 
     describe("hidden", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "hidden"
-        }
+          "ui:widget": "hidden",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=hidden]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=hidden]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: 42
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 42,
+          },
+        });
 
-        expect(node.querySelector("[type=hidden]").value)
-          .eql("42");
+        expect(node.querySelector("[type=hidden]").value).eql("42");
       });
 
       it("should map widget value to a typed state one", () => {
-        const {comp} = createFormComponent({schema, uiSchema, formData: {
-          foo: 42
-        }});
+        const { comp } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: 42,
+          },
+        });
 
         expect(comp.state.formData.foo).eql(42);
       });
@@ -760,149 +1267,168 @@ describe("uiSchema", () => {
       properties: {
         foo: {
           type: "boolean",
-        }
-      }
+        },
+      },
     };
 
     describe("radio", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "radio"
-        }
+          "ui:widget": "radio",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=radio]"))
-          .to.have.length.of(2);
-        expect(node.querySelectorAll("[type=radio]")[0])
-          .not.eql(null);
-        expect(node.querySelectorAll("[type=radio]")[1])
-          .not.eql(null);
+        expect(node.querySelectorAll("[type=radio]")).to.have.length.of(2);
+        expect(node.querySelectorAll("[type=radio]")[0]).not.eql(null);
+        expect(node.querySelectorAll("[type=radio]")[1]).not.eql(null);
       });
 
       it("should render boolean option labels", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
         const labels = [].map.call(
           node.querySelectorAll(".field-radio-group label"),
-          node => node.textContent);
+          node => node.textContent
+        );
 
-        expect(labels)
-          .eql(["true", "false"]);
+        expect(labels).eql(["yes", "no"]);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: false
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: false,
+          },
+        });
 
-        expect(node.querySelectorAll("[type=radio]")[1].checked)
-          .eql(true);
+        expect(node.querySelectorAll("[type=radio]")[1].checked).eql(true);
       });
 
       it("should update state when false is checked", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: true
-        }});
-
-        Simulate.change(node.querySelectorAll("[type=radio]")[1], {
-          target: {checked: true}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: true,
+          },
         });
 
-        expect(comp.state.formData).eql({foo: false});
+        Simulate.change(node.querySelectorAll("[type=radio]")[1], {
+          target: { checked: true },
+        });
+
+        expect(comp.state.formData).eql({ foo: false });
       });
 
       it("should update state when true is checked", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: false
-        }});
-
-        Simulate.change(node.querySelectorAll("[type=radio]")[0], {
-          target: {checked: true}
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: false,
+          },
         });
 
-        expect(comp.state.formData).eql({foo: true});
+        Simulate.change(node.querySelectorAll("[type=radio]")[0], {
+          target: { checked: true },
+        });
+
+        expect(comp.state.formData).eql({ foo: true });
       });
     });
 
     describe("select", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "select"
-        }
+          "ui:widget": "select",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("select option"))
-          .to.have.length.of(2);
+        expect(node.querySelectorAll("select option")).to.have.length.of(3);
       });
 
       it("should render boolean option labels", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("option")[0].textContent)
-          .eql("true");
-        expect(node.querySelectorAll("option")[1].textContent)
-          .eql("false");
+        expect(node.querySelectorAll("option")[1].textContent).eql("yes");
+        expect(node.querySelectorAll("option")[2].textContent).eql("no");
       });
 
       it("should update state when true is selected", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: false
-        }});
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: false,
+          },
+        });
 
         Simulate.change(node.querySelector("select"), {
           // DOM option change events always return strings
-          target: {value: "true"}
+          target: { value: "true" },
         });
 
-        expect(comp.state.formData).eql({foo: true});
+        expect(comp.state.formData).eql({ foo: true });
       });
 
       it("should update state when false is selected", () => {
-        const {comp, node} = createFormComponent({schema, uiSchema, formData: {
-          foo: false
-        }});
+        const { comp, node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: false,
+          },
+        });
 
         Simulate.change(node.querySelector("select"), {
           // DOM option change events always return strings
-          target: {value: "false"}
+          target: { value: "false" },
         });
 
-        expect(comp.state.formData).eql({foo: false});
+        expect(comp.state.formData).eql({ foo: false });
       });
     });
 
     describe("hidden", () => {
       const uiSchema = {
         foo: {
-          "ui:widget": "hidden"
-        }
+          "ui:widget": "hidden",
+        },
       };
 
       it("should accept a uiSchema object", () => {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
 
-        expect(node.querySelectorAll("[type=hidden]"))
-          .to.have.length.of(1);
+        expect(node.querySelectorAll("[type=hidden]")).to.have.length.of(1);
       });
 
       it("should support formData", () => {
-        const {node} = createFormComponent({schema, uiSchema, formData: {
-          foo: true
-        }});
+        const { node } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: true,
+          },
+        });
 
-        expect(node.querySelector("[type=hidden]").value)
-          .eql("true");
+        expect(node.querySelector("[type=hidden]").value).eql("true");
       });
 
       it("should map widget value to a typed state one", () => {
-        const {comp} = createFormComponent({schema, uiSchema, formData: {
-          foo: true
-        }});
+        const { comp } = createFormComponent({
+          schema,
+          uiSchema,
+          formData: {
+            foo: true,
+          },
+        });
 
         expect(comp.state.formData.foo).eql(true);
       });
@@ -914,31 +1440,36 @@ describe("uiSchema", () => {
       const schema = {
         type: "object",
         properties: {
-          foo: {type: "string"},
-          bar: {type: "string"},
-        }
+          foo: { type: "string" },
+          bar: { type: "string" },
+        },
       };
-      const uiSchema = {"ui:rootFieldId": "myform"};
-      const {node} = createFormComponent({schema, uiSchema});
+      const uiSchema = { "ui:rootFieldId": "myform" };
+      const { node } = createFormComponent({ schema, uiSchema });
 
-      const ids = [].map.call(node.querySelectorAll("input[type=text]"),
-                              node => node.id);
+      const ids = [].map.call(
+        node.querySelectorAll("input[type=text]"),
+        node => node.id
+      );
       expect(ids).eql(["myform_foo", "myform_bar"]);
     });
 
     it("should use a custom root field id for arrays", () => {
       const schema = {
         type: "array",
-        items: {type: "string"},
+        items: { type: "string" },
       };
-      const uiSchema = {"ui:rootFieldId": "myform"};
-      const {node} = createFormComponent({schema, uiSchema, formData: [
-        "foo",
-        "bar"
-      ]});
+      const uiSchema = { "ui:rootFieldId": "myform" };
+      const { node } = createFormComponent({
+        schema,
+        uiSchema,
+        formData: ["foo", "bar"],
+      });
 
-      const ids = [].map.call(node.querySelectorAll("input[type=text]"),
-                              node => node.id);
+      const ids = [].map.call(
+        node.querySelectorAll("input[type=text]"),
+        node => node.id
+      );
       expect(ids).eql(["myform_0", "myform_1"]);
     });
 
@@ -948,24 +1479,27 @@ describe("uiSchema", () => {
         items: {
           type: "object",
           properties: {
-            foo: {type: "string"},
-            bar: {type: "string"},
-          }
+            foo: { type: "string" },
+            bar: { type: "string" },
+          },
         },
       };
-      const uiSchema = {"ui:rootFieldId": "myform"};
-      const {node} = createFormComponent({schema, uiSchema, formData: [
-        {foo: "foo1", bar: "bar1"},
-        {foo: "foo2", bar: "bar2"},
-      ]});
+      const uiSchema = { "ui:rootFieldId": "myform" };
+      const { node } = createFormComponent({
+        schema,
+        uiSchema,
+        formData: [{ foo: "foo1", bar: "bar1" }, { foo: "foo2", bar: "bar2" }],
+      });
 
-      const ids = [].map.call(node.querySelectorAll("input[type=text]"),
-                              node => node.id);
+      const ids = [].map.call(
+        node.querySelectorAll("input[type=text]"),
+        node => node.id
+      );
       expect(ids).eql([
         "myform_0_foo",
         "myform_0_bar",
         "myform_1_foo",
-        "myform_1_bar"
+        "myform_1_bar",
       ]);
     });
   });
@@ -976,28 +1510,30 @@ describe("uiSchema", () => {
         let node;
 
         beforeEach(() => {
-          const schema = {type: "array", items: {type: "string"}};
-          const uiSchema = {"ui:disabled": true};
+          const schema = { type: "array", items: { type: "string" } };
+          const uiSchema = { "ui:disabled": true };
           const formData = ["a", "b"];
 
-          let rendered = createFormComponent({schema, uiSchema, formData});
+          let rendered = createFormComponent({ schema, uiSchema, formData });
           node = rendered.node;
         });
 
         it("should disable an ArrayField", () => {
-          const disabled = [].map.call(node.querySelectorAll("[type=text]"),
-                                       node => node.disabled);
+          const disabled = [].map.call(
+            node.querySelectorAll("[type=text]"),
+            node => node.disabled
+          );
           expect(disabled).eql([true, true]);
         });
 
         it("should disable the Add button", () => {
-          expect(node.querySelector(".array-item-add button").disabled)
-            .eql(true);
+          expect(node.querySelector(".array-item-add button").disabled).eql(
+            true
+          );
         });
 
         it("should disable the Delete button", () => {
-          expect(node.querySelector(".array-item-remove").disabled)
-            .eql(true);
+          expect(node.querySelector(".array-item-remove").disabled).eql(true);
         });
       });
 
@@ -1008,19 +1544,21 @@ describe("uiSchema", () => {
           const schema = {
             type: "object",
             properties: {
-              foo: {type: "string"},
-              bar: {type: "string"},
-            }
+              foo: { type: "string" },
+              bar: { type: "string" },
+            },
           };
-          const uiSchema = {"ui:disabled": true};
+          const uiSchema = { "ui:disabled": true };
 
-          let rendered = createFormComponent({schema, uiSchema});
+          let rendered = createFormComponent({ schema, uiSchema });
           node = rendered.node;
         });
 
         it("should disable an ObjectField", () => {
-          const disabled = [].map.call(node.querySelectorAll("[type=text]"),
-                                       node => node.disabled);
+          const disabled = [].map.call(
+            node.querySelectorAll("[type=text]"),
+            node => node.disabled
+          );
           expect(disabled).eql([true, true]);
         });
       });
@@ -1028,115 +1566,147 @@ describe("uiSchema", () => {
 
     describe("Widgets", () => {
       function shouldBeDisabled(selector, schema, uiSchema) {
-        const {node} = createFormComponent({schema, uiSchema});
+        const { node } = createFormComponent({ schema, uiSchema });
         expect(node.querySelector(selector).disabled).eql(true);
       }
 
       it("should disable a text widget", () => {
-        shouldBeDisabled("input[type=text]",
-                         {type: "string"},
-                         {"ui:disabled": true});
+        shouldBeDisabled(
+          "input[type=text]",
+          { type: "string" },
+          { "ui:disabled": true }
+        );
       });
 
       it("should disabled a file widget", () => {
-        const {node} = createFormComponent({
-          schema: {type: "string", format: "data-url"},
-          uiSchema: {"ui:disabled": true}});
-        expect(node.querySelector("input[type=file]").hasAttribute("disabled"))
-          .eql(true);
+        const { node } = createFormComponent({
+          schema: { type: "string", format: "data-url" },
+          uiSchema: { "ui:disabled": true },
+        });
+        expect(
+          node.querySelector("input[type=file]").hasAttribute("disabled")
+        ).eql(true);
       });
 
       it("should disable a textarea widget", () => {
-        shouldBeDisabled("textarea",
-                         {type: "string"},
-                         {"ui:disabled": true, "ui:widget": "textarea"});
+        shouldBeDisabled(
+          "textarea",
+          { type: "string" },
+          { "ui:disabled": true, "ui:widget": "textarea" }
+        );
       });
 
       it("should disable a number text widget", () => {
-        shouldBeDisabled("input[type=text]",
-                         {type: "number"},
-                         {"ui:disabled": true});
+        shouldBeDisabled(
+          "input[type=text]",
+          { type: "number" },
+          { "ui:disabled": true }
+        );
       });
 
       it("should disable a number widget", () => {
-        shouldBeDisabled("input[type=number]",
-                         {type: "number"},
-                         {"ui:disabled": true, "ui:widget": "updown"});
+        shouldBeDisabled(
+          "input[type=number]",
+          { type: "number" },
+          { "ui:disabled": true, "ui:widget": "updown" }
+        );
       });
 
       it("should disable a range widget", () => {
-        shouldBeDisabled("input[type=range]",
-                         {type: "number"},
-                         {"ui:disabled": true, "ui:widget": "range"});
+        shouldBeDisabled(
+          "input[type=range]",
+          { type: "number" },
+          { "ui:disabled": true, "ui:widget": "range" }
+        );
       });
 
       it("should disable a select widget", () => {
-        shouldBeDisabled("select",
-                         {type: "string", enum: ["a", "b"]},
-                         {"ui:disabled": true});
+        shouldBeDisabled(
+          "select",
+          { type: "string", enum: ["a", "b"] },
+          { "ui:disabled": true }
+        );
       });
 
       it("should disable a checkbox widget", () => {
-        shouldBeDisabled("input[type=checkbox]",
-                         {type: "boolean"},
-                         {"ui:disabled": true});
+        shouldBeDisabled(
+          "input[type=checkbox]",
+          { type: "boolean" },
+          { "ui:disabled": true }
+        );
       });
 
       it("should disable a radio widget", () => {
-        shouldBeDisabled("input[type=radio]",
-                         {type: "boolean"},
-                         {"ui:disabled": true, "ui:widget": "radio"});
+        shouldBeDisabled(
+          "input[type=radio]",
+          { type: "boolean" },
+          { "ui:disabled": true, "ui:widget": "radio" }
+        );
       });
 
       it("should disable a color widget", () => {
-        shouldBeDisabled("input[type=color]",
-                         {type: "string", format: "color"},
-                         {"ui:disabled": true});
+        shouldBeDisabled(
+          "input[type=color]",
+          { type: "string", format: "color" },
+          { "ui:disabled": true }
+        );
       });
 
       it("should disable a password widget", () => {
-        shouldBeDisabled("input[type=password]",
-                         {type: "string"},
-                         {"ui:disabled": true, "ui:widget": "password"});
+        shouldBeDisabled(
+          "input[type=password]",
+          { type: "string" },
+          { "ui:disabled": true, "ui:widget": "password" }
+        );
       });
 
       it("should disable an email widget", () => {
-        shouldBeDisabled("input[type=email]",
-                         {type: "string", format: "email"},
-                         {"ui:disabled": true});
+        shouldBeDisabled(
+          "input[type=email]",
+          { type: "string", format: "email" },
+          { "ui:disabled": true }
+        );
       });
 
       it("should disable a date widget", () => {
-        shouldBeDisabled("input[type=date]",
-                         {type: "string", format: "date"},
-                         {"ui:disabled": true});
+        shouldBeDisabled(
+          "input[type=date]",
+          { type: "string", format: "date" },
+          { "ui:disabled": true }
+        );
       });
 
       it("should disable a datetime widget", () => {
-        shouldBeDisabled("input[type=datetime-local]",
-                         {type: "string", format: "date-time"},
-                         {"ui:disabled": true});
+        shouldBeDisabled(
+          "input[type=datetime-local]",
+          { type: "string", format: "date-time" },
+          { "ui:disabled": true }
+        );
       });
 
       it("should disable an alternative date widget", () => {
-        const {node} = createFormComponent({
-          schema: {type: "string", format: "date"},
-          uiSchema: {"ui:disabled": true, "ui:widget": "alt-date"}
+        const { node } = createFormComponent({
+          schema: { type: "string", format: "date" },
+          uiSchema: { "ui:disabled": true, "ui:widget": "alt-date" },
         });
 
-        const disabled = [].map.call(node.querySelectorAll("select"),
-                                     node => node.disabled);
+        const disabled = [].map.call(
+          node.querySelectorAll("select"),
+          node => node.disabled
+        );
         expect(disabled).eql([true, true, true]);
       });
 
       it("should disable an alternative datetime widget", () => {
-        const {node} = createFormComponent({
-          schema: {type: "string", format: "date-time"},
-          uiSchema: {"ui:disabled": true, "ui:widget": "alt-datetime"}
+        const { node } = createFormComponent({
+          schema: { type: "string", format: "date-time" },
+          uiSchema: { "ui:disabled": true, "ui:widget": "alt-datetime" },
         });
 
-        const disabled = [].map.call(node.querySelectorAll("select"),
-                                     node => node.disabled);
+        const disabled = [].map.call(
+          node.querySelectorAll("select"),
+          node => node.disabled
+        );
         expect(disabled).eql([true, true, true, true, true, true]);
       });
     });
@@ -1148,28 +1718,30 @@ describe("uiSchema", () => {
         let node;
 
         beforeEach(() => {
-          const schema = {type: "array", items: {type: "string"}};
-          const uiSchema = {"ui:readonly": true};
+          const schema = { type: "array", items: { type: "string" } };
+          const uiSchema = { "ui:readonly": true };
           const formData = ["a", "b"];
 
-          let rendered = createFormComponent({schema, uiSchema, formData});
+          let rendered = createFormComponent({ schema, uiSchema, formData });
           node = rendered.node;
         });
 
         it("should mark as readonly an ArrayField", () => {
-          const disabled = [].map.call(node.querySelectorAll("[type=text]"),
-                                       node => node.hasAttribute("readonly"));
+          const disabled = [].map.call(
+            node.querySelectorAll("[type=text]"),
+            node => node.hasAttribute("readonly")
+          );
           expect(disabled).eql([true, true]);
         });
 
         it("should disable the Add button", () => {
-          expect(node.querySelector(".array-item-add button").disabled)
-            .eql(true);
+          expect(node.querySelector(".array-item-add button").disabled).eql(
+            true
+          );
         });
 
         it("should disable the Delete button", () => {
-          expect(node.querySelector(".array-item-remove").disabled)
-            .eql(true);
+          expect(node.querySelector(".array-item-remove").disabled).eql(true);
         });
       });
 
@@ -1180,19 +1752,21 @@ describe("uiSchema", () => {
           const schema = {
             type: "object",
             properties: {
-              foo: {type: "string"},
-              bar: {type: "string"},
-            }
+              foo: { type: "string" },
+              bar: { type: "string" },
+            },
           };
-          const uiSchema = {"ui:readonly": true};
+          const uiSchema = { "ui:readonly": true };
 
-          let rendered = createFormComponent({schema, uiSchema});
+          let rendered = createFormComponent({ schema, uiSchema });
           node = rendered.node;
         });
 
         it("should mark as readonly an ObjectField", () => {
-          const disabled = [].map.call(node.querySelectorAll("[type=text]"),
-                                       node => node.hasAttribute("readonly"));
+          const disabled = [].map.call(
+            node.querySelectorAll("[type=text]"),
+            node => node.hasAttribute("readonly")
+          );
           expect(disabled).eql([true, true]);
         });
       });
@@ -1200,111 +1774,142 @@ describe("uiSchema", () => {
 
     describe("Widgets", () => {
       function shouldBeReadonly(selector, schema, uiSchema) {
-        const {node} = createFormComponent({schema, uiSchema});
-        expect(node.querySelector(selector).hasAttribute("readonly"))
-          .eql(true);
+        const { node } = createFormComponent({ schema, uiSchema });
+        expect(node.querySelector(selector).hasAttribute("readonly")).eql(true);
+      }
+      function shouldBeDisabled(selector, schema, uiSchema) {
+        const { node } = createFormComponent({ schema, uiSchema });
+        expect(node.querySelector(selector).disabled).eql(true);
       }
 
       it("should mark as readonly a text widget", () => {
-        shouldBeReadonly("input[type=text]",
-                         {type: "string"},
-                         {"ui:readonly": true});
+        shouldBeReadonly(
+          "input[type=text]",
+          { type: "string" },
+          { "ui:readonly": true }
+        );
       });
 
       it("should mark as readonly a file widget", () => {
         // We mark a file widget as readonly by disabling it.
-        const {node} = createFormComponent({
-          schema: {type: "string", format: "data-url"},
-          uiSchema: {"ui:readonly": true}});
-        expect(node.querySelector("input[type=file]").hasAttribute("disabled"))
-          .eql(true);
+        const { node } = createFormComponent({
+          schema: { type: "string", format: "data-url" },
+          uiSchema: { "ui:readonly": true },
+        });
+        expect(
+          node.querySelector("input[type=file]").hasAttribute("disabled")
+        ).eql(true);
       });
 
       it("should mark as readonly a textarea widget", () => {
-        shouldBeReadonly("textarea",
-                         {type: "string"},
-                         {"ui:readonly": true, "ui:widget": "textarea"});
+        shouldBeReadonly(
+          "textarea",
+          { type: "string" },
+          { "ui:readonly": true, "ui:widget": "textarea" }
+        );
       });
 
       it("should mark as readonly a number text widget", () => {
-        shouldBeReadonly("input[type=text]",
-                         {type: "number"},
-                         {"ui:readonly": true});
+        shouldBeReadonly(
+          "input[type=text]",
+          { type: "number" },
+          { "ui:readonly": true }
+        );
       });
 
       it("should mark as readonly a number widget", () => {
-        shouldBeReadonly("input[type=number]",
-                         {type: "number"},
-                         {"ui:readonly": true, "ui:widget": "updown"});
+        shouldBeReadonly(
+          "input[type=number]",
+          { type: "number" },
+          { "ui:readonly": true, "ui:widget": "updown" }
+        );
       });
 
       it("should mark as readonly a range widget", () => {
-        shouldBeReadonly("input[type=range]",
-                         {type: "number"},
-                         {"ui:readonly": true, "ui:widget": "range"});
+        shouldBeReadonly(
+          "input[type=range]",
+          { type: "number" },
+          { "ui:readonly": true, "ui:widget": "range" }
+        );
       });
 
-      it("should mark as readonly a select widget", () => {
-        shouldBeReadonly("select",
-                         {type: "string", enum: ["a", "b"]},
-                         {"ui:readonly": true});
+      it("should mark readonly as disabled on a select widget", () => {
+        shouldBeDisabled(
+          "select",
+          { type: "string", enum: ["a", "b"] },
+          { "ui:readonly": true }
+        );
       });
 
       it("should mark as readonly a color widget", () => {
-        shouldBeReadonly("input[type=color]",
-                         {type: "string", format: "color"},
-                         {"ui:readonly": true});
+        shouldBeReadonly(
+          "input[type=color]",
+          { type: "string", format: "color" },
+          { "ui:readonly": true }
+        );
       });
 
       it("should mark as readonly a password widget", () => {
-        shouldBeReadonly("input[type=password]",
-                         {type: "string"},
-                         {"ui:readonly": true, "ui:widget": "password"});
+        shouldBeReadonly(
+          "input[type=password]",
+          { type: "string" },
+          { "ui:readonly": true, "ui:widget": "password" }
+        );
       });
 
       it("should mark as readonly a url widget", () => {
-        shouldBeReadonly("input[type=url]",
-                         {type: "string", format: "uri"},
-                         {"ui:readonly": true});
+        shouldBeReadonly(
+          "input[type=url]",
+          { type: "string", format: "uri" },
+          { "ui:readonly": true }
+        );
       });
 
       it("should mark as readonly an email widget", () => {
-        shouldBeReadonly("input[type=email]",
-                         {type: "string", format: "email"},
-                         {"ui:readonly": true});
+        shouldBeReadonly(
+          "input[type=email]",
+          { type: "string", format: "email" },
+          { "ui:readonly": true }
+        );
       });
 
       it("should mark as readonly a date widget", () => {
-        shouldBeReadonly("input[type=date]",
-                         {type: "string", format: "date"},
-                         {"ui:readonly": true});
+        shouldBeReadonly(
+          "input[type=date]",
+          { type: "string", format: "date" },
+          { "ui:readonly": true }
+        );
       });
 
       it("should mark as readonly a datetime widget", () => {
-        shouldBeReadonly("input[type=datetime-local]",
-                         {type: "string", format: "date-time"},
-                         {"ui:readonly": true});
+        shouldBeReadonly(
+          "input[type=datetime-local]",
+          { type: "string", format: "date-time" },
+          { "ui:readonly": true }
+        );
       });
 
-      it("should mark as readonly an alternative date widget", () => {
-        const {node} = createFormComponent({
-          schema: {type: "string", format: "date"},
-          uiSchema: {"ui:readonly": true, "ui:widget": "alt-date"}
+      it("should mark readonly as disabled on an alternative date widget", () => {
+        const { node } = createFormComponent({
+          schema: { type: "string", format: "date" },
+          uiSchema: { "ui:readonly": true, "ui:widget": "alt-date" },
         });
 
-        const readonly = [].map.call(node.querySelectorAll("select"),
-                                     node => node.hasAttribute("readonly"));
+        const readonly = [].map.call(node.querySelectorAll("select"), node =>
+          node.hasAttribute("disabled")
+        );
         expect(readonly).eql([true, true, true]);
       });
 
-      it("should mark as readonly an alternative datetime widget", () => {
-        const {node} = createFormComponent({
-          schema: {type: "string", format: "date-time"},
-          uiSchema: {"ui:readonly": true, "ui:widget": "alt-datetime"}
+      it("should mark readonly as disabled on an alternative datetime widget", () => {
+        const { node } = createFormComponent({
+          schema: { type: "string", format: "date-time" },
+          uiSchema: { "ui:readonly": true, "ui:widget": "alt-datetime" },
         });
 
-        const readonly = [].map.call(node.querySelectorAll("select"),
-                                     node => node.hasAttribute("readonly"));
+        const readonly = [].map.call(node.querySelectorAll("select"), node =>
+          node.hasAttribute("disabled")
+        );
         expect(readonly).eql([true, true, true, true, true, true]);
       });
     });
