@@ -1,3 +1,4 @@
+import React from "react";
 import { expect } from "chai";
 import sinon from "sinon";
 import { Simulate } from "react-addons-test-utils";
@@ -5,33 +6,42 @@ import { Simulate } from "react-addons-test-utils";
 import validateFormData, { toErrorList } from "../src/validate";
 import { createFormComponent } from "./test_utils";
 
-
 describe("Validation", () => {
   describe("validate.validateFormData()", () => {
     describe("No custom validate function", () => {
+      const illFormedKey = "bar.'\"[]()=+*&^%$#@!";
       const schema = {
         type: "object",
         properties: {
-          foo: {type: "string"}
-        }
+          foo: { type: "string" },
+          [illFormedKey]: { type: "string" },
+        },
       };
 
       let errors, errorSchema;
 
       beforeEach(() => {
-        const result = validateFormData({foo: 42}, schema);
+        const result = validateFormData(
+          { foo: 42, [illFormedKey]: 41 },
+          schema
+        );
         errors = result.errors;
         errorSchema = result.errorSchema;
       });
 
       it("should return an error list", () => {
-        expect(errors).to.have.length.of(1);
+        expect(errors).to.have.length.of(2);
         expect(errors[0].message).eql("is not of a type(s) string");
+        expect(errors[1].message).eql("is not of a type(s) string");
       });
 
       it("should return an errorSchema", () => {
         expect(errorSchema.foo.__errors).to.have.length.of(1);
         expect(errorSchema.foo.__errors[0]).eql("is not of a type(s) string");
+        expect(errorSchema[illFormedKey].__errors).to.have.length.of(1);
+        expect(errorSchema[illFormedKey].__errors[0]).eql(
+          "is not of a type(s) string"
+        );
       });
     });
 
@@ -42,9 +52,9 @@ describe("Validation", () => {
         type: "object",
         required: ["pass1", "pass2"],
         properties: {
-          pass1: {type: "string"},
-          pass2: {type: "string"},
-        }
+          pass1: { type: "string" },
+          pass2: { type: "string" },
+        },
       };
 
       beforeEach(() => {
@@ -54,7 +64,7 @@ describe("Validation", () => {
           }
           return errors;
         };
-        const formData = {pass1: "a", pass2: "b"};
+        const formData = { pass1: "a", pass2: "b" };
         const result = validateFormData(formData, schema, validate);
         errors = result.errors;
         errorSchema = result.errorSchema;
@@ -73,23 +83,57 @@ describe("Validation", () => {
 
     describe("toErrorList()", () => {
       it("should convert an errorSchema into a flat list", () => {
-        expect(toErrorList({
-          __errors: ["err1", "err2"],
-          a: {
-            b: {
-              __errors: ["err3", "err4"]
-            }
-          },
-          c: {
-            __errors: ["err5"]
-          }
-        })).eql([
-          {stack: "root: err1"},
-          {stack: "root: err2"},
-          {stack: "b: err3"},
-          {stack: "b: err4"},
-          {stack: "c: err5"},
+        expect(
+          toErrorList({
+            __errors: ["err1", "err2"],
+            a: {
+              b: {
+                __errors: ["err3", "err4"],
+              },
+            },
+            c: {
+              __errors: ["err5"],
+            },
+          })
+        ).eql([
+          { stack: "root: err1" },
+          { stack: "root: err2" },
+          { stack: "b: err3" },
+          { stack: "b: err4" },
+          { stack: "c: err5" },
         ]);
+      });
+    });
+
+    describe("transformErrors", () => {
+      const illFormedKey = "bar.'\"[]()=+*&^%$#@!";
+      const schema = {
+        type: "object",
+        properties: {
+          foo: { type: "string" },
+          [illFormedKey]: { type: "string" },
+        },
+      };
+      const newErrorMessage = "Better error message";
+      const transformErrors = errors => {
+        return [Object.assign({}, errors[0], { message: newErrorMessage })];
+      };
+
+      let errors;
+
+      beforeEach(() => {
+        const result = validateFormData(
+          { foo: 42, [illFormedKey]: 41 },
+          schema,
+          undefined,
+          transformErrors
+        );
+        errors = result.errors;
+      });
+
+      it("should use transformErrors function", () => {
+        expect(errors).not.to.be.empty;
+        expect(errors[0].message).to.equal(newErrorMessage);
       });
     });
   });
@@ -111,18 +155,22 @@ describe("Validation", () => {
           type: "object",
           required: ["foo"],
           properties: {
-            foo: {type: "string"},
-            bar: {type: "string"},
-          }
+            foo: { type: "string" },
+            bar: { type: "string" },
+          },
         };
 
         var comp, node, onError;
 
         beforeEach(() => {
           onError = sandbox.spy();
-          const compInfo = createFormComponent({schema, formData: {
-            foo: undefined
-          }, onError});
+          const compInfo = createFormComponent({
+            schema,
+            formData: {
+              foo: undefined,
+            },
+            onError,
+          });
           comp = compInfo.comp;
           node = compInfo.node;
 
@@ -130,23 +178,24 @@ describe("Validation", () => {
         });
 
         it("should validate a required field", () => {
-          expect(comp.state.errors)
-            .to.have.length.of(1);
-          expect(comp.state.errors[0].message)
-            .eql(`requires property "foo"`);
+          expect(comp.state.errors).to.have.length.of(1);
+          expect(comp.state.errors[0].message).eql('requires property "foo"');
         });
 
         it("should render errors", () => {
-          expect(node.querySelectorAll(".errors li"))
-            .to.have.length.of(1);
-          expect(node.querySelector(".errors li").textContent)
-            .eql(`instance requires property "foo"`);
+          expect(node.querySelectorAll(".errors li")).to.have.length.of(1);
+          expect(node.querySelector(".errors li").textContent).eql(
+            'instance requires property "foo"'
+          );
         });
 
         it("should trigger the onError handler", () => {
-          sinon.assert.calledWith(onError, sinon.match(errors => {
-            return errors[0].message === `requires property "foo"`;
-          }));
+          sinon.assert.calledWith(
+            onError,
+            sinon.match(errors => {
+              return errors[0].message === 'requires property "foo"';
+            })
+          );
         });
       });
 
@@ -159,16 +208,20 @@ describe("Validation", () => {
               type: "string",
               minLength: 10,
             },
-          }
+          },
         };
 
         var comp, node, onError;
 
         beforeEach(() => {
           onError = sandbox.spy();
-          const compInfo = createFormComponent({schema, formData: {
-            foo: "123456789"
-          }, onError});
+          const compInfo = createFormComponent({
+            schema,
+            formData: {
+              foo: "123456789",
+            },
+            onError,
+          });
           comp = compInfo.comp;
           node = compInfo.node;
 
@@ -176,31 +229,33 @@ describe("Validation", () => {
         });
 
         it("should validate a minLength field", () => {
-          expect(comp.state.errors)
-            .to.have.length.of(1);
-          expect(comp.state.errors[0].message)
-            .eql(`does not meet minimum length of 10`);
+          expect(comp.state.errors).to.have.length.of(1);
+          expect(comp.state.errors[0].message).eql(
+            "does not meet minimum length of 10"
+          );
         });
 
         it("should render errors", () => {
-          expect(node.querySelectorAll(".errors li"))
-            .to.have.length.of(1);
-          expect(node.querySelector(".errors li").textContent)
-            .eql("instance.foo does not meet minimum length of 10");
+          expect(node.querySelectorAll(".errors li")).to.have.length.of(1);
+          expect(node.querySelector(".errors li").textContent).eql(
+            "instance.foo does not meet minimum length of 10"
+          );
         });
 
         it("should trigger the onError handler", () => {
-          sinon.assert.calledWith(onError, sinon.match(errors => {
-            return errors[0].message ===
-              "does not meet minimum length of 10";
-          }));
+          sinon.assert.calledWith(
+            onError,
+            sinon.match(errors => {
+              return errors[0].message === "does not meet minimum length of 10";
+            })
+          );
         });
       });
     });
 
     describe("Custom Form validation", () => {
       it("should validate a simple string value", () => {
-        const schema = {type: "string"};
+        const schema = { type: "string" };
         const formData = "a";
 
         function validate(formData, errors) {
@@ -210,8 +265,12 @@ describe("Validation", () => {
           return errors;
         }
 
-        const {comp} = createFormComponent({schema, validate, liveValidate: true});
-        comp.componentWillReceiveProps({formData});
+        const { comp } = createFormComponent({
+          schema,
+          validate,
+          liveValidate: true,
+        });
+        comp.componentWillReceiveProps({ formData });
 
         expect(comp.state.errorSchema).eql({
           __errors: ["Invalid"],
@@ -219,7 +278,7 @@ describe("Validation", () => {
       });
 
       it("should submit form on valid data", () => {
-        const schema = {type: "string"};
+        const schema = { type: "string" };
         const formData = "hello";
         const onSubmit = sandbox.spy();
 
@@ -230,7 +289,7 @@ describe("Validation", () => {
           return errors;
         }
 
-        const {node} = createFormComponent({
+        const { node } = createFormComponent({
           schema,
           formData,
           validate,
@@ -243,7 +302,7 @@ describe("Validation", () => {
       });
 
       it("should prevent form submission on invalid data", () => {
-        const schema = {type: "string"};
+        const schema = { type: "string" };
         const formData = "a";
         const onSubmit = sandbox.spy();
         const onError = sandbox.spy();
@@ -255,7 +314,7 @@ describe("Validation", () => {
           return errors;
         }
 
-        const {node} = createFormComponent({
+        const { node } = createFormComponent({
           schema,
           formData,
           validate,
@@ -273,23 +332,27 @@ describe("Validation", () => {
         const schema = {
           type: "object",
           properties: {
-            pass1: {type: "string", minLength: 3},
-            pass2: {type: "string", minLength: 3},
-          }
+            pass1: { type: "string", minLength: 3 },
+            pass2: { type: "string", minLength: 3 },
+          },
         };
 
-        const formData = {pass1: "aaa", pass2: "b"};
+        const formData = { pass1: "aaa", pass2: "b" };
 
         function validate(formData, errors) {
-          const {pass1, pass2} = formData;
+          const { pass1, pass2 } = formData;
           if (pass1 !== pass2) {
             errors.pass2.addError("Passwords don't match");
           }
           return errors;
         }
 
-        const {comp} = createFormComponent({schema, validate, liveValidate: true});
-        comp.componentWillReceiveProps({formData});
+        const { comp } = createFormComponent({
+          schema,
+          validate,
+          liveValidate: true,
+        });
+        comp.componentWillReceiveProps({ formData });
 
         expect(comp.state.errorSchema).eql({
           __errors: [],
@@ -300,8 +363,64 @@ describe("Validation", () => {
             __errors: [
               "does not meet minimum length of 3",
               "Passwords don't match",
-            ]
-          }
+            ],
+          },
+        });
+      });
+
+      it("should validate an array of object", () => {
+        const schema = {
+          type: "array",
+          items: {
+            type: "object",
+            properties: {
+              pass1: { type: "string" },
+              pass2: { type: "string" },
+            },
+          },
+        };
+
+        const formData = [
+          { pass1: "a", pass2: "b" },
+          { pass1: "a", pass2: "a" },
+        ];
+
+        function validate(formData, errors) {
+          formData.forEach(({ pass1, pass2 }, i) => {
+            if (pass1 !== pass2) {
+              errors[i].pass2.addError("Passwords don't match");
+            }
+          });
+          return errors;
+        }
+
+        const { comp } = createFormComponent({
+          schema,
+          validate,
+          liveValidate: true,
+        });
+        comp.componentWillReceiveProps({ formData });
+
+        expect(comp.state.errorSchema).eql({
+          0: {
+            pass1: {
+              __errors: [],
+            },
+            pass2: {
+              __errors: ["Passwords don't match"],
+            },
+            __errors: [],
+          },
+          1: {
+            pass1: {
+              __errors: [],
+            },
+            pass2: {
+              __errors: [],
+            },
+            __errors: [],
+          },
+          __errors: [],
         });
       });
 
@@ -309,8 +428,8 @@ describe("Validation", () => {
         const schema = {
           type: "array",
           items: {
-            type: "string"
-          }
+            type: "string",
+          },
         };
 
         const formData = ["aaa", "bbb", "ccc"];
@@ -322,10 +441,17 @@ describe("Validation", () => {
           return errors;
         }
 
-        const {comp} = createFormComponent({schema, validate, liveValidate: true});
-        comp.componentWillReceiveProps({formData});
+        const { comp } = createFormComponent({
+          schema,
+          validate,
+          liveValidate: true,
+        });
+        comp.componentWillReceiveProps({ formData });
 
         expect(comp.state.errorSchema).eql({
+          0: { __errors: [] },
+          1: { __errors: [] },
+          2: { __errors: [] },
           __errors: ["Forbidden value: bbb"],
         });
       });
@@ -337,18 +463,23 @@ describe("Validation", () => {
           type: "object",
           required: ["foo"],
           properties: {
-            foo: {type: "string"},
-            bar: {type: "string"},
-          }
+            foo: { type: "string" },
+            bar: { type: "string" },
+          },
         };
 
         var comp, node, onError;
 
         beforeEach(() => {
           onError = sandbox.spy();
-          const compInfo = createFormComponent({schema, formData: {
-            foo: undefined
-          }, onError, showErrorList: false});
+          const compInfo = createFormComponent({
+            schema,
+            formData: {
+              foo: undefined,
+            },
+            onError,
+            showErrorList: false,
+          });
           comp = compInfo.comp;
           node = compInfo.node;
 
@@ -356,25 +487,51 @@ describe("Validation", () => {
         });
 
         it("should validate a required field", () => {
-          expect(comp.state.errors)
-            .to.have.length.of(1);
-          expect(comp.state.errors[0].message)
-            .eql(`requires property "foo"`);
+          expect(comp.state.errors).to.have.length.of(1);
+          expect(comp.state.errors[0].message).eql('requires property "foo"');
         });
 
         it("should not render error list if showErrorList prop true", () => {
-          expect(node.querySelectorAll(".errors li"))
-            .to.have.length.of(0);
+          expect(node.querySelectorAll(".errors li")).to.have.length.of(0);
         });
 
         it("should trigger the onError handler", () => {
-          sinon.assert.calledWith(onError, sinon.match(errors => {
-            return errors[0].message === `requires property "foo"`;
-          }));
+          sinon.assert.calledWith(
+            onError,
+            sinon.match(errors => {
+              return errors[0].message === 'requires property "foo"';
+            })
+          );
         });
       });
-
     });
 
+    describe("Custom ErrorList", () => {
+      const schema = {
+        type: "string",
+        required: true,
+        minLength: 1,
+      };
+
+      const formData = 0;
+
+      const CustomErrorList = ({ errors }) =>
+        <div className="CustomErrorList">
+          {errors.length} custom
+        </div>;
+
+      it("should use CustomErrorList", () => {
+        const { node } = createFormComponent({
+          schema,
+          liveValidate: true,
+          formData,
+          ErrorList: CustomErrorList,
+        });
+        expect(node.querySelectorAll(".CustomErrorList")).to.have.length.of(1);
+        expect(node.querySelector(".CustomErrorList").textContent).eql(
+          "1 custom"
+        );
+      });
+    });
   });
 });
